Aloe14 {metaSEM}	R Documentation

Multivariate effect sizes between classroom management self-efficacy (CMSE) and other variables reported by Aloe et al. (2014)

Description

This study reports sixteen studies on the effect sizes (correlation coefficients) between CMSE and emotional exhaustion (EE), depersonalization (DP), and (lowered) personal accomplishment (PA) reported by Aloe et al. (2014).

Usage

data("Aloe14")

Format

A data frame with 16 observations on the following 14 variables.

Study

a factor with levels Betoret Brouwers & Tomic Bumen Chang Durr Evers et al. Friedman Gold Huk Kress Kumarakulasingam Martin et al. Ozdemir Skaalvik and Skaalvik Williams

Year

Year of publication

EE

Emotional exhaustion

DP

Depersonalization

PA

(Lowered) personal accomplishment

V_EE

Sampling variance of emotional exhaustion

V_DP

Sampling variance of depersonalization

V_PA

Sampling variance of (lowered) personal accomplishment

C_EE_DP

Sampling covariance between EE and DP

C_EE_PA

Sampling covariance between EE and PA

C_DP_PA

Sampling covariance between DP and PA

Publication_type

Either Dissertation or Journal

Percentage_females

Percentage of females in the study

Years_experience

Average years of experience

Source

Aloe, A. M., Amo, L. C., & Shanahan, M. E. (2014). Classroom management self-efficacy and burnout: A multivariate meta-analysis. Educational Psychology Review, 26(1), 101-126. doi:10.1007/s10648-013-9244-0

Examples

## Not run: 
##D data(Aloe14)
##D 
##D ## Random-effects meta-analysis
##D meta1 <- meta(cbind(EE,DP,PA), cbind(V_EE, C_EE_DP, C_EE_PA, V_DP, C_DP_PA, V_PA),
##D               data=Aloe14)
##D summary(meta1)
##D 
##D ## Extract the coefficients for the variance component of the random effects
##D coef1 <- coef(meta1, select="random")
##D 
##D ## Convert it into a symmetrix matrix by row major
##D my.cov <- vec2symMat(coef1, byrow=TRUE)
##D 
##D ## Convert it into a correlation matrix
##D cov2cor(my.cov)
##D 
##D ## Plot the multivariate effect sizes
##D plot(meta1)
## End(Not run)

[Package metaSEM version 0.9.4 Index]