Becker09 {metaSEM}	R Documentation

Ten Studies of Correlation Matrices used by Becker (2009)

Description

This data set includes ten studies on the relationships between CSAI subscales and sport behavior. The original data were used in Craft et al. (2003), whereas a subset of them were illustrated in Becker (2009).

Usage

data("Becker09")

Details

A list of data with the following structure:

data

A list of 4x4 correlation matrices. The variables are Performance, Cognitive, Somatic, and Self confidence

n

A vector of sample sizes

Type_of_sport

Samples based on Individual or Team

Source

Craft, L. L., Magyar, T. M., Becker, B. J., & Feltz, D. L. (2003). The relationship between the Competitive State Anxiety Inventory-2 and sport performance: a meta-analysis. Journal of Sport and Exercise Psychology, 25(1), 44-65.

References

Becker, B. J. (2009). Model-based meta-analysis. In H. Cooper, L. V. Hedges, & J. C. Valentine (Eds.), The handbook of research synthesis and meta-analysis (2nd ed., pp. 377-395). New York: Russell Sage Foundation.

Examples

## Not run: 
##D data(Becker09)
##D 
##D #### Fixed-effects model
##D ## First stage analysis
##D fixed1 <- tssem1(Becker09$data, Becker09$n, method="FEM")
##D summary(fixed1)
##D 
##D ## Prepare a regression model using create.mxMatrix()
##D A1 <- create.mxMatrix(c(0, "0.1*Cog2Per", "0.1*SO2Per", "0.1*SC2Per",
##D                         0, 0, 0, 0,
##D                         0, 0, 0, 0,
##D                         0, "0.1*Cog2SC", "0.1*SO2SC",0),
##D                       type="Full", byrow=TRUE, ncol=4, nrow=4,
##D                       as.mxMatrix=FALSE)
##D 
##D ## This step is not necessary but it is useful for inspecting the model.
##D dimnames(A1)[[1]] <- dimnames(A1)[[2]] <- c("Per","Cog","SO","SC")
##D 
##D ## Display A1
##D A1
##D 
##D S1 <- create.mxMatrix(c("0.1*var_Per",
##D                         0, 1,
##D                         0, "0.1*cor", 1,
##D                         0, 0, 0, "0.1*var_SC"), byrow=TRUE, type="Symm",
##D                       as.mxMatrix=FALSE)
##D 
##D ## This step is not necessary but it is useful for inspecting the model.
##D dimnames(S1)[[1]] <- dimnames(S1)[[2]] <- c("Per","Cog","SO","SC")
##D 
##D ## Display S1
##D S1
##D 
##D ## Second stage analysis
##D fixed2 <- tssem2(fixed1, Amatrix=A1, Smatrix=S1, diag.constraints=TRUE,
##D                  intervals.type="LB", model.name="TSSEM2 Becker09",
##D                  mx.algebras=list( Cog=mxAlgebra(Cog2SC*SC2Per, name="Cog"),
##D                                    SO=mxAlgebra(SO2SC*SC2Per, name="SO"),
##D                                    Cog_SO=mxAlgebra(Cog2SC*SC2Per+SO2SC*SC2Per,
##D                                    name="Cog_SO")) )
##D summary(fixed2)
##D 
##D #### Fixed-effects model: with type of sport as cluster
##D ## First stage analysis
##D cluster1 <- tssem1(Becker09$data, Becker09$n, method="FEM",
##D                    cluster=Becker09$Type_of_sport)
##D summary(cluster1)
##D 
##D ## Second stage analysis
##D cluster2 <- tssem2(cluster1, Amatrix=A1, Smatrix=S1, diag.constraints=TRUE,
##D                  intervals.type="LB", model.name="TSSEM2 Becker09",
##D                  mx.algebras=list( Cog=mxAlgebra(Cog2SC*SC2Per, name="Cog"),
##D                                    SO=mxAlgebra(SO2SC*SC2Per, name="SO"),
##D                                    Cog_SO=mxAlgebra(Cog2SC*SC2Per+SO2SC*SC2Per,
##D                                    name="Cog_SO")) )
##D summary(cluster2)
##D 
##D ## Convert the model to semPlotModel object with 2 plots
##D ## Use the short forms of the variable names
##D my.plots <- lapply(X=cluster2, FUN=meta2semPlot, manNames=c("Per","Cog","SO","SC") )
##D 
##D ## Load the library
##D library("semPlot")
##D 
##D ## Setup two plots
##D layout(t(1:2))
##D ## The labels are overlapped. We may modify it by using layout="spring"
##D semPaths(my.plots[[1]], whatLabels="est", nCharNodes=10, color="orange",
##D          layout="spring", edge.label.cex=0.8)
##D title("Individual sport")
##D semPaths(my.plots[[2]], whatLabels="est", nCharNodes=10, color="skyblue",
##D          layout="spring", edge.label.cex=0.8)
##D title("Team sport")
##D 
##D 
##D #### Random-effects model
##D ## First stage analysis
##D random1 <- tssem1(Becker09$data, Becker09$n, method="REM", RE.type="Diag")
##D summary(random1)
##D 
##D ## Second stage analysis
##D random2 <- tssem2(random1, Amatrix=A1, Smatrix=S1, diag.constraints=TRUE,
##D                   intervals.type="LB", model.name="TSSEM2 Becker09",
##D                   mx.algebras=list( Cog=mxAlgebra(Cog2SC*SC2Per, name="Cog"),
##D                                     SO=mxAlgebra(SO2SC*SC2Per, name="SO"),
##D                                     Cog_SO=mxAlgebra(Cog2SC*SC2Per+SO2SC*SC2Per,
##D                                     name="Cog_SO")) )
##D summary(random2)
##D 
##D ## Load the library
##D library("semPlot")
##D 
##D ## Convert the model to semPlotModel object
##D my.plot <- meta2semPlot(random2, manNames=c("Per","Cog","SO","SC"))
##D 
##D ## Plot the model with labels
##D ## The labels are overlapped.
##D ## semPaths(my.plot, whatLabels="path", nCharEdges=10)
##D semPaths(my.plot, whatLabels="path", nCharEdges=10,  nCharNodes=10, layout="spring", color="red")
##D 
##D ## Plot the parameter estimates
##D semPaths(my.plot, whatLabels="est", nCharNodes=10, layout="spring", color="green")
## End(Not run)

[Package metaSEM version 0.9.4 Index]