Becker83 {metaSEM}	R Documentation

Studies on Sex Differences in Conformity Reported by Becker (1983)

Description

The data set includes studies on sex differences in conformity using the fictitious norm group paradigm reported by Becker (1983).

Usage

data(Becker83)

Details

The variables are:

study

study number

di

Standardized mean difference

vi

Sampling variance of the effect size

percentage

Percentage of male authors

items

Number of items

Source

Becker, B. J. (1983, April). Influence again: A comparison of methods for meta-analysis. Paper presented at the annual meeting of the American Educational Research Association, Montreal.

Hedges, L. V., & Olkin, I. (1985). Statistical methods for meta-analysis. Orlando, FL: Academic Press.

References

Cheung, M. W.-L. (2010). Fixed-effects meta-analyses as multiple-group structural equation models. Structural Equation Modeling, 17, 481-509.

Examples

data(Becker83)

## Random-effects meta-analysis
summary( meta(y=di, v=vi, data=Becker83) )

## 
## Call:
## meta(y = di, v = vi, data = Becker83)
## 
## 95% confidence intervals: z statistic approximation
## Coefficients:
##             Estimate Std.Error    lbound    ubound z value Pr(>|z|)
## Intercept1  0.174734  0.113378 -0.047482  0.396950  1.5412   0.1233
## Tau2_1_1    0.077376  0.054108 -0.028674  0.183426  1.4300   0.1527
## 
## Q statistic on the homogeneity of effect sizes: 30.64949
## Degrees of freedom of the Q statistic: 9
## P value of the Q statistic: 0.0003399239
## 
## Heterogeneity indices (based on the estimated Tau2):
##                              Estimate
## Intercept1: I2 (Q statistic)   0.6718
## 
## Number of studies (or clusters): 10
## Number of observed statistics: 10
## Number of estimated parameters: 2
## Degrees of freedom: 8
## -2 log likelihood: 7.928307 
## OpenMx status1: 0 ("0" or "1": The optimization is considered fine.
## Other values may indicate problems.)

## Mixed-effects meta-analysis with log(items) as the predictor
summary( meta(y=di, v=vi, x=log(items), data=Becker83) )

## 
## Call:
## meta(y = di, v = vi, x = log(items), data = Becker83)
## 
## 95% confidence intervals: z statistic approximation
## Coefficients:
##               Estimate   Std.Error      lbound      ubound z value
## Intercept1 -3.2015e-01  1.0981e-01 -5.3539e-01 -1.0492e-01 -2.9154
## Slope1_1    2.1088e-01  4.5084e-02  1.2251e-01  2.9924e-01  4.6774
## Tau2_1_1    1.0000e-10  2.0095e-02 -3.9386e-02  3.9386e-02  0.0000
##             Pr(>|z|)    
## Intercept1  0.003552 ** 
## Slope1_1   2.905e-06 ***
## Tau2_1_1    1.000000    
## ---
## Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
## 
## Q statistic on the homogeneity of effect sizes: 30.64949
## Degrees of freedom of the Q statistic: 9
## P value of the Q statistic: 0.0003399239
## 
## Explained variances (R2):
##                            y1
## Tau2 (no predictor)    0.0774
## Tau2 (with predictors) 0.0000
## R2                     1.0000
## 
## Number of studies (or clusters): 10
## Number of observed statistics: 10
## Number of estimated parameters: 3
## Degrees of freedom: 7
## -2 log likelihood: -4.208024 
## OpenMx status1: 0 ("0" or "1": The optimization is considered fine.
## Other values may indicate problems.)

[Package metaSEM version 0.9.5-2 Index]