Becker94 {metaSEM}	R Documentation

Five Studies of Ten Correlation Matrices reported by Becker and Schram (1994)

Description

This data set includes five studies of ten correlation matrices reported by Becker and Schram (1994).

Usage

data(Becker94)

Details

A list of data with the following structure:

data

A list of 10 correlation matrices. The variables are SAT (Math), Spatial and SAT (Verbal)

n

A vector of sample sizes

gender

Females or Males samples

Source

Becker, B. J., & Schram, C. M. (1994). Examining explanatory models through research synthesis. In H. Cooper & L. V. Hedges (Eds.), The handbook of research synthesis (pp. 357-381). New York: Russell Sage Foundation.

Examples

## \dontrun{ 
data(Becker94) 

#### Fixed-effects model 
## First stage analysis 
fixed1 <- tssem1(Becker94$data, Becker94$n, method="FEM") 
summary(fixed1) 

## 
## Call:
## tssem1FEM(my.df = my.df, n = n, cor.analysis = cor.analysis, 
##     model.name = model.name, cluster = cluster, suppressWarnings = suppressWarnings, 
##     silent = silent, run = run)
## 
## Coefficients:
##        Estimate Std.Error z value  Pr(>|z|)    
## S[1,2] 0.379961  0.037124 10.2350 < 2.2e-16 ***
## S[1,3] 0.334562  0.039947  8.3751 < 2.2e-16 ***
## S[2,3] 0.176461  0.042334  4.1683 3.069e-05 ***
## ---
## Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
## 
## Goodness-of-fit indices:
##                                      Value
## Sample size                       538.0000
## Chi-square of target model         63.6553
## DF of target model                 27.0000
## p value of target model             0.0001
## Chi-square of independence model  207.7894
## DF of independence model           30.0000
## RMSEA                               0.1590
## RMSEA lower 95% CI                  0.1096
## RMSEA upper 95% CI                  0.2117
## SRMR                                0.1586
## TLI                                 0.7709
## CFI                                 0.7938
## AIC                                 9.6553
## BIC                              -106.1169
## OpenMx status1: 0 ("0" or "1": The optimization is considered fine.
## Other values may indicate problems.)

## Prepare a regression model using create.mxMatrix() 
A1 <- create.mxMatrix(c(0,0,0,"0.2*Spatial2Aptitude", 
                        0,0,"0.2*Verbal2Aptitude",0,0), type="Full", 
                      ncol=3, nrow=3, name="A1") 
S1 <- 
create.mxMatrix(c("0.2*ErrorVarAptitude",0,0,1,"0.2*CorBetweenSpatialVerbal",1), 
                type="Symm", name="S1") 

## An alternative method to create a regression model using mxMatrix() 
# A1 <- mxMatrix("Full", ncol=3, nrow=3, value=0, free=c(FALSE,FALSE,FALSE,TRUE,FALSE, 
#                                                        FALSE,TRUE,FALSE,FALSE), 
#                label=c(NA,NA,NA,"Spatial2Aptitude",NA,NA,"Verbal2Aptitude",NA,NA), 
#                name="A1") 
# S1 <- mxMatrix("Symm", ncol=3, nrow=3, value=c(0.5,0,0,1,0.2,1), 
#                free=c(TRUE,FALSE,FALSE,FALSE,TRUE,FALSE), 
#                label=c("ErrorVarAptitude",NA,NA,NA,"CorBetweenSpatialVerbal",NA), 
#                        name="S1") 

## Second stage analysis 
fixed2 <- tssem2(fixed1, Amatrix=A1, Smatrix=S1, intervals.type="LB") 
summary(fixed2) 

## 
## Call:
## wls(Cov = coef.tssem1FEM(tssem1.obj), asyCov = vcov.tssem1FEM(tssem1.obj), 
##     n = sum(tssem1.obj$n), Amatrix = Amatrix, Smatrix = Smatrix, 
##     Fmatrix = Fmatrix, diag.constraints = diag.constraints, cor.analysis = tssem1.obj$cor.analysis, 
##     intervals.type = intervals.type, mx.algebras = mx.algebras, 
##     model.name = model.name, suppressWarnings = suppressWarnings, 
##     silent = silent, run = run)
## 
## 95% confidence intervals: Likelihood-based statistic
## Coefficients:
##                         Estimate Std.Error   lbound   ubound z value
## Spatial2Aptitude        0.331238        NA 0.258215 0.404341      NA
## Verbal2Aptitude         0.276111        NA 0.199380 0.352781      NA
## CorBetweenSpatialVerbal 0.176461        NA 0.093486 0.259436      NA
##                         Pr(>|z|)
## Spatial2Aptitude              NA
## Verbal2Aptitude               NA
## CorBetweenSpatialVerbal       NA
## 
## Goodness-of-fit indices:
##                                             Value
## Sample size                                538.00
## Chi-square of target model                   0.00
## DF of target model                           0.00
## p value of target model                      0.00
## Number of constraints imposed on "Smatrix"   0.00
## DF manually adjusted                         0.00
## Chi-square of independence model           160.47
## DF of independence model                     3.00
## RMSEA                                        0.00
## RMSEA lower 95% CI                           0.00
## RMSEA upper 95% CI                           0.00
## SRMR                                         0.00
## TLI                                          -Inf
## CFI                                          1.00
## AIC                                          0.00
## BIC                                          0.00
## OpenMx status1: 0 ("0" or "1": The optimization is considered fine.
## Other values indicate problems.)

#### Fixed-effects model: with gender as cluster 
## First stage analysis 
cluster1 <- tssem1(Becker94$data, Becker94$n, method="FEM", cluster=Becker94$gender) 
summary(cluster1) 

## $Females
## 
## Call:
## tssem1FEM(my.df = data.cluster[[i]], n = n.cluster[[i]], cor.analysis = cor.analysis, 
##     model.name = model.name, suppressWarnings = suppressWarnings)
## 
## Coefficients:
##        Estimate Std.Error z value  Pr(>|z|)    
## S[1,2] 0.455896  0.051992  8.7685 < 2.2e-16 ***
## S[1,3] 0.341583  0.061943  5.5144 3.499e-08 ***
## S[2,3] 0.171931  0.064731  2.6561  0.007905 ** 
## ---
## Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
## 
## Goodness-of-fit indices:
##                                     Value
## Sample size                      235.0000
## Chi-square of target model        43.1898
## DF of target model                12.0000
## p value of target model            0.0000
## Chi-square of independence model 123.4399
## DF of independence model          15.0000
## RMSEA                              0.2357
## RMSEA lower 95% CI                 0.1637
## RMSEA upper 95% CI                 0.3161
## SRMR                               0.2141
## TLI                                0.6405
## CFI                                0.7124
## AIC                               19.1898
## BIC                              -22.3252
## OpenMx status1: 0 ("0" or "1": The optimization is considered fine.
## Other values may indicate problems.)
## 
## $Males
## 
## Call:
## tssem1FEM(my.df = data.cluster[[i]], n = n.cluster[[i]], cor.analysis = cor.analysis, 
##     model.name = model.name, suppressWarnings = suppressWarnings)
## 
## Coefficients:
##        Estimate Std.Error z value  Pr(>|z|)    
## S[1,2] 0.318051  0.051698  6.1521 7.646e-10 ***
## S[1,3] 0.328286  0.052226  6.2859 3.261e-10 ***
## S[2,3] 0.179549  0.055944  3.2094   0.00133 ** 
## ---
## Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
## 
## Goodness-of-fit indices:
##                                     Value
## Sample size                      303.0000
## Chi-square of target model        16.4819
## DF of target model                12.0000
## p value of target model            0.1701
## Chi-square of independence model  84.3496
## DF of independence model          15.0000
## RMSEA                              0.0786
## RMSEA lower 95% CI                 0.0000
## RMSEA upper 95% CI                 0.1643
## SRMR                               0.1025
## TLI                                0.9192
## CFI                                0.9354
## AIC                               -7.5181
## BIC                              -52.0829
## OpenMx status1: 0 ("0" or "1": The optimization is considered fine.
## Other values may indicate problems.)

## Second stage analysis   
cluster2 <- tssem2(cluster1, Amatrix=A1, Smatrix=S1, intervals.type="LB") 
summary(cluster2) 

## $Females
## 
## Call:
## wls(Cov = coef.tssem1FEM(tssem1.obj), asyCov = vcov.tssem1FEM(tssem1.obj), 
##     n = sum(tssem1.obj$n), Amatrix = Amatrix, Smatrix = Smatrix, 
##     Fmatrix = Fmatrix, diag.constraints = diag.constraints, cor.analysis = tssem1.obj$cor.analysis, 
##     intervals.type = intervals.type, mx.algebras = mx.algebras, 
##     model.name = model.name, suppressWarnings = suppressWarnings, 
##     silent = silent, run = run)
## 
## 95% confidence intervals: Likelihood-based statistic
## Coefficients:
##                         Estimate Std.Error   lbound   ubound z value
## Spatial2Aptitude        0.409265        NA 0.305291 0.512538      NA
## Verbal2Aptitude         0.271217        NA 0.156124 0.386564      NA
## CorBetweenSpatialVerbal 0.171931        NA 0.045036 0.298803      NA
##                         Pr(>|z|)
## Spatial2Aptitude              NA
## Verbal2Aptitude               NA
## CorBetweenSpatialVerbal       NA
## 
## Goodness-of-fit indices:
##                                             Value
## Sample size                                235.00
## Chi-square of target model                   0.00
## DF of target model                           0.00
## p value of target model                      0.00
## Number of constraints imposed on "Smatrix"   0.00
## DF manually adjusted                         0.00
## Chi-square of independence model           105.57
## DF of independence model                     3.00
## RMSEA                                        0.00
## RMSEA lower 95% CI                           0.00
## RMSEA upper 95% CI                           0.00
## SRMR                                         0.00
## TLI                                          -Inf
## CFI                                          1.00
## AIC                                          0.00
## BIC                                          0.00
## OpenMx status1: 0 ("0" or "1": The optimization is considered fine.
## Other values indicate problems.)
## 
## $Males
## 
## Call:
## wls(Cov = coef.tssem1FEM(tssem1.obj), asyCov = vcov.tssem1FEM(tssem1.obj), 
##     n = sum(tssem1.obj$n), Amatrix = Amatrix, Smatrix = Smatrix, 
##     Fmatrix = Fmatrix, diag.constraints = diag.constraints, cor.analysis = tssem1.obj$cor.analysis, 
##     intervals.type = intervals.type, mx.algebras = mx.algebras, 
##     model.name = model.name, suppressWarnings = suppressWarnings, 
##     silent = silent, run = run)
## 
## 95% confidence intervals: Likelihood-based statistic
## Coefficients:
##                         Estimate Std.Error   lbound   ubound z value
## Spatial2Aptitude        0.267739        NA 0.166906 0.368546      NA
## Verbal2Aptitude         0.280213        NA 0.178021 0.382429      NA
## CorBetweenSpatialVerbal 0.179549        NA 0.069891 0.289199      NA
##                         Pr(>|z|)
## Spatial2Aptitude              NA
## Verbal2Aptitude               NA
## CorBetweenSpatialVerbal       NA
## 
## Goodness-of-fit indices:
##                                              Value
## Sample size                                303.000
## Chi-square of target model                   0.000
## DF of target model                           0.000
## p value of target model                      0.000
## Number of constraints imposed on "Smatrix"   0.000
## DF manually adjusted                         0.000
## Chi-square of independence model            68.564
## DF of independence model                     3.000
## RMSEA                                        0.000
## RMSEA lower 95% CI                           0.000
## RMSEA upper 95% CI                           0.000
## SRMR                                         0.000
## TLI                                           -Inf
## CFI                                          1.000
## AIC                                          0.000
## BIC                                          0.000
## OpenMx status1: 0 ("0" or "1": The optimization is considered fine.
## Other values indicate problems.)

#### Conventional fixed-effects GLS approach 
## First stage analysis 
## No random effects 
## Replicate Becker's (1992) analysis using 4 studies only 
gls1 <- tssem1(Becker92$data[1:4], Becker92$n[1:4], method="REM", RE.type="Zero", 
               model.name="Fixed effects GLS Stage 1") 
summary(gls1) 

## 
## Call:
## meta(y = ES, v = acovR, RE.constraints = matrix(0, ncol = no.es, 
##     nrow = no.es), I2 = I2, model.name = model.name, suppressWarnings = TRUE, 
##     silent = silent, run = run)
## 
## 95% confidence intervals: z statistic approximation
## Coefficients:
##            Estimate Std.Error   lbound   ubound z value  Pr(>|z|)    
## Intercept1 0.410464  0.046612 0.319106 0.501822   8.806 < 2.2e-16 ***
## Intercept2 0.471172  0.043735 0.385453 0.556890  10.773 < 2.2e-16 ***
## Intercept3 0.252607  0.052616 0.149482 0.355732   4.801 1.579e-06 ***
## ---
## Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
## 
## Q statistic on the homogeneity of effect sizes: 9.331465
## Degrees of freedom of the Q statistic: 9
## P value of the Q statistic: 0.4072567
## 
## Heterogeneity indices (based on the estimated Tau2):
##                              Estimate
## Intercept1: I2 (Q statistic)        0
## Intercept2: I2 (Q statistic)        0
## Intercept3: I2 (Q statistic)        0
## 
## Number of studies (or clusters): 4
## Number of observed statistics: 12
## Number of estimated parameters: 3
## Degrees of freedom: 9
## -2 log likelihood: -26.0304 
## OpenMx status1: 0 ("0" or "1": The optimization is considered fine.
## Other values may indicate problems.)

## Fixed-effects GLS model: Second stage analysis 
gls2 <- tssem2(gls1, Amatrix=A1, Smatrix=S1, intervals.type="LB", 
               model.name="Fixed effects GLS Stage 2") 
summary(gls2) 

## 
## Call:
## wls(Cov = pooledS, asyCov = asyCov, n = tssem1.obj$total.n, Amatrix = Amatrix, 
##     Smatrix = Smatrix, Fmatrix = Fmatrix, diag.constraints = diag.constraints, 
##     cor.analysis = cor.analysis, intervals.type = intervals.type, 
##     mx.algebras = mx.algebras, model.name = model.name, suppressWarnings = suppressWarnings, 
##     silent = silent, run = run)
## 
## 95% confidence intervals: Likelihood-based statistic
## Coefficients:
##                         Estimate Std.Error  lbound  ubound z value
## Spatial2Aptitude         0.31131        NA 0.21931 0.40291      NA
## Verbal2Aptitude          0.39253        NA 0.30338 0.48139      NA
## CorBetweenSpatialVerbal  0.25261        NA 0.14947 0.35576      NA
##                         Pr(>|z|)
## Spatial2Aptitude              NA
## Verbal2Aptitude               NA
## CorBetweenSpatialVerbal       NA
## 
## Goodness-of-fit indices:
##                                             Value
## Sample size                                311.00
## Chi-square of target model                   0.00
## DF of target model                           0.00
## p value of target model                      0.00
## Number of constraints imposed on "Smatrix"   0.00
## DF manually adjusted                         0.00
## Chi-square of independence model           172.96
## DF of independence model                     3.00
## RMSEA                                        0.00
## RMSEA lower 95% CI                           0.00
## RMSEA upper 95% CI                           0.00
## SRMR                                         0.00
## TLI                                          -Inf
## CFI                                          1.00
## AIC                                          0.00
## BIC                                          0.00
## OpenMx status1: 0 ("0" or "1": The optimization is considered fine.
## Other values indicate problems.)

## } 

[Package metaSEM version 0.9.8 Index]