Cheung00 {metaSEM}	R Documentation

Fifty Studies of Correlation Matrices used in Cheung and Chan (2000)

Description

This data set includes fifty studies of correlation matrices on theory of planned theory reported by Cheung and Chan (2000).

Usage

data(Cheung00)

Details

A list of data with the following structure:

data

A list of 50 studies of correlation matrices. The variables are attitude toward behavior att, subjective norm sn, behavioral intention bi, and behavior beh

n

A vector of sample sizes

Note

These studies were extracted from the original data set for illustration purpose. Some samples contained two or more correlation matrices and only one of them was arbitrarily selected to avoid the problem of dependence. Moreover, studies with less than 3 correlation coefficients were also excluded.

Source

Cheung, S.-F., & Chan, D. K.-S. (2000). The role of perceived behavioral control in predicting human behavior: A meta-analytic review of studies on the theory of planned behavior. Unpublished manuscript, Chinese University of Hong Kong.

References

Cheung, M.W.-L., & Cheung, S.-F. (2014). Random effects models for meta-analytic structural equation modeling: Review, issues, and illustrations. Manuscript submitted for publication.

Examples

## Not run: 
##D data(Cheung00)
##D 
##D ## Variable labels
##D labels <- colnames(Cheung00$data[[1]])
##D 
##D ## Full mediation model
##D S <- create.mxMatrix(c("1", 
##D                        ".2*cov_att_sn", "1", 
##D                        0, 0, ".2*e_bi", 
##D                        0, 0, 0, ".2*e_beh"), 
##D                      type="Symm", as.mxMatrix=FALSE, byrow=TRUE)
##D dimnames(S) <- list(labels, labels)
##D S
##D 
##D A <- matrix(c("0","0","0","0",
##D               "0","0","0","0",
##D               ".2*att2bi", ".2*sn2bi", "0", "0",
##D               "0", "0", ".2*bi2beh", "0"),
##D             byrow=TRUE, 4, 4)
##D dimnames(A) <- list(labels, labels)
##D A
##D 
##D #### Random-effects model
##D 
##D ## Stage 1 analysis
##D random_1 <- tssem1(Cheung00$data, Cheung00$n, method="REM")
##D summary(random_1)
##D 
##D ## Stage 2 analysis
##D random_2 <- tssem2(random_1, Amatrix=A, Smatrix=S, intervals.type="LB", diag.constraints=TRUE)
##D summary(random_2)
##D 
##D ## Load the library
##D library("semPlot")
##D 
##D my.plot1 <- meta2semPlot(random_2)
##D semPaths(my.plot1, whatLabels="est", col="yellow")
## End(Not run)

[Package metaSEM version 0.9.4 Index]