Craft03 {metaSEM}	R Documentation

Ten Correlation Matrices from Craft et al. (2003)

Description

This data set includes ten studies on the relationships between CSAI subscales and sport behavior. The original data were used in Craft et al. (2003), whereas a subset of them were illustrated in Becker (2009).

Usage

data(Craft03)

Details

A list of data with the following structure:

data

A list of 10 correlation matrices. The variables are Performance, Cognitive, Somatic and Self confidence

n

A vector of sample sizes

Type_of_sport

Individual or Team

Source

Craft, L. L., Magyar, T. M., Becker, B. J., & Feltz, D. L. (2003). The relationship between the Competitive State Anxiety Inventory-2 and sport performance: a meta-analysis. Journal of Sport and Exercise Psychology, 25(1), 44-65.

References

Becker, B. J. (2009). Model-based meta-analysis. In H. Cooper, L. V. Hedges, & J. C. Valentine (Eds.), The Handbook of Research Synthesis and Meta-analysis (2nd ed., pp. 377-395). New York: Russell Sage Foundation.

Examples

data(Craft03)

[Package metaSEM version 0.9-1 Index]