Diag {metaSEM}	R Documentation

Matrix Diagonals

Description

Extract or replace the diagonal of a matrix, or construct a diagonal matrix with the same behaviors as diag prior to R-3.0.0.

Usage

Diag(x, ...)
Diag(x) <- value

Arguments

x
a matrix, vector or 1D array, or missing.

...
Optional dimensions (nrow and ncol) for the result when x is not a matrix.

value
either a single value or a vector of length equal to that of the current diagonal. Should be of a mode which can be coerced to that of x.

Details

Started from R-3.0.0, diag(x) returns a numeric matrix with NA in the diagonals when x is a character vector. Although this follows exactly what the manual says, this breaks the metaSEM. The Diag has the same functions as diag except that Diag(x) works for a character vector of x by returning a square matrix of character "0" with x as the diagonals.

Note

See http://r.789695.n4.nabble.com/Behaviors-of-diag-with-character-vector-in-R-3-0-0-td4663735.html for the discussion.

Author(s)

Mike W.-L. Cheung <mikewlcheung@nus.edu.sg>

See Also

diag

Examples

v <- c("a", "b")
Diag(v)

##      [,1] [,2]
## [1,] "a"  "0" 
## [2,] "0"  "b"

[Package metaSEM version 0.9.3-1 Index]