HedgesOlkin85 {metaSEM}	R Documentation

Effects of Open Education Reported by Hedges and Olkin (1985)

Description

Effects of open education on attitude toward school and on reading achievement reported by Hedges and Olkin (1985).

Usage

data(HedgesOlkin85)

Details

The variables are:

study

Study number

d_att

Standardized mean difference on attitude

d_ach

Standardized mean difference on achievement

var_att

Sampling variance of the effect size of attitude

cov_att_ach

Sampling covariance between the effect sizes

var_ach

Sampling variance of the effect size of achievement

Source

Hedges, L. V., & Olkin, I. (1985). Statistical methods for meta-analysis. Orlando, FL: Academic Press.

References

Cheung, M. W.-L. (2010). Fixed-effects meta-analyses as multiple-group structural equation models. Structural Equation Modeling, 17, 481-509.

Examples

data(HedgesOlkin85)

## Fixed-effects meta-analysis
summary( meta(y=cbind(d_att, d_ach),
              v=cbind(var_att, cov_att_ach, var_ach),
              data=HedgesOlkin85,
              RE.constraints=matrix(0, nrow=2, ncol=2)) )

## 
## Call:
## meta(y = cbind(d_att, d_ach), v = cbind(var_att, cov_att_ach, 
##     var_ach), data = HedgesOlkin85, RE.constraints = matrix(0, 
##     nrow = 2, ncol = 2))
## 
## 95% confidence intervals: z statistic approximation
## Coefficients:
##             Estimate Std.Error    lbound    ubound z value Pr(>|z|)   
## Intercept1  0.320969  0.099492  0.125968  0.515971  3.2261 0.001255 **
## Intercept2  0.077797  0.099128 -0.116491  0.272085  0.7848 0.432564   
## ---
## Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
## 
## Q statistic on the homogeneity of effect sizes: 2.248917
## Degrees of freedom of the Q statistic: 6
## P value of the Q statistic: 0.8954419
## 
## Heterogeneity indices (based on the estimated Tau2):
##                              Estimate
## Intercept1: I2 (Q statistic)        0
## Intercept2: I2 (Q statistic)        0
## 
## Number of studies (or clusters): 4
## Number of observed statistics: 8
## Number of estimated parameters: 2
## Degrees of freedom: 6
## -2 log likelihood: -10.64183 
## OpenMx status1: 0 ("0" or "1": The optimization is considered fine.
## Other values may indicate problems.)

[Package metaSEM version 0.9.4 Index]