Hunter83 {metaSEM}	R Documentation

Fourteen Studies of Correlation Matrices reported by Hunter (1983)

Description

This data set includes fourteen studies of Correlation Matrices reported by Hunter (1983)

Usage

data(Hunter83)

Details

A list of data with the following structure:

data

A list of 14 studies of correlation matrices. The variables are Ability, Job knowledge, Work sample and Supervisor rating

n

A vector of sample sizes

Source

Hunter, J. E. (1983). A causal analysis of cognitive ability, job knowledge, job performance, and supervisor ratings. In F. Landy, S. Zedeck, & J. Cleveland (Eds.), Performance Measurement and Theory (pp. 257-266). Hillsdale, NJ: Erlbaum.

Examples

## Not run: 
##D data(Hunter83)
##D 
##D #### Fixed-effects model
##D ## First stage analysis
##D fixed1 <- tssem1(Hunter83$data, Hunter83$n, method="FEM",
##D                  model.name="TSSEM1 fixed effects model")
##D summary(fixed1)
##D 
##D #### Second stage analysis
##D ## Model without direct effect from Ability to Supervisor
##D A1 <- create.mxMatrix(c(0,"0.1*A2J","0.1*A2W",0,0,0,"0.1*J2W","0.1*J2S",
##D                         0,0,0,"0.1*W2S",0,0,0,0),
##D                         type="Full", ncol=4, nrow=4, as.mxMatrix=FALSE)
##D 
##D ## This step is not necessary but it is useful for inspecting the model.
##D dimnames(A1)[[1]] <- dimnames(A1)[[2]] <- c("Ability","Job","Work","Supervisor") 
##D A1
##D 
##D S1 <- create.mxMatrix(c(1,"0.1*Var_e_J", "0.1*Var_e_W", "0.1*Var_e_S"),
##D                       type="Diag", as.mxMatrix=FALSE)
##D dimnames(S1)[[1]] <- dimnames(S1)[[2]] <- c("Ability","Job","Work","Supervisor") 
##D S1
##D 
##D ## diag.constraints=TRUE is required as there are mediators  
##D fixed2 <- tssem2(fixed1, Amatrix=A1, Smatrix=S1, intervals.type="LB",
##D                  diag.constraints=TRUE,
##D                  model.name="TSSEM2 fixed effects model")
##D summary(fixed2)
##D 
##D ## Coefficients
##D coef(fixed2)
##D 
##D ## VCOV based on parametric bootstrap
##D vcov(fixed2)
##D 
##D #### Random-effects model with diagonal elements only
##D ## First stage analysis
##D random1 <- tssem1(Hunter83$data, Hunter83$n, method="REM", RE.type="Diag", 
##D                   model.name="TSSEM1 random effects model")
##D summary(random1)
##D 
##D ## Second stage analysis
##D ## Model without direct effect from Ability to Supervisor
##D 
##D ## diag.constraints=TRUE is required as there are mediators 
##D random2 <- tssem2(random1, Amatrix=A1, Smatrix=S1, intervals.type="LB",
##D                   diag.constraints=TRUE,
##D                   mx.algebras=
##D                   list( ind=mxAlgebra(A2J*J2S+A2J*J2W*W2S+A2W*W2S, name="ind") ),
##D                   model.name="TSSEM2 random effects model")
##D summary(random2)
##D 
##D ## Load the library
##D library("semPlot")
##D 
##D ## Convert the model to semPlotModel object
##D my.plot <- meta2semPlot(random2)
##D 
##D ## Plot the model with labels
##D semPaths(my.plot, whatLabels="path", nCharEdges=10, nCharNodes=10, color="red")
##D 
##D ## Plot the parameter estimates
##D semPaths(my.plot, whatLabels="est", nCharNodes=10, color="green")
## End(Not run)

[Package metaSEM version 0.9.4 Index]