Jaramillo05 {metaSEM}	R Documentation

Dataset from Jaramillo, Mulki & Marshall (2005)

Description

A dataset of the relationship between organizational commitment (OC) and salesperson job performance (JP) from Jaramillo, Mulki & Marshall (2005).

Usage

data(Jaramillo05)

Format

A data frame with 61 observations on the following 10 variables.

Author

a character vector of study

Sample_size

sample size of the study

Sales

sample type; either "mixed", "nonsales" or "sales"

Country

a character vector of country of study

IDV

Hofstede's (1997) individualism index

OC_scale

scale of OC; either "Porter or Mowday", "Meyer" or "other"

OC_alpha

Coefficient alpha of organizational commitment

JP_alpha

Coefficient alpha of job performance

r

correlation between organizational commitment and job performance

r_v

sampling variance of r

Source

Jaramillo, F., Mulki, J. P., & Marshall, G. W. (2005). A meta-analysis of the relationship between organizational commitment and salesperson job performance: 25 years of research. Journal of Business Research, 58(6), 705-714. doi:10.1016/j.jbusres.2003.10.004

Examples

## Not run: 
##D ## Research question 4.4.1
##D summary(meta(r, r_v, data=Jaramillo05))
##D 
##D ## Research question 4.4.2
##D ## Select cases with either "sales" or "nonsales"
##D Sales.df <- subset(Jaramillo05, Sales %in% c("sales", "nonsales"))
##D 
##D ## Create a predictor with 1 and 0 when they are "sales" or "nonsales", respectively
##D predictor <- ifelse(Jaramillo05$Sales=="sales", yes=1, no=0)
##D 
##D ## Mixed-effects meta-analysis
##D summary( meta(y = r, v = r_v, x = predictor, data = Jaramillo05) )
##D 
##D ## Research question 4.4.3
##D summary(meta(r, r_v, x=IDV, data=Jaramillo05))
## End(Not run)

[Package metaSEM version 0.9.4 Index]