Norton13 {metaSEM}	R Documentation

Studies on the Hospital Anxiety and Depression Scale Reported by Norton et al. (2013)

Description

The data set includes 28 studies on 14 items measuring the Hospital Anxiety and Depression Scale (HADS) Reported by Norton et al. (2013).

Usage

data(Norton13)

Details

The variables are:

data

A list of 28 studies of correlation matrices. The variables are 14 items (x1 to x14) measuring HADS.

n

A vector of sample sizes

population

A vector of population of the data

group

A vector of classification into patients vs. non-patients based on population

Source

Norton, S., Cosco, T., Doyle, F., Done, J., & Sacker, A. (2013). The Hospital Anxiety and Depression Scale: A meta confirmatory factor analysis. Journal of Psychosomatic Research, 74(1), 74-81.

References

Jak, S., & Cheung, M. W.-L. (2015). Addressing heterogeneity in meta-analytic structural equation modeling using subgroup analysis. Manuscript in preparation.

Examples

data(Norton13) 

[Package metaSEM version 0.9.8 Index]