asyCov {metaSEM}	R Documentation

Compute Asymptotic Covariance Matrix of a Correlation/Covariance Matrix

Description

It computes the asymptotic sampling covariance matrix of a correlation/covariance matrix under the assumption of multivariate normality.

Usage

asyCov(x, n, cor.analysis = TRUE, dropNA = FALSE, as.matrix = TRUE,
       suppressWarnings = TRUE,  silent = TRUE, run = TRUE, ...)

Arguments

x
A correlation/covariance matrix or a list of correlation/covariance matrices. NA on the variables or other values defined in na.strings will be removed before the analysis. Note that it only checks the diagonal elements of the matrices. If there are missing values, make sure that the diagonals are coded with NA or values defined in na.strings.

n
Sample size or a vector of sample sizes

cor.analysis
Logical. The output is either a correlation or covariance matrix.

dropNA
Logical. If it is TRUE, the resultant dimensions will be reduced by dropping the missing variables. If it is FALSE, the resultant dimensions are the same as the input by keeping the missing variables.

as.matrix
Logical. If it is TRUE and x is a list of correlation/covariance matrices with the same dimensions, the asymptotic covariance matrices will be column vectorized and stacked together. If it is FALSE, the output will be a list of asymptotic covariance matrices. Note that if it is TRUE, dropNA will be FALSE automatically. This option is useful when passing the asymptotic covariance matrices to meta

suppressWarnings
Logical. If TRUE, warnings are suppressed. Argument to be passed to mxRun.

silent
Logical. Argument to be passed to mxRun

run
Logical. If FALSE, only return the mx model without running the analysis.

...
Futher arguments to be passed to mxRun

Value

An asymptotic covariance matrix of the vectorized correlation/covariance matrix or a list of these matrices. If as.matrix=TRUE and x is a list of matrices, the output is a stacked matrix.

Author(s)

Mike W.-L. Cheung <mikewlcheung@nus.edu.sg>

References

Cheung, M. W.-L., & Chan, W. (2004). Testing dependent correlation coefficients via structural equation modeling. Organizational Research Methods, 7, 206-223.

Examples

C1 <- matrix(c(1,0.5,0.4,0.5,1,0.2,0.4,0.2,1), ncol=3)  
asyCov(C1, n=100)

##              x2x1         x3x1        x3x2
## x2x1 0.0056818219 0.0006363572 0.002590906
## x3x1 0.0006363572 0.0071272787 0.003818178
## x3x2 0.0025909062 0.0038181776 0.009309103

## Data with missing values
C2 <- matrix(c(1,0.4,NA,0.4,1,NA,NA,NA,NA), ncol=3)  
C3 <- matrix(c(1,0.2,0.2,1), ncol=2)

## Output is a list of asymptotic covariance matrices
asyCov(list(C1,C2,C3), n=c(100,50,50), dropNA=TRUE, as.matrix=FALSE)

## [[1]]
##              x2x1         x3x1        x3x2
## x2x1 0.0056818143 0.0006363488 0.002590897
## x3x1 0.0006363488 0.0071272639 0.003818169
## x3x2 0.0025908973 0.0038181689 0.009309094
## 
## [[2]]
##            x2x1
## x2x1 0.01440001
## 
## [[3]]
##            x2x1
## x2x1 0.01880814

## Output is a stacked matrix of asymptotic covariance matrices
asyCov(list(C1,C2), n=c(100,50), as.matrix=TRUE)

##      cov(x2x1_x2x1) cov(x3x1_x2x1) cov(x3x2_x2x1) cov(x3x1_x3x1)
## [1,]    0.005681821   0.0006363485    0.002590896    0.007127279
## [2,]    0.014399992             NA             NA             NA
##      cov(x3x2_x3x1) cov(x3x2_x3x2)
## [1,]    0.003818179    0.009309097
## [2,]             NA             NA

## Output is a stacked matrix of asymptotic covariance matrices
asyCov(list(C3,C3), n=c(100,50), as.matrix=TRUE)

##      cov(x2x1_x2x1)
## [1,]    0.009309085
## [2,]    0.018808220

[Package metaSEM version 0.9-2 Index]