create.mxMatrix {metaSEM}	R Documentation

Create a Vector into MxMatrix-class

Description

It converts a vector into MxMatrix-class via mxMatrix.

Usage

create.mxMatrix(x, type=c("Full","Symm","Diag","Stand"), ncol=NA,
nrow=NA, as.mxMatrix=TRUE, byrow=FALSE, ...) 

Arguments

x
A character or numeric vector

type
Matrix type similar to those listed in mxMatrix

ncol
Number of columns. It is necessary when type="Full". It is ignored and determined by the length of x for the other types of matrices.

nrow
Number of rows. It is necessary when type="Full". It is ignored and determined by the length of x for the other types of matrices.

as.mxMatrix
Logical. If it is TRUE, the output is a matrix of MxMatrix-class. If it is FALSE, it is a numeric matrix.

byrow
Logical. If FALSE (the default) the matrix is filled by columns, otherwise the matrix is filled by rows.

...
Further arguments to be passed to mxMatrix. Please note that type, nrow, ncol, values, free and labels will be created automatically. Thus, these arguments excepts labels should be avoided in ...

Details

If there are non-numeric values in x, they are treated as the labels of the free parameters. If an "*" is present, the numeric value on the left hand side will be treated as the starting value for a free parameter or a fixed value for a fixed parameter. If it is a matrix of numeric values, there is no free parameters in the output matrix. nrow and ncol will be calculated from the length of x unless type="Full" is specified.

Value

A MxMatrix-class object with the same dimensions as x

Author(s)

Mike W.-L. Cheung <mikewlcheung@nus.edu.sg>

See Also

mxMatrix, create.mxMatrix, create.Fmatrix

Examples

## a and b are free parameters with starting values and labels
(a1 <- c(1:4, "5*a", 6, "7*b", 8, 9))

## [1] "1"   "2"   "3"   "4"   "5*a" "6"   "7*b" "8"   "9"

(mat1 <- create.mxMatrix(a1, ncol=3, nrow=3, name="mat1"))

## FullMatrix 'mat1' 
## 
## $labels
##      [,1] [,2] [,3]
## [1,] NA   NA   "b" 
## [2,] NA   "a"  NA  
## [3,] NA   NA   NA  
## 
## $values
##      [,1] [,2] [,3]
## [1,]    1    4    7
## [2,]    2    5    8
## [3,]    3    6    9
## 
## $free
##       [,1]  [,2]  [,3]
## [1,] FALSE FALSE  TRUE
## [2,] FALSE  TRUE FALSE
## [3,] FALSE FALSE FALSE
## 
## $lbound: No lower bounds assigned.
## 
## $ubound: No upper bounds assigned.

## Arrange the elements by row
(mat2 <- create.mxMatrix(a1, ncol=3, nrow=3, as.mxMatrix=FALSE, byrow=TRUE))

##      [,1]  [,2]  [,3]
## [1,] "1"   "2"   "3" 
## [2,] "4"   "5*a" "6" 
## [3,] "7*b" "8"   "9"

(a3 <- c(1:3, "4*f4", "5*f5", "6*f6"))

## [1] "1"    "2"    "3"    "4*f4" "5*f5" "6*f6"

(mat3 <- create.mxMatrix(a3, type="Symm", name="mat3"))

## SymmMatrix 'mat3' 
## 
## $labels
##      [,1] [,2] [,3]
## [1,] NA   NA   NA  
## [2,] NA   "f4" "f5"
## [3,] NA   "f5" "f6"
## 
## $values
##      [,1] [,2] [,3]
## [1,]    1    2    3
## [2,]    2    4    5
## [3,]    3    5    6
## 
## $free
##       [,1]  [,2]  [,3]
## [1,] FALSE FALSE FALSE
## [2,] FALSE  TRUE  TRUE
## [3,] FALSE  TRUE  TRUE
## 
## $lbound: No lower bounds assigned.
## 
## $ubound: No upper bounds assigned.

## Create character matrix
(mat4 <- create.mxMatrix(a3, type="Symm", as.mxMatrix=FALSE))

##      [,1] [,2]   [,3]  
## [1,] "1"  "2"    "3"   
## [2,] "2"  "4*f4" "5*f5"
## [3,] "3"  "5*f5" "6*f6"

## Arrange the elements by row
(mat5 <- create.mxMatrix(a3, type="Symm", as.mxMatrix=FALSE, byrow=TRUE))

##      [,1]   [,2]   [,3]  
## [1,] "1"    "2"    "4*f4"
## [2,] "2"    "3"    "5*f5"
## [3,] "4*f4" "5*f5" "6*f6"

(mat6 <- create.mxMatrix(a3, type="Diag", lbound=6:1, name="mat6"))

## DiagMatrix 'mat6' 
## 
## $labels
##      [,1] [,2] [,3] [,4] [,5] [,6]
## [1,] NA   NA   NA   NA   NA   NA  
## [2,] NA   NA   NA   NA   NA   NA  
## [3,] NA   NA   NA   NA   NA   NA  
## [4,] NA   NA   NA   "f4" NA   NA  
## [5,] NA   NA   NA   NA   "f5" NA  
## [6,] NA   NA   NA   NA   NA   "f6"
## 
## $values
##      [,1] [,2] [,3] [,4] [,5] [,6]
## [1,]    1    0    0    0    0    0
## [2,]    0    2    0    0    0    0
## [3,]    0    0    3    0    0    0
## [4,]    0    0    0    4    0    0
## [5,]    0    0    0    0    5    0
## [6,]    0    0    0    0    0    6
## 
## $free
##       [,1]  [,2]  [,3]  [,4]  [,5]  [,6]
## [1,] FALSE FALSE FALSE FALSE FALSE FALSE
## [2,] FALSE FALSE FALSE FALSE FALSE FALSE
## [3,] FALSE FALSE FALSE FALSE FALSE FALSE
## [4,] FALSE FALSE FALSE  TRUE FALSE FALSE
## [5,] FALSE FALSE FALSE FALSE  TRUE FALSE
## [6,] FALSE FALSE FALSE FALSE FALSE  TRUE
## 
## $lbound
##      [,1] [,2] [,3] [,4] [,5] [,6]
## [1,]    6   NA   NA   NA   NA   NA
## [2,]   NA    5   NA   NA   NA   NA
## [3,]   NA   NA    4   NA   NA   NA
## [4,]   NA   NA   NA    3   NA   NA
## [5,]   NA   NA   NA   NA    2   NA
## [6,]   NA   NA   NA   NA   NA    1
## 
## $ubound: No upper bounds assigned.

[Package metaSEM version 0.9-1 Index]