deltaMethod {metaSEM}	R Documentation

Approximate Variance or Covariance Matrix of Functions Using the Delta Method

Description

It calculates the approximate variance or covariance matrix of functions using the delta method.

Usage

deltaMethod(fn, variables, Mean, Cov)

Arguments

fn
A list of functions in characters.

variables
Variables used in fn, Mean, and Cov

Mean
The sample mean or the parameter estimates

Cov
The sample covariance matrix or the sampling covariance of the paramter estimates

Value

The approximate variance or the covariance matrix of fn

Note

This function is modified from the deltamethod() in the msm package

Author(s)

Mike W.-L. Cheung <mikewlcheung@nus.edu.sg>

Examples

## Not run: 
##D ## Example taken from http://lavaan.ugent.be/tutorial/mediation.html
##D ## Generate artificial data to estimate the indirect effect
##D set.seed(1234)
##D X <- rnorm(100)
##D M <- 0.5*X + rnorm(100)
##D Y <- 0.7*M + rnorm(100)
##D Data <- data.frame(X=X, Y=Y, M=M)
##D 
##D model <- '# direct effect
##D           Y ~ c*X
##D           # mediator
##D           M ~ a*X
##D           Y ~ b*M
##D           # indirect effect (a*b)
##D           ab := a*b
##D           # total effect
##D           total := c + (a*b)'
##D fit <- sem(model, data = Data)
##D summary(fit)
## End(Not run)

## Approximate standard errors of the indirect and total effects using the delta method.
## The first and second functions are the direct and indirect effects, respectively.
fn <- list("a*b", "a*b+c")
variables <- c("a", "b", "c")

## Results from the above analysis
Mean <- c(0.47391503, 0.78831930, 0.03635467)

Cov <- structure(c(0.0105540501465901, -5.36944694410567e-19,
4.78727359106282e-18, 3.56339238409499e-19, 0.00852258573549811,
-0.00403898148221173, 1.60078511618195e-18, -0.00403898148221173,
0.0109089137570649), .Dim = c(3L, 3L), .Dimnames = list(c("a", "b", "c"), c("a", 
"b", "c")))

ind.Cov <- deltaMethod(fn=fn, variables=variables, Mean=Mean, Cov=Cov)

## Standard errors
sqrt(diag(ind.Cov))

## [1] 0.09204847 0.12471394

## Approximate sampling variance of log(p/(1-p))
## Number of success
x <- 20
## Total number of trials
n <- 100
## Estimated proportion of success
p <- x/n
## Sampling variance of p
v <- p*(1-p)/n

## logit(p)
fn <- "log(p/(1-p))"
variables <- c("p")

## Sampling variance of logit(p)
deltaMethod(fn=fn, variables=variables, Mean=p, Cov=v)

##        [,1]
## [1,] 0.0625

[Package metaSEM version 0.9-2 Index]