issp05 {metaSEM}	R Documentation

Data Set from ISSP (2005)

Description

Thirty-two covariance matrices on work-related attitudes were extracted from the International Social Survey Programme 2005: Work Orientation III (ISSP, 2005). Seven variables were selected for demonstration purposes. They were grouped into three constructs: Importance of Job Prospects measured by job security (JP1), high income (JP2), and opportunity for advancement (JP3); Importance of Job Autonomy measured by work independently (JA1) and decide time of work (JA2); and Importance of Contributions to Society measured by help other people (CS1) and a job useful to society (CS2).

Usage

data(issp05)

Details

A list of data with the following structure:

data

A list of 32 covariance matrices

n

A vector of sample sizes

Author(s)

Mike W.-L. Cheung <mikewlcheung@nus.edu.sg>

Source

ISSP Research Group (2007): International Social Survey Programme 2005: Work Orientation III (ISSP 2005). GESIS Data Archive, Cologne. ZA4350 Data file Version 1.0.0, doi:10.4232/1.4350

See Also

issp89

Examples

## Not run: 
##D data(issp05)
##D 
##D #### Fixed-effects TSSEM
##D fixed1 <- tssem1(issp05$data, issp05$n, method="FEM")
##D summary(fixed1)
##D 
##D ## Prepare a model implied matrix
##D ## Factor correlation matrix
##D Phi <- create.mxMatrix( c("0.3*corf2f1","0.3*corf3f1","0.3*corf3f2"),
##D                         type="Stand", as.mxMatrix=FALSE )
##D 
##D ## Error variances
##D Psi <- create.mxMatrix( paste("0.2*e", 1:7, sep=""), type="Diag",
##D                         as.mxMatrix=FALSE )
##D 
##D ## Create Smatrix
##D S1 <- bdiagMat(list(Psi, Phi))
##D ## dimnames(S1)[[1]] <- dimnames(S1)[[2]] <- c(paste("x",1:7,sep=""),
##D ##                                             paste("f",1:3,sep=""))
##D ## S1
##D S1 <- as.mxMatrix(S1)
##D 
##D ## Factor loadings
##D Lambda <- create.mxMatrix( c(".3*f1x1",".3*f1x2",".3*f1x3",rep(0,7),
##D                              ".3*f2x4",".3*f2x5",rep(0,7),".3*f3x6",".3*f3x7"),
##D                            type="Full", ncol=3, nrow=7, as.mxMatrix=FALSE )
##D Zero1 <- matrix(0, nrow=7, ncol=7)
##D Zero2 <- matrix(0, nrow=3, ncol=10)
##D 
##D ## Create Amatrix
##D A1 <- rbind( cbind(Zero1, Lambda),
##D              Zero2 )
##D ## dimnames(A1)[[1]] <- dimnames(A1)[[2]] <- c(paste("x",1:7,sep=""),
##D ##                                             paste("f",1:3,sep=""))
##D ## A1
##D A1 <- as.mxMatrix(A1)
##D 
##D ## Create Fmatrix
##D F1 <- create.Fmatrix(c(rep(1,7), rep(0,3)))
##D 
##D #### Fixed-effects model: Stage 2 analysis
##D fixed2 <- tssem2(fixed1, Amatrix=A1, Smatrix=S1, Fmatrix=F1,
##D                  intervals.type="LB")
##D summary(fixed2)
## End(Not run)

[Package metaSEM version 0.9.4 Index]