issp89 {metaSEM}	R Documentation

Data Set from Cheung and Chan (2005; 2009)

Description

Eleven covariance matrices on work-related attitudes were extracted from the Inter-University Consortium for Political and Social Research (1989). Nine variables were selected by Cheung and Chan (2005; 2009) for demonstration purposes. They were grouped into three constructs: Job Prospects measured by job security (JP1), income (JP2), and advancement opportunity (JP3); Job Nature measured by interesting job (JN1), independent work (JN2), help other people (JN3), and useful to society (JN4); and Time Demand measured by flexible working hours (TD1) and lots of leisure time (TD2).

Usage

data(issp89)

Details

A list of data with the following structure:

data

A list of 11 studies of covariance matrices

n

A vector of sample sizes

Author(s)

Mike W.-L. Cheung <mikewlcheung@nus.edu.sg>

Source

Inter-University Consortium for Political and Social Research. (1989). International Social Survey Program: Work orientation. Ann Arbor, MI: Author.

References

Cheung, M. W.-L., & Chan, W. (2005). Meta-analytic structural equation modeling: A two-stage approach. Psychological Methods, 10, 40-64.

Cheung, M. W.-L., & Chan, W. (2009). A two-stage approach to synthesizing covariance matrices in meta-analytic structural equation modeling. Structural Equation Modeling, 16, 28-53.

See Also

issp05

Examples

## Not run: 
##D data(issp89)
##D 
##D #### Analysis of correlation structure in Cheung and Chan (2005)
##D #### Fixed-effects model: Stage 1 analysis
##D cor1 <- tssem1(issp89$data, issp89$n, method="FEM", cor.analysis=TRUE)
##D summary(cor1)
##D   
##D ## Prepare a model implied matrix
##D ## Factor correlation matrix
##D Phi <- create.mxMatrix( c("0.3*corf2f1","0.3*corf3f1","0.3*corf3f2"),
##D                         type="Stand", as.mxMatrix=FALSE )
##D ## Error variances
##D Psi <- create.mxMatrix( paste("0.2*e", 1:9, sep=""), type="Diag",
##D                         as.mxMatrix=FALSE )
##D 
##D ## Create Smatrix
##D S1 <- bdiagMat(list(Psi, Phi))
##D ## dimnames(S1)[[1]] <- dimnames(S1)[[2]] <- c(paste("x",1:9,sep=""),
##D ##                                             paste("f",1:3,sep=""))
##D ## S1
##D S1 <- as.mxMatrix(S1)
##D 
##D ## Factor loadings
##D Lambda <- create.mxMatrix( c(".3*f1x1",".3*f1x2",".3*f1x3",rep(0,9),
##D                              ".3*f2x4",".3*f2x5",".3*f2x6",".3*f2x7",
##D                              rep(0,9),".3*f3x8",".3*f3x9"), type="Full",
##D                              ncol=3, nrow=9, as.mxMatrix=FALSE )
##D Zero1 <- matrix(0, nrow=9, ncol=9)
##D Zero2 <- matrix(0, nrow=3, ncol=12)
##D 
##D ## Create Amatrix
##D A1 <- rbind( cbind(Zero1, Lambda),
##D              Zero2 )
##D ## dimnames(A1)[[1]] <- dimnames(A1)[[2]] <- c(paste("x",1:9,sep=""),
##D ##                                             paste("f",1:3,sep=""))
##D ## A1
##D A1 <- as.mxMatrix(A1)
##D 
##D ## Create Fmatrix
##D F1 <- create.Fmatrix(c(rep(1,9), rep(0,3)))
##D   
##D #### Fixed-effects model: Stage 2 analysis
##D cor2 <- tssem2(cor1, Amatrix=A1, Smatrix=S1, Fmatrix=F1, intervals.type="LB")
##D summary(cor2)
##D 
##D #### Analysis of covariance structure in Cheung and Chan (2009)
##D #### Fixed-effects model: Stage 1 analysis
##D cov1 <- tssem1(issp89$data, issp89$n, method="FEM", cor.analysis=FALSE)
##D summary(cov1)
##D   
##D #### Fixed-effects model: Stage 2 analysis
##D cov2 <- tssem2(cov1, Amatrix=A1, Smatrix=S1, Fmatrix=F1)              
##D summary(cov2)
## End(Not run)

[Package metaSEM version 0.9.4 Index]