lavaan2RAM {metaSEM}	R Documentation

Convert lavaan models to RAM models

Description

It converts models specified in lavaan model syntax to RAM models.

Usage

 
lavaan2RAM(model, obs.variables = NULL, A.notation = "ON", 
           S.notation = "WITH", auto.var = TRUE, std.lv = TRUE, ...) 

Arguments

model
A model using the lavaan model syntax.

obs.variables
A character vector of the observed variables. The observed variables in the RAM specification will follow the order specified in obs.variables. It is important to check whether the order of the observed variables matches the order in the dataset.

A.notation
A character to be used in the A matrix if the labels are not included in the lavaan model. For example, the label will be "yONx" for regressing "y" on "x".

S.notation
A character to be used in the S matrix if the labels are not included in the lavaan model. For example, the label will be "yWITHx" for the covariance between "y" with "x" and "yWITHy" for the (error) variance of "y".

auto.var
Logical. If TRUE, the residual variances and the variances of exogenous latent variables are included in the model and set free. See lavaanify.

std.lv
Logical. If TRUE, the metric of each latent variable is determined by fixing their variances to 1.0. If FALSE, the metric of each latent variable is determined by fixing the factor loading of the first indicator to 1.0. See lavaanify.

...
Futher arguments to be passed to lavaanify

. Please note that fixed.x is set at FALSE. Thus it cannot be passed to ....

Details

It uses the lavaanify to do the conversion.

Value

A list of RAM specification with A, S, and F matrices.

Author(s)

Mike W.-L. Cheung <mikewlcheung@nus.edu.sg>

See Also

ramModel, Becker92, Becker09, Digman97, Hunter83

Examples

## Regression model on correlation matrix 
model1 <- "y~x1+x2+x3 
           ## Fix the independent variables at 1 
           x1~~1*x1 
           x2~~1*x2 
           x3~~1*x3 
           ## Declare the correlations among the independent variables 
           x1~~x2 
           x1~~x3 
           x2~~x3" 

## Compare the arrangements of variables with and without 
## specifying the obs.variables arguments.  
lavaan2RAM(model1, obs.variables=c("y", "x1", "x2", "x3")) 

## $A
##    y   x1        x2        x3       
## y  "0" "0*yONx1" "0*yONx2" "0*yONx3"
## x1 "0" "0"       "0"       "0"      
## x2 "0" "0"       "0"       "0"      
## x3 "0" "0"       "0"       "0"      
## 
## $S
##    y          x1           x2           x3          
## y  "0*yWITHy" "0"          "0"          "0"         
## x1 "0"        "1"          "0*x1WITHx2" "0*x1WITHx3"
## x2 "0"        "0*x1WITHx2" "1"          "0*x2WITHx3"
## x3 "0"        "0*x1WITHx3" "0*x2WITHx3" "1"         
## 
## $F
##    y x1 x2 x3
## y  1  0  0  0
## x1 0  1  0  0
## x2 0  0  1  0
## x3 0  0  0  1

## Two-factor CFA model 
model2 <- "f1=~x1+x2+x3 
           f2=~x4+x5+x6 
           ## Declare the correlation between f1 and f2 
           ## and label it with cor_f1f2 
           f1~~cor_f1f2*f2" 

lavaan2RAM(model2) 

## $A
##    x1  x2  x3  x4  x5  x6  f1         f2        
## x1 "0" "0" "0" "0" "0" "0" "0*x1ONf1" "0"       
## x2 "0" "0" "0" "0" "0" "0" "0*x2ONf1" "0"       
## x3 "0" "0" "0" "0" "0" "0" "0*x3ONf1" "0"       
## x4 "0" "0" "0" "0" "0" "0" "0"        "0*x4ONf2"
## x5 "0" "0" "0" "0" "0" "0" "0"        "0*x5ONf2"
## x6 "0" "0" "0" "0" "0" "0" "0"        "0*x6ONf2"
## f1 "0" "0" "0" "0" "0" "0" "0"        "0"       
## f2 "0" "0" "0" "0" "0" "0" "0"        "0"       
## 
## $S
##    x1           x2           x3           x4           x5          
## x1 "0*x1WITHx1" "0"          "0"          "0"          "0"         
## x2 "0"          "0*x2WITHx2" "0"          "0"          "0"         
## x3 "0"          "0"          "0*x3WITHx3" "0"          "0"         
## x4 "0"          "0"          "0"          "0*x4WITHx4" "0"         
## x5 "0"          "0"          "0"          "0"          "0*x5WITHx5"
## x6 "0"          "0"          "0"          "0"          "0"         
## f1 "0"          "0"          "0"          "0"          "0"         
## f2 "0"          "0"          "0"          "0"          "0"         
##    x6           f1           f2          
## x1 "0"          "0"          "0"         
## x2 "0"          "0"          "0"         
## x3 "0"          "0"          "0"         
## x4 "0"          "0"          "0"         
## x5 "0"          "0"          "0"         
## x6 "0*x6WITHx6" "0"          "0"         
## f1 "0"          "1"          "0*cor_f1f2"
## f2 "0"          "0*cor_f1f2" "1"         
## 
## $F
##    x1 x2 x3 x4 x5 x6 f1 f2
## x1  1  0  0  0  0  0  0  0
## x2  0  1  0  0  0  0  0  0
## x3  0  0  1  0  0  0  0  0
## x4  0  0  0  1  0  0  0  0
## x5  0  0  0  0  1  0  0  0
## x6  0  0  0  0  0  1  0  0

[Package metaSEM version 0.9.8 Index]