meta {metaSEM}	R Documentation

Univariate and Multivariate Meta-Analysis with Maximum Likelihood Estimation

Description

It conducts univariate and multivariate meta-analysis with maximum likelihood estimation method. Mixed-effects meta-analysis can be conducted by including study characteristics as predictors. Equality constraints on intercepts, regression coefficients and variance components can be easily imposed by setting the same labels on the parameter estimates.

Usage

meta(y, v, x, data, intercept.constraints = NULL, coef.constraints = NULL,
     RE.constraints = NULL, RE.startvalues=0.1, RE.lbound = 1e-10,
     intervals.type = c("z", "LB"), I2="I2q", R2=TRUE,
     model.name="Meta analysis with ML", suppressWarnings = TRUE,
     silent = TRUE, run = TRUE, ...)

Arguments

y
A vector of effect size for univariate meta-analysis or a k x p matrix of effect sizes for multivariate meta-analysis where k is the number of studies and p is the number of effect sizes.

v
A vector of the sampling variance of the effect size for univarite meta-analysis or a k x p* matrix of the sampling covariance matrix of the effect sizes for multivariate meta-analysis where p* = p(p+1)/2. It is arranged by column major as used by vech.

x
A predictor or a k x m matrix of predictors where m is the number of predictors.

data
An optional data frame containing the variables in the model.

intercept.constraints
A 1 x p matrix specifying whether the intercepts of the effect sizes are fixed or free. If the input is not a matrix, the input is converted into a 1 x p matrix with t(as.matrix(intercept.constraints)). The default is that the intercepts are free. When there is no predictor, these intercepts are the same as the pooled effect sizes. The format of this matrix follows as.mxMatrix. The intercepts can be constrained equally by using the same labels.

coef.constraints
A p x m matrix specifying how the predictors predict the effect sizes. If the input is not a matrix, it is converted into a matrix by as.matrix(). The default is that all m predictors predict all p effect sizes. The format of this matrix follows as.mxMatrix. The regression coefficients can be constrained equally by using the same labels.

RE.constraints
A p x p matrix specifying the variance componets of the random effects. If the input is not a matrix, it is converted into a matrix by as.matrix(). The default is that all covariance/variance components are free. The format of this matrix follows as.mxMatrix. Elements of the variance components can be constrained equally by using the same labels. If a zero matrix is specified, it becomes a fixed-effects meta-analysis.

RE.startvalues
A vector of p starting values on the diagonals of the variance component of the random effects. If only one scalar is given, it will be duplicated across the diagonals. Starting values for the off-diagonals of the variance component are all 0. A p x p symmetric matrix of starting values is also accepted.

RE.lbound
A vector of p lower bounds on the diagonals of the variance component of the random effects. If only one scalar is given, it will be duplicated across the diagonals. Lower bounds for the off-diagonals of the variance component are set at NA. A p x p symmetric matrix of the lower bounds is also accepted.

intervals.type
Either z (default if missing) or LB. If it is z, it calculates the 95% Wald confidence intervals (CIs) based on the z statistic. If it is LB, it calculates the 95% likelihood-based CIs on the parameter estimates. Note that the z values and their associated p values are based on the z statistic. They are not related to the likelihood-based CIs.

I2
Possible options are "I2q", "I2hm" and "I2am". They represent the I2 calculated by using a typical within-study sampling variance from the Q statistic, the harmonic mean and the arithmatic mean of the within-study sampling variances (Xiong, Miller, & Morris, 2010). More than one options are possible. If intervals.type="LB", 95% confidence intervals on the heterogeneity indices will be constructed.

R2
Logical. If TRUE and there are predictors, R2 is calculated (Raudenbush, 2009).

model.name
A string for the model name in mxModel.

suppressWarnings
Logical. If TRUE, warnings are suppressed. Argument to be passed to mxRun.

silent
Logical. Argument to be passed to mxRun

run
Logical. If FALSE, only return the mx model without running the analysis.

...
Futher arguments to be passed to mxRun

Value

An object of class meta with a list of

call
Object returned by match.call

data
A data matrix of y, v and x

no.y
No. of effect sizes

no.x
No. of predictors

miss.x
A vector indicating whether the predictors are missing. Studies will be removed before the analysis if they are TRUE

I2
Types of I2 calculated

R2
Logical

mx.fit
A fitted object returned from mxRun

mx0.fit
A fitted object without any predictor returned from mxRun

Note

Missing values (NA) in y and their related elements in v will be removed automatically. When there are missing values in v but not in y, missing values will be replaced by 1e5. Effectively, these effect sizes will have little impact on the analysis.

Author(s)

Mike W.-L. Cheung <mikewlcheung@nus.edu.sg>

References

Cheung, M. W.-L. (2008). A model for integrating fixed-, random-, and mixed-effects meta-analyses into structural equation modeling. Psychological Methods, 13, 182-202.

Cheung, M. W.-L. (2009). Constructing approximate confidence intervals for parameters with structural equation models. Structural Equation Modeling, 16, 267-294.

Cheung, M. W.-L. (2013). Multivariate meta-analysis as structural equation models. Structural Equation Modeling, 20, 429-454.

Hardy, R. J., & Thompson, S. G. (1996). A likelihood approach to meta-analysis with random effects. Statistics in Medicine, 15, 619-629.

Neale, M. C., & Miller, M. B. (1997). The use of likelihood-based confidence intervals in genetic models. Behavior Genetics, 27, 113-120.

Raudenbush, S. W. (2009). Analyzing effect sizes: random effects models. In H. M. Cooper, L. V. Hedges, & J. C. Valentine (Eds.), The handbook of research synthesis and meta-analysis (2nd ed., pp. 295-315). New York: Russell Sage Foundation.

Xiong, C., Miller, J. P., & Morris, J. C. (2010). Measuring study-specific heterogeneity in meta-analysis: application to an antecedent biomarker study of alzheimer's disease. Statistics in Biopharmaceutical Research, 2(3), 300-309. doi:10.1198/sbr.2009.0067

See Also

reml, Hox02, Berkey98, wvs94a

[Package metaSEM version 0.9.4 Index]