meta2semPlot {metaSEM}	R Documentation

Convert metaSEM objects into semPlotModel objects for plotting

Description

It converts objects in class wls into objects of class semPlotModel.

Usage

meta2semPlot(object, manNames = NULL, latNames = NULL, labels = c("labels", "RAM"), ...)

Arguments

object
An object of class wls returned from wls() or tssem2().

manNames
A character vector of the manifest names. The program will try to get it from the object if it is not given.

latNames
A character vector of the latent names. The program will create it by using "L1", "L2", etc if it is not given.

labels
Either labels (default if missing) or RAM. If labels, the labels of the parameters are used in plotting. If RAM, the RAM notations are used in plotting.

...
Further arguments to be passed to ramModel

Details

It uses the ramModel() to do the conversion.

Value

A "semPlotModel" object.

Author(s)

Mike W.-L. Cheung <mikewlcheung@nus.edu.sg>

See Also

ramModel, Becker92, Becker09, Digman97, Hunter83

[Package metaSEM version 0.9.4 Index]