meta3 {metaSEM}	R Documentation

Three-Level Univariate Meta-Analysis with Maximum Likelihood Estimation

Description

It conducts three-level univariate meta-analysis with maximum likelihood estimation method. Mixed-effects meta-analysis can be conducted by including study characteristics as predictors. Equality constraints on the intercepts, regression coefficients and variance components on the level-2 and on the level-3 can be easily imposed by setting the same labels on the parameter estimates.

Usage

meta3(y, v, cluster, x, data, intercept.constraints = NULL,
      coef.constraints = NULL , RE2.constraints = NULL,
      RE2.lbound = 1e-10, RE3.constraints = NULL, RE3.lbound = 1e-10,
      intervals.type = c("z", "LB"), I2="I2q",
      R2=TRUE, model.name = "Meta analysis with ML",
      suppressWarnings = TRUE, silent = TRUE, run = TRUE, ...)
meta3X(y, v, cluster, x2, x3, av2, av3, data, intercept.constraints=NULL,
       coef.constraints=NULL, RE2.constraints=NULL, RE2.lbound=1e-10,
       RE3.constraints=NULL, RE3.lbound=1e-10, intervals.type=c("z", "LB"),
       R2=TRUE, model.name="Meta analysis with ML",
       suppressWarnings=TRUE, silent = TRUE, run = TRUE, ...) 

Arguments

y
A vector of k studies of effect size.

v
A vector of k studies of sampling variance.

cluster
A vector of k characters or numbers indicating the clusters.

x
A predictor or a k x m matrix of level-2 and level-3 predictors where m is the number of predictors.

x2
A predictor or a k x m matrix of level-2 predictors where m is the number of predictors.

x3
A predictor or a k x m matrix of level-3 predictors where m is the number of predictors.

av2
A predictor or a k x m matrix of level-2 auxiliary variables where m is the number of variables.

av3
A predictor or a k x m matrix of level-3 auxiliary variables where m is the number of variables.

data
An optional data frame containing the variables in the model.

intercept.constraints
A 1 x 1 matrix specifying whether the intercept of the effect size is fixed or constrained. The format of this matrix follows as.mxMatrix. The intercept can be constrained with other parameters by using the same label.

coef.constraints
A 1 x m matrix specifying how the level-2 and level-3 predictors predict the effect sizes. If the input is not a matrix, it is converted into a matrix by as.matrix(). The default is that all m predictors predict the effect size. The format of this matrix follows as.mxMatrix. The regression coefficients can be constrained equally by using the same labels.

RE2.constraints
A scalar or a 1 x 1 matrix specifying the variance componets of the random effects. The default is that the variance components is free. The format of this matrix follows as.mxMatrix. Elements of the variance components can be constrained equally by using the same label.

RE2.lbound
A scalar or a 1 x 1 matrix of lower bound on the level-2 variance component of the random effects.

RE3.constraints
A scalar of a 1 x 1 matrix specifying the variance componets of the random effects at level-3. The default is that the variance components is free. The format of this matrix follows as.mxMatrix. Elements of the variance components can be constrained equally by using the same label.

RE3.lbound
A scalar or a 1 x 1 matrix of lower bound on the level-3 variance component of the random effects.

intervals.type
Either z (default if missing) or LB. If it is z, it calculates the 95% Wald confidence intervals (CIs) based on the z statistic. If it is LB, it calculates the 95% likelihood-based CIs on the parameter estimates. Note that the z values and their associated p values are based on the z statistic. They are not related to the likelihood-based CIs.

I2
Possible options are "I2q", "I2hm", "I2am" and "ICC". They represent the I2 calculated by using a typical within-study sampling variance from the Q statistic, the harmonic mean, the arithmatic mean of the within-study sampling variances, and the intra-class correlation. More than one options are possible. If intervals.type="LB", 95% confidence intervals on the heterogeneity indices will be constructed.

R2
Logical. If TRUE and there are predictors, R2 is calculated.

model.name
A string for the model name in mxModel.

suppressWarnings
Logical. If TRUE, warnings are suppressed. Argument to be passed to mxRun.

silent
Logical. Argument to be passed to mxRun

run
Logical. If FALSE, only return the mx model without running the analysis.

...
Futher arguments to be passed to mxRun

Details

y_{ij} = β_0 + \mathbf{β'}*\mathbf{x}_{ij} + u_{(2)ij} + u_{(3)j} + e_{ij}

where y_{ij} is the effect size for the ith study in the jth cluster, β_0 is the intercept, \mathbf{β} is the regression coefficients, \mathbf{x}_{ij} is a vector of predictors, u_{(2)ij}~ N(0, tau^2_2) and u_{(3)j}~ N(0, tau^2_3) are the level-2 and level-3 heterogeneity variances, respectively, and e_{ij}~ N(0, v_{ij}) is the conditional known sampling variance.

meta3() does not differentiate between level-2 or level-3 variables in x since both variables are treated as a design matrix. When there are missing values in x, the data will be deleted. meta3X() treats the predictors x2 and x3 as level-2 and level-3 variables. Thus, their means and covariance matrix will be estimated. Missing values in x2 and x3 will be handled by (full information) maximum likelihood (FIML) in meta3X(). Moreover, auxiliary variables av2 at level-2 and av3 at level-3 may be included to improve the estimation. Although meta3X() is more flexible in handling missing covariates, it is more likely to encounter estimation problems.

Author(s)

Mike W.-L. Cheung <mikewlcheung@nus.edu.sg>

References

Cheung, M. W.-L. (2014). Modeling dependent effect sizes with three-level meta-analyses: A structural equation modeling approach. Psychological Methods, 19, 211-229.

Enders, C. K. (2010). Applied missing data analysis. New York: Guilford Press.

Graham, J. (2003). Adding missing-data-relevant variables to FIML-based structural equation models. Structural Equation Modeling: A Multidisciplinary Journal, 10(1), 80-100.

Konstantopoulos, S. (2011). Fixed effects and variance components estimation in three-level meta-analysis. Research Synthesis Methods, 2, 61-76.

See Also

reml3, Cooper03, Bornmann07

[Package metaSEM version 0.9.5-2 Index]