pattern.n {metaSEM}	R Documentation

Display the Accumulative Sample Sizes for the Covariance Matrix

Description

It displays the accumulative sample sizes for the covariance matrix.

Usage

pattern.n(x, n)

Arguments

x
A list of square matrices

n
A vector of sample sizes.

Value

A square matrix of the accumulative sample sizes of the input matrices.

Author(s)

Mike W.-L. Cheung <mikewlcheung@nus.edu.sg>

Examples

## Show the pattern of missing data
pattern.n(Hunter83$data, Hunter83$n)

##             Ability Knowledge Work sample Supervisor
## Ability        3815      3372        3281       3605
## Knowledge      3372      3532        2998       3322
## Work sample    3281      2998        3441       3231
## Supervisor     3605      3322        3231       3765

#             Ability Knowledge Work sample Supervisor
# Ability        3815      3372        3281       3605
# Knowledge      3372      3532        2998       3322
# Work sample    3281      2998        3441       3231
# Supervisor     3605      3322        3231       3765

[Package metaSEM version 0.9-2 Index]