pattern.na {metaSEM}	R Documentation

Display the Pattern of Missing Data of a List of Square Matrices

Description

It displays the pattern of missing data (or pattern of data that are present) of a list of square matrices with the same dimensions.

Usage

pattern.na(x, show.na = TRUE)

Arguments

x
A list of square matrices

show.na
If it is TRUE, it shows the pattern of missing data. If it is FALSE, it shows the pattern of data that are present.

Value

A square matrix of numerical values with the same dimensions of the input matrices.

Author(s)

Mike W.-L. Cheung <mikewlcheung@nus.edu.sg>

Examples

## Show the pattern of missing data
pattern.na(Hunter83$data, show.na=TRUE)

##             Ability Knowledge Work sample Supervisor
## Ability           1         3           3          2
## Knowledge         3         2           4          3
## Work sample       3         4           2          3
## Supervisor        2         3           3          1

#             Ability Knowledge Work sample Supervisor
# Ability           1         3           3          2
# Knowledge         3         2           4          3
# Work sample       3         4           2          3
# Supervisor        2         3           3          1

## Show the pattern of data that are present
pattern.na(Hunter83$data, show.na=FALSE)

##             Ability Knowledge Work sample Supervisor
## Ability          13        11          11         12
## Knowledge        11        12          10         11
## Work sample      11        10          12         11
## Supervisor       12        11          11         13

#             Ability Knowledge Work sample Supervisor
# Ability          13        11          11         12
# Knowledge        11        12          10         11
# Work sample      11        10          12         11
# Supervisor       12        11          11         13

[Package metaSEM version 0.9.3-1 Index]