reml {metaSEM}	R Documentation

Estimate Variance Components with Restricted (Residual) Maximum Likelihood Estimation

Description

It estimates the variance components of random-effects in univariate and multivariate meta-analysis with restricted (residual) maximum likelihood (REML) estimation method.

Usage

reml(y, v, x, data, RE.constraints = NULL, RE.startvalues = 0.1,
     RE.lbound = 1e-10, intervals.type = c("z", "LB"),
     model.name="Variance component with REML",
     suppressWarnings = TRUE, silent = TRUE, run = TRUE, ...)

Arguments

y
A vector of effect size for univariate meta-analysis or a k x p matrix of effect sizes for multivariate meta-analysis where k is the number of studies and p is the number of effect sizes.

v
A vector of the sampling variance of the effect size for univarite meta-analysis or a k x p* matrix of the sampling covariance matrix of the effect sizes for multivariate meta-analysis where p* = p(p+1)/2. It is arranged by column major as used by vech.

x
A predictor or a k x m matrix of predictors where m is the number of predictors.

data
An optional data frame containing the variables in the model.

RE.constraints
A p x p matrix specifying the variance componets of the random effects. If the input is not a matrix, it is converted into a matrix by as.matrix(). The default is that all covariance/variance components are free. The format of this matrix follows as.mxMatrix. Elements of the variance components can be constrained equally by using the same labels. If a zero matrix is specified, it becomes a fixed-effects meta-analysis.

RE.startvalues
A vector of p starting values on the diagonals of the variance component of the random effects. If only one scalar is given, it will be repeated across the diagonals. Starting values for the off-diagonals of the variance component are all 0. A p x p symmetric matrix of starting values is also accepted.

RE.lbound
A vector of p lower bounds on the diagonals of the variance component of the random effects. If only one scalar is given, it will be repeated across the diagonals. Lower bounds for the off-diagonals of the variance component are set at NA. A p x p symmetric matrix of the lower bounds is also accepted.

intervals.type
Either z (default if missing) or LB. If it is z, it calculates the 95% Wald confidence intervals (CIs) based on the z statistic. If it is LB, it calculates the 95% likelihood-based CIs on the parameter estimates. Note that the z values and their associated p values are based on the z statistic. They are not related to the likelihood-based CIs.

model.name
A string for the model name in mxModel.

suppressWarnings
Logical. If TRUE, warnings are suppressed. Argument to be passed to mxRun.

silent
Logical. Argument to be passed to mxRun

run
Logical. If FALSE, only return the mx model without running the analysis.

...
Futher arguments to be passed to mxRun

Details

Restricted (residual) maximum likelihood obtains the parameter estimates on the transformed data that do not include the fixed-effects parameters. A transformation matrix M=I-X(X'X)^{-1}X' is created based on the design matrix X which is just a column vector when there is no predictor in x. The last N redundant rows of M is removed where N is the rank of X. After pre-multiplying by M on y, the parameters of fixed-effects are removed from the model. Thus, only the parameters of random-effects are estimated.

An alternative but equivalent approach is to minimize the -2*log-likelihood function:

log(det|V+T^2|)+log(det|X'(V+T^2)^{-1}X|)+(y-Xα)'(V+T^2)^{-1}(y-X*α)

where V is the known conditional sampling covariance matrix of y, T^2 is the variance component of the random effects, and \hat{α}=(t(X)(V+T^2)^{-1}X)^{-1}t(X)(V+T^2)^{-1}y. reml() minimizes the above likelihood function to obtain the parameter estimates.

Value

An object of class reml with a list of

call
Object returned by match.call

data
A data matrix of y, v and x

no.y
No. of effect sizes

no.x
No. of predictors

miss.vec
A vector indicating missing data. Studies will be removed before the analysis if they are TRUE

mx.fit
A fitted object returned from mxRun

Note

reml is more computional intensive than meta. Moreover, reml is more likely to encounter errors during optimization. Since a likelihood function is directly employed to obtain the parameter estimates, there is no number of studies and number of observed statistics returned by mxRun. Ad-hoc steps are used to modify mx.fit@runstate$objectives[[1]]@numObs and mx.fit@runstate$objectives[[1]]@numStats.

Author(s)

Mike W.-L. Cheung <mikewlcheung@nus.edu.sg>

References

Cheung, M. W.-L. (2013). Implementing restricted maximum likelihood estimation in structural equation models. Structural Equation Modeling, 20(1), 157-167.

Mehta, P. D., & Neale, M. C. (2005). People Are Variables Too: Multilevel Structural Equations Modeling. Psychological Methods, 10(3), 259-284.

Searle, S. R., Casella, G., & McCulloch, C. E. (1992). Variance components. New York: Wiley.

Viechtbauer, W. (2005). Bias and efficiency of meta-analytic variance estimators in the random-effects model. Journal of Educational and Behavioral Statistics, 30(3), 261-293.

See Also

meta, reml3, Hox02, Berkey98

[Package metaSEM version 0.9-1 Index]