reml3 {metaSEM}	R Documentation

Estimate Variance Components in Three-Level Univariate Meta-Analysis with Restricted (Residual) Maximum Likelihood Estimation

Description

It estimates the variance components of random-effects in three-level univariate meta-analysis with restricted (residual) maximum likelihood (REML) estimation method.

Usage

 
reml3(y, v, cluster, x, data, RE2.startvalue=0.1, RE2.lbound=1e-10, 
      RE3.startvalue=RE2.startvalue, RE3.lbound=RE2.lbound, RE.equal=FALSE, 
      intervals.type=c("z", "LB"), model.name="Variance component with REML", 
      suppressWarnings=TRUE, silent=TRUE, run=TRUE, ...)  

Arguments

y
A vector of k studies of effect size.

v
A vector of k studies of sampling variance.

cluster
A vector of k characters or numbers indicating the clusters.

x
A predictor or a k x m matrix of level-2 and level-3 predictors where m is the number of predictors.

data
An optional data frame containing the variables in the model.

RE2.startvalue
Starting value for the level-2 variance.

RE2.lbound
Lower bound for the level-2 variance.

RE3.startvalue
Starting value for the level-3 variance.

RE3.lbound
Lower bound for the level-3 variance.

RE.equal
Logical. Whether the variance components at level-2 and level-3 are constrained equally.

intervals.type
Either z (default if missing) or LB. If it is z, it calculates the 95% Wald confidence intervals (CIs) based on the z statistic. If it is LB, it calculates the 95% likelihood-based CIs on the parameter estimates. Note that the z values and their associated p values are based on the z statistic. They are not related to the likelihood-based CIs.

model.name
A string for the model name in mxModel.

suppressWarnings
Logical. If TRUE, warnings are suppressed. Argument to be passed to mxRun.

silent
Logical. Argument to be passed to mxRun

run
Logical. If FALSE, only return the mx model without running the analysis.

...
Futher arguments to be passed to mxRun

Details

Restricted (residual) maximum likelihood obtains the parameter estimates on the transformed data that do not include the fixed-effects parameters. A transformation matrix M=I-X(X'X)^{-1}X' is created based on the design matrix X which is just a column vector when there is no predictor in x. The last N redundant rows of M is removed where N is the rank of X. After pre-multiplying by M on y, the parameters of fixed-effects are removed from the model. Thus, only the parameters of random-effects are estimated.

An alternative but equivalent approach is to minimize the -2*log-likelihood function:

log(det|V+T^2|)+log(det|X'(V+T^2)^{-1}X|)+(y-Xα)'(V+T^2)^{-1}(y-X*α)

where V is the known conditional sampling covariance matrix of y, T^2 is the variance component combining level-2 and level-3 random effects, and \hat{α}=(t(X)(V+T^2)^{-1}X)^{-1}t(X)(V+T^2)^{-1}y. reml() minimizes the above likelihood function to obtain the parameter estimates.

Value

An object of class reml with a list of

call
Object returned by match.call

data
A data matrix of y, v and x

mx.fit
A fitted object returned from mxRun

Note

reml is more computional intensive than meta. Moreover, reml is more likely to encounter errors during optimization. Since a likelihood function is directly employed to obtain the parameter estimates, there is no number of studies and number of observed statistics returned by mxRun. Ad-hoc steps are used to modify mx.fit@runstate$objectives[[1]]@numObs and mx.fit@runstate$objectives[[1]]@numStats.

Author(s)

Mike W.-L. Cheung <mikewlcheung@nus.edu.sg>

References

Cheung, M. W.-L. (2013). Implementing restricted maximum likelihood estimation in structural equation models. Structural Equation Modeling, 20(1), 157-167.

Cheung, M. W.-L. (2014). Modeling dependent effect sizes with three-level meta-analyses: A structural equation modeling approach. Psychological Methods, 19, 211-229.

Mehta, P. D., & Neale, M. C. (2005). People Are Variables Too: Multilevel Structural Equations Modeling. Psychological Methods, 10(3), 259-284.

Searle, S. R., Casella, G., & McCulloch, C. E. (1992). Variance components. New York: Wiley.

See Also

meta3, reml, Cooper03, Bornmann07

[Package metaSEM version 0.9.8 Index]