summary {metaSEM}	R Documentation

Summary Method for tssem1, wls, meta and meta3X Objects

Description

It summaries results for either class tssem1FEM, class tssem1FEM.cluster, class tssem1REM, class wls, class wls.cluster, class meta, class meta3X or class reml.

Usage

## S3 method for class 'tssem1FEM'
summary(object, ...)
## S3 method for class 'tssem1FEM.cluster'
summary(object, ...)
## S3 method for class 'tssem1REM'
summary(object, ...)
## S3 method for class 'wls'
summary(object, df.adjustment=0, R=50, ...)
## S3 method for class 'wls.cluster'
summary(object, df.adjustment=0, R=50, ...)
## S3 method for class 'meta'
summary(object, homoStat=TRUE, ...)
## S3 method for class 'meta3X'
summary(object, allX=FALSE, ...)
## S3 method for class 'reml'
summary(object, ...)
## S3 method for class 'tssem1FEM'
print.summary(x, ...)
## S3 method for class 'wls'
print.summary(x, ...)
## S3 method for class 'meta'
print.summary(x, ...)
## S3 method for class 'meta3X'
print.summary(x, ...)
## S3 method for class 'reml'
print.summary(x, ...)

Arguments

object
An object returned from either class tssem1FEM, class tssem1FEM.cluster, class tssem1REM, class wls, class wls.cluster, class meta, class meta3X or class reml.

x
An object returned from either class summary.tssem1FEM, class tssem1FEM.cluster, class summary.wls, class summary.meta, class summary.meta3X or class summary.reml.

homoStat
Logical. Whether to conduct a homogeneity test on the effect sizes.

allX
Logical. Whether to report the predictors and the auxiliary variables.

df.adjustment
Numeric. Adjust the degrees of freedom manually. It may be necessary if the df calculated is incorrect when diag.constraints=TRUE.

R
Positive integer. The number of parameter bootstrap replicates when diag.constraints=TRUE.

...
Further arguments to be passed to printCoefmat

Note

If the OpenMx status1 is either 0 or 1, the estimation is considered fine. If the OpenMx status1 is other values, it indicates estimation problems. Users should refer to http://openmx.psyc.virginia.edu/wiki/errors for more details.

Author(s)

Mike W.-L. Cheung <mikewlcheung@nus.edu.sg>

See Also

tssem1, wls, meta, reml

[Package metaSEM version 0.9-1 Index]