tssem1 {metaSEM}	R Documentation

First Stage of the Two-Stage Structural Equation Modeling (TSSEM)

Description

It conducts the first stage analysis of TSSEM by pooling correlation/covariance matrices. tssem1FEM() and tssem1REM() use fixed- and random-effects models, respectively. tssem1() is a wrapper of these functions.

Usage

tssem1(my.df, n, method=c("FEM","REM"), cor.analysis = TRUE, cluster=NULL,
       RE.type=c("Symm", "Diag", "Zero"), RE.startvalues=0.1, RE.lbound=1e-10, I2="I2q",
       model.name=NULL, suppressWarnings=TRUE, silent=TRUE, run=TRUE, ...)
tssem1FEM(my.df, n, cor.analysis=TRUE, model.name=NULL,
          cluster=NULL, suppressWarnings=TRUE, silent=TRUE, run=TRUE, ...)
tssem1REM(my.df, n, cor.analysis=TRUE, RE.type=c("Symm", "Diag", "Zero"),
          RE.startvalues=0.1, RE.lbound=1e-10, I2="I2q", model.name=NULL,
          suppressWarnings=TRUE, silent=TRUE, run=TRUE, ...)

Arguments

my.df
A list of correlation/covariance matrices

n
A vector of sample sizes

method
Either "FEM" (default if missing) or "REM". If it is "FEM", fixed-effects meta-analysis will be applied. If it is "REM", random-effects meta-analysis will be applied.

cor.analysis
Logical. The output is either a pooled correlation or a covariance matrix.

cluster
A vector of characters or numbers indicating the clusters. Analyses will be conducted for each cluster. It will be ignored when method="REM".

RE.type
Either "Symm", "Diag" or "Zero". If it is "Symm" (default if missing), a symmetric matrix is used for the random effects on the covariances among the correlation (or covariance) vectors. If it is "Diag", a diagonal matrix is used for the random effects meaning that the random effects are independent. If it is "Zero", there is no random effects which is similar to the conventional Generalized Least Squares (GLS) approach to fixed-effects analysis. This argument will be ignored when method="FEM".

RE.startvalues
Starting values on the diagonals of the variance component of the random effects. It will be ignored when method="FEM".

RE.lbound
Lower bounds on the diagonals of the variance component of the random effects. It will be ignored when method="FEM".

I2
Possible options are "I2q", "I2hm" and "I2am". They represent the I2 calculated by using a typical within-study sampling variance from the Q statistic, the harmonic mean and the arithmatic mean of the within-study sampling variances (Xiong, Miller, & Morris, 2010). More than one options are possible. If intervals.type="LB", 95% confidence intervals on the heterogeneity indices will be constructed.

model.name
A string for the model name in mxModel.

suppressWarnings
Logical. If TRUE, warnings are suppressed. Argument to be passed to mxRun.

silent
Logical. Argument to be passed to mxRun

run
Logical. If FALSE, only return the mx model without running the analysis.

...
Further arguments to be passed to mxRun

Value

Either an object of class tssem1FEM for fixed-effects TSSEM, an object of class tssem1FEM.cluster for fixed-effects TSSEM with cluster argument, or an object of class tssem1REM for random-effects TSSEM.

Note

If the cluster argument is used, it returns a list of results based for each cluster.

Author(s)

Mike W.-L. Cheung <mikewlcheung@nus.edu.sg>

References

Cheung, M. W.-L. (2014). Fixed- and random-effects meta-analytic structural equation modeling: Examples and analyses in R. Behavior Research Methods, 46, 29-40.

Cheung, M. W.-L. (2013). Multivariate meta-analysis as structural equation models. Structural Equation Modeling, 20, 429-454.

Cheung, M. W.-L., & Chan, W. (2005). Meta-analytic structural equation modeling: A two-stage approach. Psychological Methods, 10, 40-64.

Cheung, M. W.-L., & Chan, W. (2009). A two-stage approach to synthesizing covariance matrices in meta-analytic structural equation modeling. Structural Equation Modeling, 16, 28-53.

See Also

wls, Cheung09, Becker92, Digman97, issp89, issp05

[Package metaSEM version 0.9-2 Index]