vcov {metaSEM}	R Documentation

Extract Covariance Matrix Parameter Estimates from Various Objects

Description

It extracts the variance-covariance matrix of the parameter estimates from either tssem1FEM, tssem1FEM.cluster, tssem1REM, wls, wls.cluster, meta, meta3X, reml or MxRAMModel objects.

Usage

## S3 method for class 'tssem1FEM'
vcov(object, ...)
## S3 method for class 'tssem1FEM.cluster'
vcov(object, ...)
## S3 method for class 'tssem1REM'
vcov(object, select = c("all", "fixed", "random"), ...)
## S3 method for class 'wls'
vcov(object, R=50, ...)
## S3 method for class 'wls.cluster'
vcov(object, R=50, ...)
## S3 method for class 'meta'
vcov(object, select = c("all", "fixed", "random"), ...)
## S3 method for class 'meta3X'
vcov(object, select = c("all", "fixed", "random", "allX"), ...)
## S3 method for class 'reml'
vcov(object, ...)
## S3 method for class 'MxRAMModel'
vcov(object, ...)

Arguments

object
An object returned from either class tssem1FEM, class tssem1FEM.cluster, class tssem1REM, class wls, class wls.cluster, class meta, class reml or class MxRAMModel

select
Select all for both fixed- and random-effects parameters, fixed for the fixed-effects parameters or random for the random-effects parameters. For meta3X objects, allX is used to extract all parameters including the predictors and auxiliary variables.

R
Positive integer. The number of parameter bootstrap replicates when diag.constraints=TRUE.

...
Further arguments; currently none is used

Value

A variance-covariance matrix of the parameter estimates.

Note

vcov returns NA when diag.constraints=TRUE argument is used in wls objects.

Author(s)

Mike W.-L. Cheung <mikewlcheung@nus.edu.sg>

See Also

tssem1, wls, meta, reml

Examples

## Random-effects meta-analysis
model1 <- meta(y=yi, v=vi, data=Hox02)
vcov(model1)

##              Intercept1     Tau2_1_1
## Intercept1 0.0110460939 0.0004852251
## Tau2_1_1   0.0004852251 0.0054075547

## Fixed-effects only
vcov(model1, select="fixed")

##            Intercept1
## Intercept1 0.01104609

## Random-effects only
vcov(model1, select="random")

##             Tau2_1_1
## Tau2_1_1 0.005407555

[Package metaSEM version 0.9.3-1 Index]