wls {metaSEM}	R Documentation

Conduct a Correlation/Covariance Structure Analysis with WLS

Description

It fits a correlation or covariance structure with weighted least squares (WLS) estimation method where the inverse of the asymptotic covariance matrix is used as the weight matrix. tssem2 conducts the second stage analysis of the two-stage strutural equation modeling (TSSEM). tssem2 is a wrapper of wls.

Usage

wls(Cov, asyCov, n, Amatrix=NULL, Smatrix=NULL, Fmatrix=NULL,
    diag.constraints=FALSE, cor.analysis=TRUE, intervals.type=c("z","LB"),
    mx.algebras=NULL, model.name=NULL, suppressWarnings=TRUE,
    silent=TRUE, run=TRUE, ...)
tssem2(tssem1.obj, Amatrix=NULL, Smatrix=NULL, Fmatrix=NULL,
    diag.constraints=FALSE, intervals.type=c("z", "LB"), mx.algebras=NULL,
    model.name=NULL, suppressWarnings=TRUE, silent=TRUE, run=TRUE, ...) 

Arguments

tssem1.obj
An object of either class tssem1FEM, class tssem1FEM.cluster or class tssem1REM returned from tssem1()

Cov
A p x p sample correlation/covariance matrix where p is the number of variables.

asyCov
A p* x p* asymptotic sampling covariance matrix of either vechs (Cov) or vech (Cov) where p* = p(p-1)/2 for correlation matrix and p* = p(p+1)/2 for covariance matrix.

n
Sample size.

Amatrix
An asymmetric matrix in the RAM approach with MxMatrix-class. If it is NULL, a matrix of zero will be created. If it is a matrix, it will be converted into MxMatrix-class by the as.mxMatrix function.

Smatrix
A symmetric matrix in the RAM approach with MxMatrix-class. If it is a matrix, it will be converted into MxMatrix-class by the as.mxMatrix function.

Fmatrix
A filter matrix in the RAM approach with MxMatrix-class. If it is NULL (the default), an identity matrix with the same dimensions of Cov will be created. If it is a matrix, it will be converted into MxMatrix-class by the as.mxMatrix function. It is not required when there is no latent variable.

diag.constraints
Logical. This argument is ignored when cor.analysis=FALSE. If diag.constraints=TRUE, the diagonals of the model implied matrix will be constrained at 1 by nonlinear constraints. The drawback is that standard error will not be generated. Parametric bootstrap is used to estimate the standard error by drawing samples from N(vech(Cov), asyCov) for covariance analysis and N(vechs(Cov), asyCov) for correlation analysis while asyCov is treated as fixed. This process is computational intensive. A better approach is to request likelihood-based confidence intervals (CIs) by specifying intervals.type="LB".

If diag.constraints=FALSE and cor.analysis=TRUE, the diagonals are automatically constrained as ones by treating the error variances as computed values rather than as parameters. Since the error variances are not parameters, they are not reported. It is usually not necessary to set diag.constraints=TRUE unless there are mediators (variables acting as both a predictor and dependent variable) in the model. See the examples in Hunter83.

cor.analysis
Logical. Analysis of correlation or covariance structure. If cor.analysis=TRUE, vechs is used to vectorize S; otherwise, vech is used to vectorize S.

intervals.type
Either z (default if missing) or LB. If it is z, it calculates the 95% Wald CIs based on the z statistic. If it is LB, it calculates the 95% likelihood-based CIs on the parameter estimates. Please note that the z values and their associated p values are based on the z statistic. They are not related to the likelihood-based CIs.

mx.algebras
A list of mxMatrix or mxAlgebra objects on the Amatrix, Smatrix and Fmatrx. It can be used to define new functions of parameters and their LBCIs. For example, if the regression coefficients to calculate an indirect effect are stored in A[1,2] and A[1,3], we may define list(ind=mxAlgebra(Amatrix[1,2]*Amatrix[1,3], name="ind")) See the examples in Becker92 and Hunter83. It should be noted that Fmatrix, Amatrix, Smatrix, Iden (a p x p identity matrix), sampleS (sample correlation or covariance matrix), impliedS1, impliedS (model implied correlation or covariance matrix), vecS, invAcov, obj, One, select and constraint and Ematrix (computed error variances when diag.constraints=FALSE) have been defined internally. You should not create new matrices using these names.

model.name
A string for the model name in mxModel. If it is missing, the default is "TSSEM2 (or WLS) Analysis of Correlation Structure" for cor.analysis=TRUE and "TSSEM2 (or WLS) Analysis of Covariance Structure" for cor.analysis=FALSE.

suppressWarnings
Logical. If TRUE, warnings are suppressed. Argument to be passed to mxRun.

silent
Logical. Argument to be passed to mxRun

run
Logical. If FALSE, only return the mx model without running the analysis.

...
Futher arguments to be passed to mxRun.

Value

An object of class wls with a list of

call
The matched call

Cov
Input data of either a covariance or correlation matrix

asyCov
Asymptotic covariance matrix of the input data

noObservedStat
Number of observed statistics

n
Sample size

cor.analysis
logical

noConstraints
Number of constraints imposed on S

indepModelChisq
Chi-square statistic of the independent model returned by .indepwlsChisq

indepModelDf
Degrees of freedom of the independent model returned by .indepwlsChisq

mx.fit
A fitted object returned from mxRun

Note

If the input is a list of tssem1.obj, it returns a list of results for each cluster.

Author(s)

Mike W.-L. Cheung <mikewlcheung@nus.edu.sg>

References

Bentler, P.M., & Savalei, V. (2010). Analysis of correlation structures: current status and open problems. In Kolenikov, S., Thombs, L., & Steinley, D. (Eds.). Recent Methodological Developments in Social Science Statistics (pp. 1-36). Hoboken, NJ: Wiley.

Cheung, M. W.-L. (2010). Fixed-effects meta-analyses as multiple-group structural equation models. Structural Equation Modeling, 17, 481-509.

Cheung, M. W.-L. (2014). Fixed- and random-effects meta-analytic structural equation modeling: Examples and analyses in R. Behavior Research Methods, 46, 29-40.

Cheung, M. W.-L., & Chan, W. (2005). Meta-analytic structural equation modeling: A two-stage approach. Psychological Methods, 10, 40-64.

Cheung, M. W.-L., & Chan, W. (2009). A two-stage approach to synthesizing covariance matrices in meta-analytic structural equation modeling. Structural Equation Modeling, 16, 28-53.

Joreskog, K. G., Sorbom, D., Du Toit, S., & Du Toit, M. (1999). LISREL 8: New Statistical Features. Chicago: Scientific Software International.

McArdle, J. J., & MacDonald, R. P. (1984). Some algebraic properties of the Reticular Action Model for moment structures. British Journal of Mathematical and Statistical Psychology, 37, 234-251.

See Also

tssem1, Becker92, Digman97, Hunter83, issp89, issp05

Examples

#### Analysis of correlation structure
R1 <- matrix(c(1.00, 0.22, 0.24, 0.18,
               0.22, 1.00, 0.30, 0.22,
               0.24, 0.30, 1.00, 0.24,
               0.18, 0.22, 0.24, 1.00), ncol=4, nrow=4)
n <- 1000
acovR1 <- asyCov(R1, n)

## One-factor CFA model
(A1 <- cbind(matrix(0, nrow=5, ncol=4),
             matrix(c("0.2*a1","0.2*a2","0.2*a3","0.2*a4",0),
             ncol=1)))

##      [,1] [,2] [,3] [,4] [,5]    
## [1,] "0"  "0"  "0"  "0"  "0.2*a1"
## [2,] "0"  "0"  "0"  "0"  "0.2*a2"
## [3,] "0"  "0"  "0"  "0"  "0.2*a3"
## [4,] "0"  "0"  "0"  "0"  "0.2*a4"
## [5,] "0"  "0"  "0"  "0"  "0"

(S1 <- Diag(c("0.2*e1","0.2*e2","0.2*e3","0.2*e4",1)))

##      [,1]     [,2]     [,3]     [,4]     [,5]
## [1,] "0.2*e1" "0"      "0"      "0"      "0" 
## [2,] "0"      "0.2*e2" "0"      "0"      "0" 
## [3,] "0"      "0"      "0.2*e3" "0"      "0" 
## [4,] "0"      "0"      "0"      "0.2*e4" "0" 
## [5,] "0"      "0"      "0"      "0"      "1"

## The first 4 variables are observed while the last one is latent.
(F1 <- create.Fmatrix(c(1,1,1,1,0), name="F1"))

## FullMatrix 'F1' 
## 
## $labels: No labels assigned.
## 
## $values
##      [,1] [,2] [,3] [,4] [,5]
## [1,]    1    0    0    0    0
## [2,]    0    1    0    0    0
## [3,]    0    0    1    0    0
## [4,]    0    0    0    1    0
## 
## $free: No free parameters.
## 
## $lbound: No lower bounds assigned.
## 
## $ubound: No upper bounds assigned.

wls.fit1 <- wls(Cov=R1, asyCov=acovR1, n=n, Fmatrix=F1, Smatrix=S1, Amatrix=A1,
                 cor.analysis=TRUE, intervals="LB")
summary(wls.fit1)

## 
## Call:
## wls(Cov = R1, asyCov = acovR1, n = n, Amatrix = A1, Smatrix = S1, 
##     Fmatrix = F1, cor.analysis = TRUE, intervals.type = "LB")
## 
## 95% confidence intervals: Likelihood-based statistic
## Coefficients:
##    Estimate Std.Error  lbound  ubound z value Pr(>|z|)
## a1  0.42159        NA 0.34435 0.50017      NA       NA
## a2  0.52376        NA 0.44603 0.60472      NA       NA
## a3  0.57092        NA 0.49204 0.65446      NA       NA
## a4  0.42159        NA 0.34436 0.50018      NA       NA
## 
## Goodness-of-fit indices:
##                                                Value
## Sample size                                1000.0000
## Chi-square of target model                    0.0134
## DF of target model                            2.0000
## p value of target model                       0.9933
## Number of constraints imposed on "Smatrix"    0.0000
## DF manually adjusted                          0.0000
## Chi-square of independence model            207.8699
## DF of independence model                      6.0000
## RMSEA                                         0.0000
## SRMR                                          0.0012
## TLI                                           1.0295
## CFI                                           1.0000
## AIC                                          -3.9866
## BIC                                         -13.8021
## OpenMx status1: 0 ("0" or "1": The optimization is considered fine.
## Other values indicate problems.)

#### Multiple regression analysis
## Variables in R2: y, x1, x2
R2 <- matrix(c(1.00, 0.22, 0.24, 
               0.22, 1.00, 0.30, 
               0.24, 0.30, 1.00, 
               0.18, 0.22, 0.24), ncol=3, nrow=3)
acovR2 <- asyCov(R2, n)

## A2: Regression coefficents
#    y x1 x2
# y  F T  T 
# x1 F F  F 
# x2 F F  F 
(A2 <- mxMatrix("Full", ncol=3, nrow=3, byrow=TRUE,
               free=c(FALSE, rep(TRUE, 2), rep(FALSE, 6)), name="A2"))

## FullMatrix 'A2' 
## 
## $labels: No labels assigned.
## 
## $values
##      [,1] [,2] [,3]
## [1,]    0    0    0
## [2,]    0    0    0
## [3,]    0    0    0
## 
## $free
##       [,1]  [,2]  [,3]
## [1,] FALSE  TRUE  TRUE
## [2,] FALSE FALSE FALSE
## [3,] FALSE FALSE FALSE
## 
## $lbound: No lower bounds assigned.
## 
## $ubound: No upper bounds assigned.

## S2: Covariance matrix of free parameters
#    y x1 x2
# y  T F  F 
# x1 F F  F 
# x2 F T  F
(S2 <- mxMatrix("Symm", ncol=3, nrow=3, values=c(0.2,0,0,1,0.2,1),
               free=c(TRUE,FALSE,FALSE,FALSE,TRUE,FALSE), name="S2"))

## SymmMatrix 'S2' 
## 
## $labels: No labels assigned.
## 
## $values
##      [,1] [,2] [,3]
## [1,]  0.2  0.0  0.0
## [2,]  0.0  1.0  0.2
## [3,]  0.0  0.2  1.0
## 
## $free
##       [,1]  [,2]  [,3]
## [1,]  TRUE FALSE FALSE
## [2,] FALSE FALSE  TRUE
## [3,] FALSE  TRUE FALSE
## 
## $lbound: No lower bounds assigned.
## 
## $ubound: No upper bounds assigned.

## F may be ignored as there is no latent variable.
wls.fit2 <- wls(Cov=R2, asyCov=acovR2, n=n, Amatrix=A2, Smatrix=S2,
                cor.analysis=TRUE, intervals="LB")
summary(wls.fit2)

## 
## Call:
## wls(Cov = R2, asyCov = acovR2, n = n, Amatrix = A2, Smatrix = S2, 
##     cor.analysis = TRUE, intervals.type = "LB")
## 
## 95% confidence intervals: Likelihood-based statistic
## Coefficients:
##              Estimate Std.Error  lbound  ubound z value Pr(>|z|)
## Amatrix[1,2]  0.16264        NA 0.10103 0.22433      NA       NA
## Amatrix[1,3]  0.19121        NA 0.12990 0.25262      NA       NA
## Smatrix[2,3]  0.30000        NA 0.24354 0.35646      NA       NA
## 
## Goodness-of-fit indices:
##                                             Value
## Sample size                                1000.0
## Chi-square of target model                    0.0
## DF of target model                            0.0
## p value of target model                       0.0
## Number of constraints imposed on "Smatrix"    0.0
## DF manually adjusted                          0.0
## Chi-square of independence model            162.1
## DF of independence model                      3.0
## RMSEA                                         0.0
## SRMR                                          0.0
## TLI                                          -Inf
## CFI                                           1.0
## AIC                                           0.0
## BIC                                           0.0
## OpenMx status1: 0 ("0" or "1": The optimization is considered fine.
## Other values indicate problems.)

#### Analysis of covariance structure
R3 <- matrix(c(1.50, 0.22, 0.24, 0.18,
               0.22, 1.60, 0.30, 0.22,
               0.24, 0.30, 1.80, 0.24,
               0.18, 0.22, 0.24, 1.30), ncol=4, nrow=4)
n <- 1000
acovS3 <- asyCov(R3, n, cor.analysis=FALSE)

(A3 <- cbind(matrix(0, nrow=5, ncol=4),
             matrix(c("0.2*a1","0.2*a2","0.2*a3","0.2*a4",0),ncol=1)))

##      [,1] [,2] [,3] [,4] [,5]    
## [1,] "0"  "0"  "0"  "0"  "0.2*a1"
## [2,] "0"  "0"  "0"  "0"  "0.2*a2"
## [3,] "0"  "0"  "0"  "0"  "0.2*a3"
## [4,] "0"  "0"  "0"  "0"  "0.2*a4"
## [5,] "0"  "0"  "0"  "0"  "0"

(S3 <- Diag(c("0.2*e1","0.2*e2","0.2*e3","0.2*e4",1)))

##      [,1]     [,2]     [,3]     [,4]     [,5]
## [1,] "0.2*e1" "0"      "0"      "0"      "0" 
## [2,] "0"      "0.2*e2" "0"      "0"      "0" 
## [3,] "0"      "0"      "0.2*e3" "0"      "0" 
## [4,] "0"      "0"      "0"      "0.2*e4" "0" 
## [5,] "0"      "0"      "0"      "0"      "1"

F3 <- c(TRUE,TRUE,TRUE,TRUE,FALSE)
(F3 <- create.Fmatrix(F3, name="F3", as.mxMatrix=FALSE))

##      [,1] [,2] [,3] [,4] [,5]
## [1,]    1    0    0    0    0
## [2,]    0    1    0    0    0
## [3,]    0    0    1    0    0
## [4,]    0    0    0    1    0

wls.fit3 <- wls(Cov=R3, asyCov=acovS3, n=n, Amatrix=A3, Smatrix=S3,
                Fmatrix=F3, cor.analysis=FALSE)
summary(wls.fit3)

## 
## Call:
## wls(Cov = R3, asyCov = acovS3, n = n, Amatrix = A3, Smatrix = S3, 
##     Fmatrix = F3, cor.analysis = FALSE)
## 
## 95% confidence intervals: z statistic approximation
## Coefficients:
##    Estimate Std.Error   lbound   ubound z value  Pr(>|z|)    
## a1 0.422193  0.063787 0.297173 0.547212  6.6188 3.621e-11 ***
## a2 0.523188  0.069694 0.386590 0.659786  7.5069 6.062e-14 ***
## a3 0.570613  0.074823 0.423963 0.717262  7.6262 2.420e-14 ***
## a4 0.421946  0.060419 0.303526 0.540366  6.9836 2.877e-12 ***
## e1 1.321745  0.073423 1.177837 1.465652 18.0017 < 2.2e-16 ***
## e2 1.326269  0.085075 1.159526 1.493012 15.5895 < 2.2e-16 ***
## e3 1.474395  0.097523 1.283254 1.665536 15.1185 < 2.2e-16 ***
## e4 1.121954  0.065237 0.994093 1.249816 17.1983 < 2.2e-16 ***
## ---
## Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
## 
## Goodness-of-fit indices:
##                                                Value
## Sample size                                1000.0000
## Chi-square of target model                    0.0050
## DF of target model                            2.0000
## p value of target model                       0.9975
## Number of constraints imposed on "Smatrix"    0.0000
## DF manually adjusted                          0.0000
## Chi-square of independence model             78.2765
## DF of independence model                      6.0000
## RMSEA                                         0.0000
## SRMR                                          0.0006
## TLI                                           1.0828
## CFI                                           1.0000
## AIC                                          -3.9950
## BIC                                         -13.8105
## OpenMx status1: 0 ("0" or "1": The optimization is considered fine.
## Other values indicate problems.)

[Package metaSEM version 0.9-1 Index]