wvs94a {metaSEM}	R Documentation

Forty-four Studies from Cheung (2013)

Description

Between 1990 and 1993, 57,561 adults aged 18 and above from 42 nations were interviewed by local academic institutes in Eastern European nations and by professional survey organizations in other nations. Standardized mean difference (SMD) between males and females on life satisfaction and on life control in each country were calculated as the effect sizes. Positive values indicate that males have higher scores than females do.

Usage

data(wvs94a)

Details

The variables are:

country

Country

lifesat

SMD on life satisfaction

lifecon

SMD on life control

lifesat_var

Sampling variance of lifesat

inter_cov

Sampling covariance between lifesat and lifecon

lifecon_var

Sampling variance of lifecon

gnp

Gross National Product

Source

World Values Study Group. (1994). World Values Survey, 1981-1984 and 1990-1993 [Computer file]. Ann Arbor, MI: Inter-university Consortium for Political and Social Research.

References

Au, K., & Cheung, M. W.-L. (2004). Intra-cultural variation and job autonomy in 42 countries. Organization Studies, 25, 1339-1362.

Cheung, M. W.-L. (2013). Multivariate meta-analysis as structural equation models. Structural Equation Modeling, 20, 429-454.

Examples

## Not run: 
##D data(wvs94a)
##D 
##D ## Random-effects model
##D random.ma1 <- meta(y=cbind(lifesat, lifecon),
##D                    v=cbind(lifesat_var, inter_cov, lifecon_var), data=wvs94a,
##D                    model.name="Random effects model")
##D summary(random.ma1)
##D 
##D ## Random-effects model with both population effect sizes fixed at 0
##D random.ma2 <- meta(y=cbind(lifesat, lifecon),
##D                    v=cbind(lifesat_var, inter_cov, lifecon_var), data=wvs94a,
##D                    intercept.constraints=matrix(0, nrow=1, ncol=2),
##D                    model.name="Effect sizes are fixed at 0")
##D summary(random.ma2)
##D 
##D ## Compare the nested models
##D anova(random.ma1, random.ma2)
##D 
##D ## Fixed-effects model by fixing the variance component at 0 
##D fixed.ma <- meta(y=cbind(lifesat, lifecon),
##D                  v=cbind(lifesat_var, inter_cov, lifecon_var), data=wvs94a,
##D                  RE.constraints=matrix(0, ncol=2, nrow=2),
##D                  model.name="Fixed effects model")
##D summary(fixed.ma)
##D 
##D ## Mixed-effects model
##D ## gnp is divided by 10000 and centered by using 
##D ## scale(gnp/10000, scale=FALSE)
##D mixed.ma1 <- meta(y=cbind(lifesat, lifecon),
##D                   v=cbind(lifesat_var, inter_cov, lifecon_var),
##D                   x=scale(gnp/10000, scale=FALSE), data=wvs94a,
##D                   model.name="GNP as a predictor")
##D summary(mixed.ma1)
##D 
##D ## Mixed-effects model with equal regression coefficients
##D mixed.ma2 <- meta(y=cbind(lifesat, lifecon),
##D                   v=cbind(lifesat_var, inter_cov, lifecon_var),
##D                   x=scale(gnp/10000, scale=FALSE), data=wvs94a,
##D                   coef.constraints=matrix(c("0.0*Eq_slope",
##D                                             "0.0*Eq_slope"), nrow=2),
##D                   model.name="GNP as a predictor with equal slope")
##D summary(mixed.ma2)
##D 
##D ## Compare the nested models
##D anova(mixed.ma1, mixed.ma2)
##D 
##D ## Plot the multivariate effect sizes
##D plot(random.ma1, main="Estimated effect sizes and their 95% confidence ellipses",
##D      axis.label=c("Gender difference on life satisfaction",
##D                   "Gender difference on life control"))
## End(Not run)

[Package metaSEM version 0.9.4 Index]