wvs94b {metaSEM}	R Documentation

Forty-four Covariance Matrices on Life Satisfaction, Job Satisfaction, and Job Autonomy

Description

Between 1990 and 1993, 57,561 adults aged 18 and above from 42 nations were interviewed by local academic institutes in Eastern European nations and by professional survey organizations in other nations. The covariance matrices among Life Satisfaction, Job Satisfaction, and Job Autonomy were calculated.

Usage

data(wvs94b)

Details

The variables are:

data

Covariance matrix among Life Satisfaction (LS), Job Satisfaction (JS), and Job Autonomy (JA)

n

Sample size in the country

Source

World Values Study Group. (1994). World Values Survey, 1981-1984 and 1990-1993 [Computer file]. Ann Arbor, MI: Inter-university Consortium for Political and Social Research.

References

Au, K., & Cheung, M. W.-L. (2004). Intra-cultural variation and job autonomy in 42 countries. Organization Studies, 25, 1339-1362.

Cheung, M.W.-L., & Cheung, S.-F. (in press). Random-effects models for meta-analytic structural equation modeling: Review, issues, and illustrations. Research Synthesis Methods.

Examples

## \dontrun{ 
data(wvs94b) 

## Get the indirect and the direct effects and 
## their sampling covariance matrices for each study 
indirect1 <- indirectEffect(wvs94b$data, wvs94b$n) 
indirect1 

##             ind_eff       dir_eff      ind_var   ind_dir_cov      dir_var
## france   0.15978141  0.0827522461 0.0006412924 -4.649454e-04 0.0024374059
## britain  0.14168820 -0.0002263007 0.0002888310 -1.626571e-04 0.0012216599
## wgermany 0.17524960  0.1527287443 0.0003158081 -2.684000e-04 0.0010394339
## italy    0.20321234  0.0296684559 0.0002994225 -1.503761e-04 0.0009550049
## nethland 0.07862686  0.1185534103 0.0004741621 -4.167198e-04 0.0025940000
## denmark  0.08921949  0.0357811049 0.0002580830 -1.633167e-04 0.0015946257
## belgium  0.18234893  0.0406592228 0.0002430000 -1.638546e-04 0.0008544332
## spain    0.15684423  0.0166711391 0.0001385408 -8.530734e-05 0.0005433748
## ireland  0.20451372 -0.0635826273 0.0005530175 -2.113689e-04 0.0016560481
## nireland 0.14117359  0.1715056590 0.0016206049 -1.130507e-03 0.0066716297
## usa      0.21260852  0.0097460305 0.0003352506 -2.565373e-04 0.0010396340
## canada   0.15592998  0.0559244096 0.0002674737 -1.924936e-04 0.0010345513
## japan    0.29279253 -0.1210110214 0.0010138128 -9.983440e-04 0.0020297107
## mexico   0.27017766  0.0899476460 0.0003898565 -2.364167e-04 0.0010783334
## s africa 0.22436811  0.1495906062 0.0002662679 -2.111175e-04 0.0007991767
## hungary  0.11562915  0.1698171076 0.0003129404 -2.141807e-04 0.0015091523
## norway   0.11717420 -0.0022129617 0.0002263148 -1.070791e-04 0.0010993010
## sweden   0.14566226 -0.0139555980 0.0003337621 -2.806301e-04 0.0011881523
## iceland  0.10093202 -0.0039303487 0.0003330384 -2.163154e-04 0.0018302047
## argentin 0.12691382 -0.0226665411 0.0003765071 -1.909892e-04 0.0017249274
## finland  0.22153359  0.0911586195 0.0011790783 -1.065998e-03 0.0032807259
## skorea   0.26277918  0.1425267795 0.0004037433 -2.710684e-04 0.0011330304
## poland   0.11064977 -0.0597440130 0.0005223964 -4.938223e-04 0.0015451010
## switz    0.22223141  0.0093824791 0.0004566021 -3.589338e-04 0.0013705545
## brazil   0.13247467  0.1370104166 0.0002408544 -1.562138e-04 0.0010596832
## nigeria  0.11195614  0.1413833557 0.0002809950 -1.339513e-04 0.0013614816
## chile    0.11531262  0.0395808318 0.0002566581 -1.571078e-04 0.0012826150
## byelorus 0.11559917  0.1265705841 0.0002316383 -1.473846e-04 0.0011406506
## india    0.15666943  0.0614854223 0.0002108708 -9.139084e-05 0.0007803188
## czech    0.09205198  0.0373221695 0.0001602023 -1.109053e-04 0.0009425369
## egermany 0.12373178  0.2026104203 0.0003151898 -2.791858e-04 0.0011981493
## slovenia 0.11637801  0.0818733474 0.0003582661 -1.800883e-04 0.0017417019
## bulgaria 0.18526507  0.1151116394 0.0005613585 -4.807255e-04 0.0017873385
## romania  0.16122303  0.0311905394 0.0005047971 -3.320222e-04 0.0019349102
## china    0.16410105  0.1558057384 0.0002975283 -1.823033e-04 0.0011034250
## portugal 0.10308383  0.0423772969 0.0003103805 -1.961384e-04 0.0016971350
## austria  0.08144816  0.0812577522 0.0002187153 -7.889336e-05 0.0013385159
## turkey   0.09635417  0.1202190099 0.0004596091 -2.861366e-04 0.0026423311
## lithuan  0.04971958  0.0906342006 0.0001545078 -1.259299e-04 0.0013986628
## latvia   0.13141244  0.1555683548 0.0003631150 -2.769157e-04 0.0015455950
## estonia  0.13213639  0.0910846395 0.0003962934 -3.409209e-04 0.0015314712
## russia   0.11256496  0.0784023947 0.0001841914 -1.407663e-04 0.0008853345

## Multivariate meta-analysis on the indirect and direct effects 
indirect2 <- meta(indirect1[, c("ind_eff", "dir_eff")], 
                  indirect1[, c("ind_var", "ind_dir_cov", "dir_var")]) 

summary(indirect2)   

## 
## Call:
## meta(y = indirect1[, c("ind_eff", "dir_eff")], v = indirect1[, 
##     c("ind_var", "ind_dir_cov", "dir_var")])
## 
## 95% confidence intervals: z statistic approximation
## Coefficients:
##               Estimate   Std.Error      lbound      ubound z value
## Intercept1  1.4809e-01  8.0686e-03  1.3228e-01  1.6390e-01 18.3538
## Intercept2  6.8795e-02  1.0638e-02  4.7944e-02  8.9646e-02  6.4667
## Tau2_1_1    2.3571e-03  6.0243e-04  1.1764e-03  3.5379e-03  3.9127
## Tau2_2_1   -8.0175e-05  5.7034e-04 -1.1980e-03  1.0377e-03 -0.1406
## Tau2_2_2    3.3174e-03  1.0386e-03  1.2818e-03  5.3529e-03  3.1942
##             Pr(>|z|)    
## Intercept1 < 2.2e-16 ***
## Intercept2 1.002e-10 ***
## Tau2_1_1   9.129e-05 ***
## Tau2_2_1    0.888208    
## Tau2_2_2    0.001402 ** 
## ---
## Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
## 
## Q statistic on the homogeneity of effect sizes: 561.9791
## Degrees of freedom of the Q statistic: 82
## P value of the Q statistic: 0
## 
## Heterogeneity indices (based on the estimated Tau2):
##                              Estimate
## Intercept1: I2 (Q statistic)   0.8838
## Intercept2: I2 (Q statistic)   0.7198
## 
## Number of studies (or clusters): 42
## Number of observed statistics: 84
## Number of estimated parameters: 5
## Degrees of freedom: 79
## -2 log likelihood: -232.1644 
## OpenMx status1: 0 ("0" or "1": The optimization is considered fine.
## Other values may indicate problems.)

## } 

[Package metaSEM version 0.9.8 Index]