Becker94 {metaSEM}	R Documentation

Five Studies of Ten Correlation Matrices reported by Becker and Schram (1994)

Description

This data set includes five studies of ten correlation matrices reported by Becker and Schram (1994).

Usage

data(Becker94)

Details

A list of data with the following structure:

data

A list of 10 correlation matrices. The variables are SAT (Math), Spatial and SAT (Verbal)

n

A vector of sample sizes

gender

Females or Males samples

Source

Becker, B. J., & Schram, C. M. (1994). Examining explanatory models through research synthesis. In H. Cooper & L. V. Hedges (Eds.), The handbook of research synthesis (pp. 357-381). New York: Russell Sage Foundation.

Examples

## Not run: 
##D data(Becker94)
##D 
##D #### Fixed-effects model
##D ## First stage analysis
##D fixed1 <- tssem1(Becker94$data, Becker94$n, method="FEM")
##D summary(fixed1)
##D 
##D ## Prepare a regression model using create.mxMatrix()
##D A1 <- create.mxMatrix(c(0,0,0,"0.2*Spatial2Aptitude",
##D                         0,0,"0.2*Verbal2Aptitude",0,0), type="Full",
##D                       ncol=3, nrow=3, name="A1")
##D S1 <-
##D create.mxMatrix(c("0.2*ErrorVarAptitude",0,0,1,"0.2*CorBetweenSpatialVerbal",1),
##D                 type="Symm", name="S1")
##D 
##D ## An alternative method to create a regression model using mxMatrix()
##D # A1 <- mxMatrix("Full", ncol=3, nrow=3, value=0, free=c(FALSE,FALSE,FALSE,TRUE,FALSE,
##D #                                                        FALSE,TRUE,FALSE,FALSE),
##D #                label=c(NA,NA,NA,"Spatial2Aptitude",NA,NA,"Verbal2Aptitude",NA,NA),
##D #                name="A1")
##D # S1 <- mxMatrix("Symm", ncol=3, nrow=3, value=c(0.5,0,0,1,0.2,1),
##D #                free=c(TRUE,FALSE,FALSE,FALSE,TRUE,FALSE),
##D #                label=c("ErrorVarAptitude",NA,NA,NA,"CorBetweenSpatialVerbal",NA),
##D #                        name="S1")
##D 
##D ## Second stage analysis
##D fixed2 <- tssem2(fixed1, Amatrix=A1, Smatrix=S1, intervals.type="LB")
##D summary(fixed2)
##D 
##D 
##D #### Fixed-effects model: with gender as cluster
##D ## First stage analysis
##D cluster1 <- tssem1(Becker94$data, Becker94$n, method="FEM", cluster=Becker94$gender)
##D summary(cluster1)
##D 
##D ## Second stage analysis  
##D cluster2 <- tssem2(cluster1, Amatrix=A1, Smatrix=S1, intervals.type="LB")
##D summary(cluster2)
##D 
##D 
##D #### Conventional fixed-effects GLS approach
##D ## First stage analysis
##D ## No random effects
##D ## Replicate Becker's (1992) analysis using 4 studies only
##D gls1 <- tssem1(Becker92$data[1:4], Becker92$n[1:4], method="REM", RE.type="Zero",
##D                model.name="Fixed effects GLS Stage 1")
##D summary(gls1)
##D 
##D ## Fixed-effects GLS model: Second stage analysis
##D gls2 <- tssem2(gls1, Amatrix=A1, Smatrix=S1, intervals.type="LB",
##D                model.name="Fixed effects GLS Stage 2")
##D summary(gls2)
## End(Not run)

[Package metaSEM version 0.9.4 Index]