Berkey98 {metaSEM}	R Documentation

Five Published Trails from Berkey et al. (1998)

Description

The data set includes five published trials, reported by Berkey et al. (1998), comparing surgical and non-surgical treatments for medium-severity periodontal disease, one year after treatment.

Usage

data(Berkey98)

Details

The variables are:

trial

Trial number

pub_year

Publication year

no_of_patients

Number of patients

PD

Patient improvements (mm) in probing depth

AL

Patient improvements (mm) in attachment level

var_PD

Sampling variance of PD

cov_PD_AL

Sampling covariance between PD and AD

var_AL

Sampling variance of AL

Source

Berkey, C. S., Hoaglin, D. C., Antczak-Bouckoms, A., Mosteller, F, & Colditz, G. A. (1998). Meta-analysis of multiple outcomes by regression with random effects. Statistics in Medicine, 17, 2537-2550.

Examples

data(Berkey98) 

#### ML estimation method 
## Multivariate meta-analysis 
summary( meta(y=cbind(PD, AL), v=cbind(var_PD, cov_PD_AL, var_AL), 
              data=Berkey98, I2=c("I2q", "I2hm", "I2am")) ) 

## 
## Call:
## meta(y = cbind(PD, AL), v = cbind(var_PD, cov_PD_AL, var_AL), 
##     data = Berkey98, I2 = c("I2q", "I2hm", "I2am"))
## 
## 95% confidence intervals: z statistic approximation
## Coefficients:
##              Estimate  Std.Error     lbound     ubound z value  Pr(>|z|)
## Intercept1  0.3448392  0.0536312  0.2397239  0.4499544  6.4298 1.278e-10
## Intercept2 -0.3379381  0.0812480 -0.4971812 -0.1786951 -4.1593 3.192e-05
## Tau2_1_1    0.0070020  0.0090497 -0.0107351  0.0247391  0.7737    0.4391
## Tau2_2_1    0.0094607  0.0099698 -0.0100797  0.0290010  0.9489    0.3427
## Tau2_2_2    0.0261445  0.0177409 -0.0086270  0.0609161  1.4737    0.1406
##               
## Intercept1 ***
## Intercept2 ***
## Tau2_1_1      
## Tau2_2_1      
## Tau2_2_2      
## ---
## Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
## 
## Q statistic on the homogeneity of effect sizes: 128.2267
## Degrees of freedom of the Q statistic: 8
## P value of the Q statistic: 0
## 
## Heterogeneity indices (based on the estimated Tau2):
##                                  Estimate
## Intercept1: I2 (Q statistic)       0.6021
## Intercept2: I2 (Q statistic)       0.9250
## Intercept1: I2 (harmonic mean)     0.6264
## Intercept2: I2 (harmonic mean)     0.9359
## Intercept1: I2 (arithmetic mean)   0.5148
## Intercept2: I2 (arithmetic mean)   0.7577
## 
## Number of studies (or clusters): 5
## Number of observed statistics: 10
## Number of estimated parameters: 5
## Degrees of freedom: 5
## -2 log likelihood: -11.68131 
## OpenMx status1: 0 ("0" or "1": The optimization is considered fine.
## Other values may indicate problems.)

## Multivariate meta-analysis with "publication year-1979" as the predictor 
summary( meta(y=cbind(PD, AL), v=cbind(var_PD, cov_PD_AL, var_AL), 
              x=scale(pub_year, center=1979), data=Berkey98, 
              RE.lbound=NA) ) 

## 
## Call:
## meta(y = cbind(PD, AL), v = cbind(var_PD, cov_PD_AL, var_AL), 
##     x = scale(pub_year, center = 1979), data = Berkey98, RE.lbound = NA)
## 
## 95% confidence intervals: z statistic approximation
## Coefficients:
##              Estimate  Std.Error     lbound     ubound z value  Pr(>|z|)
## Intercept1  0.3440001  0.0857659  0.1759021  0.5120981  4.0109 6.048e-05
## Intercept2 -0.2918175  0.1312795 -0.5491206 -0.0345143 -2.2229   0.02622
## Slope1_1    0.0063540  0.1078234 -0.2049761  0.2176841  0.0589   0.95301
## Slope2_1   -0.0705888  0.1620964 -0.3882918  0.2471143 -0.4355   0.66322
## Tau2_1_1    0.0080405  0.0101206 -0.0117955  0.0278766  0.7945   0.42692
## Tau2_2_1    0.0093413  0.0105515 -0.0113392  0.0300218  0.8853   0.37599
## Tau2_2_2    0.0250135  0.0170788 -0.0084603  0.0584873  1.4646   0.14303
##               
## Intercept1 ***
## Intercept2 *  
## Slope1_1      
## Slope2_1      
## Tau2_1_1      
## Tau2_2_1      
## Tau2_2_2      
## ---
## Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
## 
## Q statistic on the homogeneity of effect sizes: 128.2267
## Degrees of freedom of the Q statistic: 8
## P value of the Q statistic: 0
## 
## Explained variances (R2):
##                               y1     y2
## Tau2 (no predictor)    0.0070020 0.0261
## Tau2 (with predictors) 0.0080405 0.0250
## R2                     0.0000000 0.0433
## 
## Number of studies (or clusters): 5
## Number of observed statistics: 10
## Number of estimated parameters: 7
## Degrees of freedom: 3
## -2 log likelihood: -12.00859 
## OpenMx status1: 0 ("0" or "1": The optimization is considered fine.
## Other values may indicate problems.)

## Multivariate meta-analysis with equality constraint on the regression coefficients 
summary( meta(y=cbind(PD, AL), v=cbind(var_PD, cov_PD_AL, var_AL), 
              x=scale(pub_year, center=1979), data=Berkey98, 
              coef.constraints=matrix(c("0.3*Eq_slope", "0.3*Eq_slope"), 
              nrow=2)) ) 

## 
## Call:
## meta(y = cbind(PD, AL), v = cbind(var_PD, cov_PD_AL, var_AL), 
##     x = scale(pub_year, center = 1979), data = Berkey98, coef.constraints = matrix(c("0.3*Eq_slope", 
##         "0.3*Eq_slope"), nrow = 2))
## 
## 95% confidence intervals: z statistic approximation
## Coefficients:
##              Estimate  Std.Error     lbound     ubound z value  Pr(>|z|)
## Intercept1  0.3437612  0.0849829  0.1771979  0.5103245  4.0451 5.231e-05
## Intercept2 -0.3390010  0.1041005 -0.5430344 -0.1349677 -3.2565  0.001128
## Eq_slope    0.0016748  0.1024443 -0.1991123  0.2024619  0.0163  0.986956
## Tau2_1_1    0.0070474  0.0094638 -0.0115013  0.0255962  0.7447  0.456471
## Tau2_2_1    0.0095165  0.0105668 -0.0111940  0.0302269  0.9006  0.367800
## Tau2_2_2    0.0261979  0.0180773 -0.0092330  0.0616288  1.4492  0.147278
##               
## Intercept1 ***
## Intercept2 ** 
## Eq_slope      
## Tau2_1_1      
## Tau2_2_1      
## Tau2_2_2      
## ---
## Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
## 
## Q statistic on the homogeneity of effect sizes: 128.2267
## Degrees of freedom of the Q statistic: 8
## P value of the Q statistic: 0
## 
## Explained variances (R2):
##                               y1     y2
## Tau2 (no predictor)    0.0070020 0.0261
## Tau2 (with predictors) 0.0070474 0.0262
## R2                     0.0000000 0.0000
## 
## Number of studies (or clusters): 5
## Number of observed statistics: 10
## Number of estimated parameters: 6
## Degrees of freedom: 4
## -2 log likelihood: -11.68158 
## OpenMx status1: 0 ("0" or "1": The optimization is considered fine.
## Other values may indicate problems.)

#### REML estimation method 
## Multivariate meta-analysis 
summary( reml(y=cbind(PD, AL), v=cbind(var_PD, cov_PD_AL, var_AL), 
              data=Berkey98, 
              model.name="Multivariate meta analysis with REML") ) 

## 
## Call:
## reml(y = cbind(PD, AL), v = cbind(var_PD, cov_PD_AL, var_AL), 
##     data = Berkey98, model.name = "Multivariate meta analysis with REML")
## 
## 95% confidence intervals: z statistic approximation
## Coefficients:
##           Estimate Std.Error    lbound    ubound z value Pr(>|z|)
## Tau2_1_1  0.011733  0.013645 -0.015011  0.038477  0.8599   0.3899
## Tau2_2_1  0.011916  0.014416 -0.016340  0.040172  0.8266   0.4085
## Tau2_2_2  0.032651  0.024402 -0.015176  0.080479  1.3380   0.1809
## 
## Number of studies (or clusters): 5
## Number of observed statistics: 8
## Number of estimated parameters: 3
## Degrees of freedom: 5
## -2 log likelihood: -18.86768 
## OpenMx status1: 0 ("0" or "1": The optimization is considered fine.
## Other values may indicate problems.)

## Multivariate meta-analysis with "publication year-1979" as the predictor 
summary( reml(y=cbind(PD, AL), v=cbind(var_PD, cov_PD_AL, var_AL), 
              x=scale(pub_year, center=1979), data=Berkey98) ) 

## 
## Call:
## reml(y = cbind(PD, AL), v = cbind(var_PD, cov_PD_AL, var_AL), 
##     x = scale(pub_year, center = 1979), data = Berkey98)
## 
## 95% confidence intervals: z statistic approximation
## Coefficients:
##             Estimate Std.Error lbound ubound z value Pr(>|z|)
## Tau2_1_1  0.00023168        NA     NA     NA      NA       NA
## Tau2_2_1 -0.01160166        NA     NA     NA      NA       NA
## Tau2_2_2  0.00185663        NA     NA     NA      NA       NA
## 
## Number of studies (or clusters): 5
## Number of observed statistics: 6
## Number of estimated parameters: 3
## Degrees of freedom: 3
## -2 log likelihood: -48.55547 
## OpenMx status1: 6 ("0" or "1": The optimization is considered fine.
## Other values may indicate problems.)

[Package metaSEM version 0.9.8 Index]