Bornmann07 {metaSEM}	R Documentation

A Dataset from Bornmann et al. (2007)

Description

A dataset from Bornmann et al. (2007) for three-level meta-analysis.

Usage

data(Bornmann07)

Details

The variables are:

ID

ID of the study

Study

Study name

Cluster

Cluster for effect sizes

logOR

Effect size: log odds ratio

v

Sampling variance of logOR

Year

Year of publication

Type

Type of proposal: either Grant or Fellowship

Discipline

Discipline of proposal: either Physical sciences, Life sciences/biology, Social sciences/humanities or Multidisciplinary)

Country

Country of proposal: either United States, Canada, Australia, United Kingdom or Europe

Source

Bornmann, L., Mutz, R., & Daniel, H.-D. (2007). Gender differences in grant peer review: A meta-analysis. Journal of Informetrics, 1(3), 226-238. doi:10.1016/j.joi.2007.03.001

References

Cheung, M. W.-L. (2014). Modeling dependent effect sizes with three-level meta-analyses: A structural equation modeling approach. Psychological Methods, 19, 211-229.

Marsh, H. W., Bornmann, L., Mutz, R., Daniel, H.-D., & O'Mara, A. (2009). Gender Effects in the Peer Reviews of Grant Proposals: A Comprehensive Meta-Analysis Comparing Traditional and Multilevel Approaches. Review of Educational Research, 79(3), 1290-1326. doi:10.3102/0034654309334143

Examples

data(Bornmann07)

#### ML estimation method
## No predictor
summary( meta3(y=logOR, v=v, cluster=Cluster, data=Bornmann07) )

## 
## Call:
## meta3(y = logOR, v = v, cluster = Cluster, data = Bornmann07)
## 
## 95% confidence intervals: z statistic approximation
## Coefficients:
##             Estimate  Std.Error     lbound     ubound z value Pr(>|z|)  
## Intercept -0.1007784  0.0401327 -0.1794371 -0.0221198 -2.5111  0.01203 *
## Tau2_2     0.0037965  0.0027210 -0.0015367  0.0091297  1.3952  0.16295  
## Tau2_3     0.0141352  0.0091445 -0.0037877  0.0320580  1.5458  0.12216  
## ---
## Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
## 
## Q statistic on the homogeneity of effect sizes: 221.2809
## Degrees of freedom of the Q statistic: 65
## P value of the Q statistic: 0
## 
## Heterogeneity indices (based on the estimated Tau2):
##                               Estimate
## I2_2 (Typical v: Q statistic)   0.1568
## I2_3 (Typical v: Q statistic)   0.5839
## 
## Number of studies (or clusters): 21
## Number of observed statistics: 66
## Number of estimated parameters: 3
## Degrees of freedom: 63
## -2 log likelihood: 25.80256 
## OpenMx status1: 0 ("0" or "1": The optimization is considered fine.
## Other values may indicate problems.)

## Type as a predictor
## Grant: 0
## Fellowship: 1
summary( meta3(y=logOR, v=v, x=(as.numeric(Type)-1),
               cluster=Cluster, data=Bornmann07) )

## 
## Call:
## meta3(y = logOR, v = v, cluster = Cluster, x = (as.numeric(Type) - 
##     1), data = Bornmann07)
## 
## 95% confidence intervals: z statistic approximation
## Coefficients:
##             Estimate  Std.Error     lbound     ubound z value  Pr(>|z|)
## Intercept -0.0066071  0.0371125 -0.0793462  0.0661320 -0.1780 0.8587001
## Slope_1   -0.1955869  0.0541649 -0.3017483 -0.0894256 -3.6110 0.0003051
## Tau2_2     0.0035335  0.0024306 -0.0012303  0.0082974  1.4538 0.1460058
## Tau2_3     0.0029079  0.0031183 -0.0032039  0.0090197  0.9325 0.3510704
##              
## Intercept    
## Slope_1   ***
## Tau2_2       
## Tau2_3       
## ---
## Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
## 
## Q statistic on the homogeneity of effect sizes: 221.2809
## Degrees of freedom of the Q statistic: 65
## P value of the Q statistic: 0
## 
## Explained variances (R2):
##                          Level 2 Level 3
## Tau2 (no predictor)    0.0037965  0.0141
## Tau2 (with predictors) 0.0035335  0.0029
## R2                     0.0692595  0.7943
## 
## Number of studies (or clusters): 21
## Number of observed statistics: 66
## Number of estimated parameters: 4
## Degrees of freedom: 62
## -2 log likelihood: 17.62569 
## OpenMx status1: 0 ("0" or "1": The optimization is considered fine.
## Other values may indicate problems.)

## Centered Year as a predictor
summary( meta3(y=logOR, v=v, x=scale(Year, scale=FALSE),
               cluster=Cluster, data=Bornmann07) )

## 
## Call:
## meta3(y = logOR, v = v, cluster = Cluster, x = scale(Year, scale = FALSE), 
##     data = Bornmann07)
## 
## 95% confidence intervals: z statistic approximation
## Coefficients:
##             Estimate  Std.Error     lbound     ubound z value Pr(>|z|)   
## Intercept -0.1007172  0.0388512 -0.1768642 -0.0245702 -2.5924 0.009531 **
## Slope_1    0.0045812  0.0064555 -0.0080714  0.0172338  0.7097 0.477919   
## Tau2_2     0.0041930  0.0029068 -0.0015043  0.0098904  1.4425 0.149171   
## Tau2_3     0.0125756  0.0088638 -0.0047972  0.0299483  1.4188 0.155970   
## ---
## Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
## 
## Q statistic on the homogeneity of effect sizes: 221.2809
## Degrees of freedom of the Q statistic: 65
## P value of the Q statistic: 0
## 
## Explained variances (R2):
##                          Level 2 Level 3
## Tau2 (no predictor)    0.0037965  0.0141
## Tau2 (with predictors) 0.0041930  0.0126
## R2                     0.0000000  0.1103
## 
## Number of studies (or clusters): 21
## Number of observed statistics: 66
## Number of estimated parameters: 4
## Degrees of freedom: 62
## -2 log likelihood: 25.28968 
## OpenMx status1: 0 ("0" or "1": The optimization is considered fine.
## Other values may indicate problems.)

#### REML estimation method
## No predictor
summary( reml3(y=logOR, v=v, cluster=Cluster, data=Bornmann07) )

## 
## Call:
## reml3(y = logOR, v = v, cluster = Cluster, data = Bornmann07)
## 
## 95% confidence intervals: z statistic approximation
## Coefficients:
##          Estimate  Std.Error     lbound     ubound z value Pr(>|z|)
## Tau2_2  0.0037528  0.0026831 -0.0015060  0.0090116  1.3987   0.1619
## Tau2_3  0.0160846  0.0104008 -0.0043005  0.0364698  1.5465   0.1220
## 
## Number of studies (or clusters): 66
## Number of observed statistics: 65
## Number of estimated parameters: 2
## Degrees of freedom: 63
## -2 log likelihood: -89.10382 
## OpenMx status1: 0 ("0" or "1": The optimization is considered fine.
## Other values may indicate problems.)

## Type as a predictor
## Grants: 0
## Fellowship: 1
summary( reml3(y=logOR, v=v, x=(as.numeric(Type)-1),
               cluster=Cluster, data=Bornmann07) )

## 
## Call:
## reml3(y = logOR, v = v, cluster = Cluster, x = (as.numeric(Type) - 
##     1), data = Bornmann07)
## 
## 95% confidence intervals: z statistic approximation
## Coefficients:
##          Estimate  Std.Error     lbound     ubound z value Pr(>|z|)
## Tau2_2  0.0035470  0.0024639 -0.0012821  0.0083760  1.4396   0.1500
## Tau2_3  0.0044844  0.0045692 -0.0044711  0.0134400  0.9814   0.3264
## 
## Number of studies (or clusters): 66
## Number of observed statistics: 64
## Number of estimated parameters: 2
## Degrees of freedom: 62
## -2 log likelihood: -90.66179 
## OpenMx status1: 0 ("0" or "1": The optimization is considered fine.
## Other values may indicate problems.)

## Centered Year as a predictor
summary( reml3(y=logOR, v=v, x=scale(Year, scale=FALSE),
               cluster=Cluster, data=Bornmann07) )

## 
## Call:
## reml3(y = logOR, v = v, cluster = Cluster, x = scale(Year, scale = FALSE), 
##     data = Bornmann07)
## 
## 95% confidence intervals: z statistic approximation
## Coefficients:
##          Estimate  Std.Error     lbound     ubound z value Pr(>|z|)
## Tau2_2  0.0044586  0.0030516 -0.0015225  0.0104397  1.4610   0.1440
## Tau2_3  0.0150012  0.0103867 -0.0053563  0.0353587  1.4443   0.1487
## 
## Number of studies (or clusters): 66
## Number of observed statistics: 64
## Number of estimated parameters: 2
## Degrees of freedom: 62
## -2 log likelihood: -79.4153 
## OpenMx status1: 0 ("0" or "1": The optimization is considered fine.
## Other values may indicate problems.)

## Handling missing covariates with FIML
## MCAR
## Set seed for replication
set.seed(1000000)

## Copy Bornmann07 to my.df
my.df <- Bornmann07
## "Fellowship": 1; "Grant": 0
my.df$Type_MCAR <- ifelse(Bornmann07$Type=="Fellowship", yes=1, no=0)

## Create 17 out of 66 missingness with MCAR
my.df$Type_MCAR[sample(1:66, 17)] <- NA
summary(meta3X(y=logOR, v=v, cluster=Cluster, x2=Type_MCAR, data=my.df))

## 
## Call:
## meta3X(y = logOR, v = v, cluster = Cluster, x2 = Type_MCAR, data = my.df)
## 
## 95% confidence intervals: z statistic approximation
## Coefficients:
##             Estimate  Std.Error     lbound     ubound z value Pr(>|z|)   
## Intercept -0.0106318  0.0397685 -0.0885766  0.0673131 -0.2673 0.789206   
## SlopeX2_1 -0.1753249  0.0582619 -0.2895161 -0.0611336 -3.0093 0.002619 **
## Tau2_2     0.0030338  0.0026839 -0.0022266  0.0082941  1.1304 0.258324   
## Tau2_3     0.0036839  0.0042817 -0.0047082  0.0120759  0.8604 0.389586   
## ---
## Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
## 
## Explained variances (R2):
##                          Level 2 Level 3
## Tau2 (no predictor)    0.0037965  0.0141
## Tau2 (with predictors) 0.0030338  0.0037
## R2                     0.2009069  0.7394
## 
## Number of studies (or clusters): 21
## Number of observed statistics: 115
## Number of estimated parameters: 7
## Degrees of freedom: 108
## -2 log likelihood: 49.76372 
## OpenMx status1: 0 ("0" or "1": The optimization is considered fine.
## Other values may indicate problems.)

## MAR
Type_MAR <- ifelse(Bornmann07$Type=="Fellowship", yes=1, no=0)

## Create 27 out of 66 missingness with MAR for cases Year<1996
index_MAR <- ifelse(Bornmann07$Year<1996, yes=TRUE, no=FALSE)
Type_MAR[index_MAR] <- NA

## Include auxiliary variable
summary(meta3X(y=logOR, v=v, cluster=Cluster, x2=Type_MAR, av2=Year, data=my.df))

## 
## Call:
## meta3X(y = logOR, v = v, cluster = Cluster, x2 = Type_MAR, av2 = Year, 
##     data = my.df)
## 
## 95% confidence intervals: z statistic approximation
## Coefficients:
##             Estimate  Std.Error     lbound     ubound z value Pr(>|z|)   
## Intercept -0.0264058  0.0571965 -0.1385089  0.0856974 -0.4617 0.644320   
## SlopeX2_1 -0.2003999  0.0691031 -0.3358395 -0.0649603 -2.9000 0.003731 **
## Tau2_2     0.0029970  0.0022371 -0.0013877  0.0073816  1.3397 0.180358   
## Tau2_3     0.0030212  0.0032462 -0.0033412  0.0093836  0.9307 0.352010   
## ---
## Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
## 
## Explained variances (R2):
##                          Level 2 Level 3
## Tau2 (no predictor)    0.0049237  0.0088
## Tau2 (with predictors) 0.0029970  0.0030
## R2                     0.3913243  0.6571
## 
## Number of studies (or clusters): 21
## Number of observed statistics: 171
## Number of estimated parameters: 14
## Degrees of freedom: 157
## -2 log likelihood: 377.3479 
## OpenMx status1: 0 ("0" or "1": The optimization is considered fine.
## Other values may indicate problems.)

[Package metaSEM version 0.9.5-2 Index]