Cheung09 {metaSEM}	R Documentation

A Data Set from TSSEM User's Guide Version 1.11 by Cheung (2009)

Description

Four studies were selected from the data set used by Cheung and Chan (2005; 2009). Some variables were randomly deleted to illustrate the analysis with missing data.

Usage

data(Cheung09)

Details

A list of data with the following structure:

data

A list of 4 studies of correlation matrices

n

A vector of sample sizes

Source

Cheung, M. W.-L. (2009). TSSEM: A LISREL syntax generator for two-stage structural equation modeling (Version 1.11) [Computer software]. Retrieved from http://courses.nus.edu.sg/course/psycwlm/internet/tssem.zip.

References

Cheung, M. W.-L., & Chan, W. (2005). Meta-analytic structural equation modeling: A two-stage approach. Psychological Methods, 10, 40-64.

Cheung, M. W.-L., & Chan, W. (2009). A two-stage approach to synthesizing covariance matrices in meta-analytic structural equation modeling. Structural Equation Modeling, 16, 28-53.

Examples

## \dontrun{ 
data(Cheung09) 

#### Fixed-effects model: Stage 1 analysis 
fixed1 <- tssem1(Cheung09$data, Cheung09$n, method="FEM") 
summary(fixed1) 

## 
## Call:
## tssem1FEM(my.df = my.df, n = n, cor.analysis = cor.analysis, 
##     model.name = model.name, cluster = cluster, suppressWarnings = suppressWarnings, 
##     silent = silent, run = run)
## 
## Coefficients:
##        Estimate Std.Error z value  Pr(>|z|)    
## S[1,2] 0.295208  0.019248 15.3371 < 2.2e-16 ***
## S[1,3] 0.195328  0.024735  7.8967 2.887e-15 ***
## S[1,4] 0.218779  0.024106  9.0758 < 2.2e-16 ***
## S[1,5] 0.219895  0.024393  9.0147 < 2.2e-16 ***
## S[1,6] 0.188089  0.020230  9.2975 < 2.2e-16 ***
## S[1,7] 0.192204  0.020232  9.5001 < 2.2e-16 ***
## S[1,8] 0.098339  0.025639  3.8355 0.0001253 ***
## S[1,9] 0.098793  0.025889  3.8161 0.0001356 ***
## S[2,3] 0.369555  0.022685 16.2910 < 2.2e-16 ***
## S[2,4] 0.254053  0.023512 10.8053 < 2.2e-16 ***
## S[2,5] 0.228530  0.024084  9.4888 < 2.2e-16 ***
## S[2,6] 0.114155  0.020509  5.5660 2.606e-08 ***
## S[2,7] 0.045237  0.020782  2.1768 0.0294980 *  
## S[2,8] 0.148069  0.025231  5.8684 4.400e-09 ***
## S[2,9] 0.101846  0.025699  3.9631 7.399e-05 ***
## S[3,4] 0.307684  0.019814 15.5286 < 2.2e-16 ***
## S[3,5] 0.190550  0.021184  8.9949 < 2.2e-16 ***
## S[3,6] 0.168648  0.020879  8.0772 6.661e-16 ***
## S[3,7] 0.131200  0.021062  6.2293 4.684e-10 ***
## S[3,8] 0.165582  0.026993  6.1343 8.553e-10 ***
## S[3,9] 0.125631  0.027561  4.5584 5.155e-06 ***
## S[4,5] 0.451826  0.017497 25.8237 < 2.2e-16 ***
## S[4,6] 0.385270  0.018388 20.9520 < 2.2e-16 ***
## S[4,7] 0.356282  0.018816 18.9349 < 2.2e-16 ***
## S[4,8] 0.118673  0.026662  4.4511 8.545e-06 ***
## S[4,9] 0.056846  0.027172  2.0920 0.0364361 *  
## S[5,6] 0.312554  0.019523 16.0092 < 2.2e-16 ***
## S[5,7] 0.231892  0.020356 11.3918 < 2.2e-16 ***
## S[5,8] 0.238516  0.026139  9.1249 < 2.2e-16 ***
## S[5,9] 0.064726  0.027395  2.3627 0.0181416 *  
## S[6,7] 0.553733  0.012903 42.9149 < 2.2e-16 ***
## S[6,8] 0.162001  0.021020  7.7069 1.288e-14 ***
## S[6,9] 0.053055  0.021643  2.4513 0.0142330 *  
## S[7,8] 0.135421  0.021197  6.3887 1.673e-10 ***
## S[7,9] 0.066484  0.021635  3.0730 0.0021192 ** 
## S[8,9] 0.198039  0.021239  9.3245 < 2.2e-16 ***
## ---
## Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
## 
## Goodness-of-fit indices:
##                                      Value
## Sample size                      2902.0000
## Chi-square of target model        172.9686
## DF of target model                 57.0000
## p value of target model             0.0000
## Chi-square of independence model 3251.3935
## DF of independence model           93.0000
## RMSEA                               0.0530
## RMSEA lower 95% CI                  0.0441
## RMSEA upper 95% CI                  0.0621
## SRMR                                0.0492
## TLI                                 0.9401
## CFI                                 0.9633
## AIC                                58.9686
## BIC                              -281.5013
## OpenMx status1: 0 ("0" or "1": The optimization is considered fine.
## Other values may indicate problems.)

## Prepare a model implied matrix 
## Factor correlation matrix 
Phi <- create.mxMatrix( c("0.3*corf2f1","0.3*corf3f1","0.3*corf3f2"), 
                        type="Stand", as.mxMatrix=FALSE ) 
## Error variances 
Psi <- create.mxMatrix( paste("0.2*e", 1:9, sep=""), type="Diag", 
                        as.mxMatrix=FALSE ) 

## Create Smatrix 
S1 <- bdiagMat(list(Psi, Phi)) 
## dimnames(S1)[[1]] <- dimnames(S1)[[2]] <- c(paste("x",1:9,sep=""), 
##                                             paste("f",1:3,sep="")) 
## S1 
S1 <- as.mxMatrix(S1) 

## Factor loadings 
Lambda <- create.mxMatrix( c(".3*f1x1",".3*f1x2",".3*f1x3",rep(0,9), 
                             ".3*f2x4",".3*f2x5",".3*f2x6",".3*f2x7", 
                             rep(0,9),".3*f3x8",".3*f3x9"), type="Full", 
                             ncol=3, nrow=9, as.mxMatrix=FALSE ) 
Zero1 <- matrix(0, nrow=9, ncol=9) 
Zero2 <- matrix(0, nrow=3, ncol=12) 

## Create Amatrix 
A1 <- rbind( cbind(Zero1, Lambda), 
             Zero2 ) 
## dimnames(A1)[[1]] <- dimnames(A1)[[2]] <- c(paste("x",1:9,sep=""), 
##                                             paste("f",1:3,sep="")) 
## A1 
A1 <- as.mxMatrix(A1) 

## Create Fmatrix 
F1 <- create.Fmatrix(c(rep(1,9), rep(0,3))) 

#### Fixed-effects model: Stage 2 analysis 
fixed2 <- tssem2(fixed1, Amatrix=A1, Smatrix=S1, Fmatrix=F1, 
                 intervals.type="LB") 
summary(fixed2) 

## 
## Call:
## wls(Cov = coef.tssem1FEM(tssem1.obj), asyCov = vcov.tssem1FEM(tssem1.obj), 
##     n = sum(tssem1.obj$n), Amatrix = Amatrix, Smatrix = Smatrix, 
##     Fmatrix = Fmatrix, diag.constraints = diag.constraints, cor.analysis = tssem1.obj$cor.analysis, 
##     intervals.type = intervals.type, mx.algebras = mx.algebras, 
##     model.name = model.name, suppressWarnings = suppressWarnings, 
##     silent = silent, run = run)
## 
## 95% confidence intervals: Likelihood-based statistic
## Coefficients:
##         Estimate Std.Error  lbound  ubound z value Pr(>|z|)
## f1x1     0.51573        NA 0.46775 0.56414      NA       NA
## f1x2     0.56900        NA 0.52180 0.61693      NA       NA
## f1x3     0.58853        NA 0.53621 0.64170      NA       NA
## f2x4     0.70588        NA 0.67411 0.73741      NA       NA
## f2x5     0.57726        NA 0.54110 0.61323      NA       NA
## f2x6     0.74770        NA 0.72062 0.77501      NA       NA
## f2x7     0.69393        NA 0.66542 0.72273      NA       NA
## f3x8     0.62056        NA 0.52535 0.74448      NA       NA
## f3x9     0.33091        NA 0.26545 0.39717      NA       NA
## corf2f1  0.54547        NA 0.49391 0.59759      NA       NA
## corf3f1  0.48561        NA 0.37639 0.60082      NA       NA
## corf3f2  0.38878        NA 0.31165 0.47256      NA       NA
## 
## Goodness-of-fit indices:
##                                                Value
## Sample size                                2902.0000
## Chi-square of target model                  379.8705
## DF of target model                           24.0000
## p value of target model                       0.0000
## Number of constraints imposed on "Smatrix"    0.0000
## DF manually adjusted                          0.0000
## Chi-square of independence model           3309.5988
## DF of independence model                     36.0000
## RMSEA                                         0.0715
## RMSEA lower 95% CI                            0.0653
## RMSEA upper 95% CI                            0.0779
## SRMR                                          0.0727
## TLI                                           0.8369
## CFI                                           0.8913
## AIC                                         331.8705
## BIC                                         188.5148
## OpenMx status1: 0 ("0" or "1": The optimization is considered fine.
## Other values indicate problems.)

## } 

[Package metaSEM version 0.9.8 Index]