Hox02 {metaSEM}	R Documentation

Simulated Effect Sizes Reported by Hox (2002)

Description

Twenty stimulated studies on standardized mean difference and one continuous study characteristic reported by Hox (2002).

Usage

data(Hox02)

Details

The variables are:

study

Study number

yi

Effect size (standardized mean difference)

vi

Sampling variance of the effect size

weeks

Duration of the experimental intervention in terms of weeks

Source

Hox, J. J. (2002). Multilevel analysis: Techniques and applications. Mahwah, N.J.: Lawrence Erlbaum Associates.

References

Cheung, M. W.-L. (2008). A model for integrating fixed-, random-, and mixed-effects meta-analyses into structural equation modeling. Psychological Methods, 13, 182-202.

Examples

data(Hox02)

#### ML estimation method
## Random-effects meta-analysis
summary( meta(y=yi, v=vi, data=Hox02, I2=c("I2q", "I2hm"), intervals.type="LB") ) 

## 
## Call:
## meta(y = yi, v = vi, data = Hox02, intervals.type = "LB", I2 = c("I2q", 
##     "I2hm"))
## 
## 95% confidence intervals: Likelihood-based statistic
## Coefficients:
##            Estimate Std.Error  lbound  ubound z value Pr(>|z|)
## Intercept1  0.57903        NA 0.36541 0.80025      NA       NA
## Tau2_1_1    0.13152        NA 0.03239 0.35883      NA       NA
## 
## Q statistic on the homogeneity of effect sizes: 49.5852
## Degrees of freedom of the Q statistic: 19
## P value of the Q statistic: 0.000150801
## 
## Heterogeneity indices (I2) and their 95% likelihood-based CIs:
##                                 lbound Estimate ubound
## Intercept1: I2 (Q statistic)   0.27625  0.60780 0.6749
## Intercept1: I2 (harmonic mean) 0.27713  0.60885 0.6766
## 
## Number of studies (or clusters): 20
## Number of observed statistics: 20
## Number of estimated parameters: 2
## Degrees of freedom: 18
## -2 log likelihood: 27.79916 
## OpenMx status1: 0 ("0" or "1": The optimization is considered fine.
## Other values may indicate problems.)

## Fixed-effects meta-analysis
summary( meta(y=yi, v=vi, data=Hox02, RE.constraints=0,
              model.name="Fixed effects model") )

## 
## Call:
## meta(y = yi, v = vi, data = Hox02, RE.constraints = 0, model.name = "Fixed effects model")
## 
## 95% confidence intervals: z statistic approximation
## Coefficients:
##            Estimate Std.Error   lbound   ubound z value  Pr(>|z|)    
## Intercept1 0.550206  0.064998 0.422813 0.677599   8.465 < 2.2e-16 ***
## ---
## Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
## 
## Q statistic on the homogeneity of effect sizes: 49.5852
## Degrees of freedom of the Q statistic: 19
## P value of the Q statistic: 0.000150801
## 
## Heterogeneity indices (based on the estimated Tau2):
##                              Estimate
## Intercept1: I2 (Q statistic)        0
## 
## Number of studies (or clusters): 20
## Number of observed statistics: 20
## Number of estimated parameters: 1
## Degrees of freedom: 19
## -2 log likelihood: 37.70073 
## OpenMx status1: 0 ("0" or "1": The optimization is considered fine.
## Other values may indicate problems.)

## Mixed-effects meta-analysis with "weeks" as a predictor
## Request likelihood-based CI
summary( meta(y=yi, v=vi, x=weeks, data=Hox02, intervals.type="LB",
              model.name="Mixed effects meta analysis with LB CI") )

## 
## Call:
## meta(y = yi, v = vi, x = weeks, data = Hox02, intervals.type = "LB", 
##     model.name = "Mixed effects meta analysis with LB CI")
## 
## 95% confidence intervals: Likelihood-based statistic
## Coefficients:
##             Estimate Std.Error    lbound    ubound z value Pr(>|z|)
## Intercept1 -0.213561        NA -0.618064  0.179466      NA       NA
## Slope1_1    0.138658        NA  0.074675  0.206899      NA       NA
## Tau2_1_1    0.023252        NA        NA  0.023252      NA       NA
## 
## Q statistic on the homogeneity of effect sizes: 49.5852
## Degrees of freedom of the Q statistic: 19
## P value of the Q statistic: 0.000150801
## 
## Explained variances (R2):
##                            y1
## Tau2 (no predictor)    0.1315
## Tau2 (with predictors) 0.0233
## R2                     0.8232
## 
## Number of studies (or clusters): 20
## Number of observed statistics: 20
## Number of estimated parameters: 3
## Degrees of freedom: 17
## -2 log likelihood: 13.90435 
## OpenMx status1: 0 ("0" or "1": The optimization is considered fine.
## Other values may indicate problems.)

#### REML estimation method
## Random-effects meta-analysis with REML
summary( VarComp <- reml(y=yi, v=vi, data=Hox02) )

## 
## Call:
## reml(y = yi, v = vi, data = Hox02)
## 
## 95% confidence intervals: z statistic approximation
## Coefficients:
##           Estimate Std.Error    lbound    ubound z value Pr(>|z|)  
## Tau2_1_1  0.144609  0.079766 -0.011729  0.300947  1.8129  0.06984 .
## ---
## Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
## 
## Number of studies (or clusters): 20
## Number of observed statistics: 19
## Number of estimated parameters: 1
## Degrees of freedom: 18
## -2 log likelihood: -4.477744 
## OpenMx status1: 0 ("0" or "1": The optimization is considered fine.
## Other values may indicate problems.)

## Extract the variance component
VarComp_REML <- matrix( coef(VarComp), ncol=1, nrow=1 )

## Meta-analysis by treating the variance component as fixed
summary( meta(y=yi, v=vi, data=Hox02, RE.constraints=VarComp_REML) )

## 
## Call:
## meta(y = yi, v = vi, data = Hox02, RE.constraints = VarComp_REML)
## 
## 95% confidence intervals: z statistic approximation
## Coefficients:
##            Estimate Std.Error  lbound  ubound z value  Pr(>|z|)    
## Intercept1  0.58015   0.10800 0.36847 0.79182  5.3716 7.802e-08 ***
## ---
## Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
## 
## Q statistic on the homogeneity of effect sizes: 49.5852
## Degrees of freedom of the Q statistic: 19
## P value of the Q statistic: 0.000150801
## 
## Heterogeneity indices (based on the estimated Tau2):
##                              Estimate
## Intercept1: I2 (Q statistic)   0.6302
## 
## Number of studies (or clusters): 20
## Number of observed statistics: 20
## Number of estimated parameters: 1
## Degrees of freedom: 19
## -2 log likelihood: 27.82858 
## OpenMx status1: 0 ("0" or "1": The optimization is considered fine.
## Other values may indicate problems.)

## Mixed-effects meta-analysis with "weeks" as a predictor
## Request likelihood-based CI
summary( reml(y=yi, v=vi, x=weeks, intervals.type="LB",
              data=Hox02, model.name="REML with LB CI") )

## 
## Call:
## reml(y = yi, v = vi, x = weeks, data = Hox02, intervals.type = "LB", 
##     model.name = "REML with LB CI")
## 
## 95% confidence intervals: Likelihood-based statistic
## Coefficients:
##          Estimate Std.Error lbound   ubound z value Pr(>|z|)
## Tau2_1_1 0.036582        NA     NA 0.175443      NA       NA
## 
## Number of studies (or clusters): 20
## Number of observed statistics: 18
## Number of estimated parameters: 1
## Degrees of freedom: 17
## -2 log likelihood: -10.86705 
## OpenMx status1: 0 ("0" or "1": The optimization is considered fine.
## Other values may indicate problems.)

[Package metaSEM version 0.9.5-2 Index]