issp05 {metaSEM}	R Documentation

Data Set from ISSP (2005)

Description

Thirty-two covariance matrices on work-related attitudes were extracted from the International Social Survey Programme 2005: Work Orientation III (ISSP, 2005). Seven variables were selected for demonstration purposes. They were grouped into three constructs: Importance of Job Prospects measured by job security (JP1), high income (JP2), and opportunity for advancement (JP3); Importance of Job Autonomy measured by work independently (JA1) and decide time of work (JA2); and Importance of Contributions to Society measured by help other people (CS1) and a job useful to society (CS2).

Usage

data(issp05)

Details

A list of data with the following structure:

data

A list of 32 covariance matrices

n

A vector of sample sizes

Author(s)

Mike W.-L. Cheung <mikewlcheung@nus.edu.sg>

Source

ISSP Research Group (2007): International Social Survey Programme 2005: Work Orientation III (ISSP 2005). GESIS Data Archive, Cologne. ZA4350 Data file Version 1.0.0, doi:10.4232/1.4350

See Also

issp89

Examples

## \dontrun{ 
data(issp05) 

#### Fixed-effects TSSEM 
fixed1 <- tssem1(issp05$data, issp05$n, method="FEM") 
summary(fixed1) 

## 
## Call:
## tssem1FEM(my.df = my.df, n = n, cor.analysis = cor.analysis, 
##     model.name = model.name, cluster = cluster, suppressWarnings = suppressWarnings, 
##     silent = silent, run = run)
## 
## Coefficients:
##         Estimate Std.Error z value  Pr(>|z|)    
## S[1,2] 0.3458536 0.0043955  78.683 < 2.2e-16 ***
## S[1,3] 0.2957012 0.0045543  64.927 < 2.2e-16 ***
## S[1,4] 0.1451564 0.0048873  29.700 < 2.2e-16 ***
## S[1,5] 0.0902511 0.0049407  18.267 < 2.2e-16 ***
## S[1,6] 0.2058902 0.0047674  43.187 < 2.2e-16 ***
## S[1,7] 0.2178484 0.0047418  45.942 < 2.2e-16 ***
## S[2,3] 0.4039957 0.0041659  96.976 < 2.2e-16 ***
## S[2,4] 0.1941360 0.0047938  40.497 < 2.2e-16 ***
## S[2,5] 0.1905006 0.0047923  39.751 < 2.2e-16 ***
## S[2,6] 0.1348540 0.0048943  27.553 < 2.2e-16 ***
## S[2,7] 0.1432304 0.0048833  29.331 < 2.2e-16 ***
## S[3,4] 0.2780719 0.0046158  60.243 < 2.2e-16 ***
## S[3,5] 0.1882161 0.0048136  39.101 < 2.2e-16 ***
## S[3,6] 0.2530847 0.0046802  54.076 < 2.2e-16 ***
## S[3,7] 0.2500475 0.0046855  53.367 < 2.2e-16 ***
## S[4,5] 0.3628171 0.0043392  83.614 < 2.2e-16 ***
## S[4,6] 0.3779783 0.0042854  88.201 < 2.2e-16 ***
## S[4,7] 0.3177996 0.0044889  70.797 < 2.2e-16 ***
## S[5,6] 0.2647916 0.0046363  57.112 < 2.2e-16 ***
## S[5,7] 0.2541295 0.0046669  54.453 < 2.2e-16 ***
## S[6,7] 0.6623172 0.0028137 235.393 < 2.2e-16 ***
## ---
## Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
## 
## Goodness-of-fit indices:
##                                       Value
## Sample size                      40596.0000
## Chi-square of target model        5309.8751
## DF of target model                 651.0000
## p value of target model              0.0000
## Chi-square of independence model 61921.2659
## DF of independence model           672.0000
## RMSEA                                0.0751
## RMSEA lower 95% CI                   0.0733
## RMSEA upper 95% CI                   0.0770
## SRMR                                 0.0902
## TLI                                  0.9215
## CFI                                  0.9239
## AIC                               4007.8751
## BIC                              -1598.1624
## OpenMx status1: 6 ("0" or "1": The optimization is considered fine.
## Other values may indicate problems.)

## Prepare a model implied matrix 
## Factor correlation matrix 
Phi <- create.mxMatrix( c("0.3*corf2f1","0.3*corf3f1","0.3*corf3f2"), 
                        type="Stand", as.mxMatrix=FALSE ) 

## Error variances 
Psi <- create.mxMatrix( paste("0.2*e", 1:7, sep=""), type="Diag", 
                        as.mxMatrix=FALSE ) 

## Create Smatrix 
S1 <- bdiagMat(list(Psi, Phi)) 
## dimnames(S1)[[1]] <- dimnames(S1)[[2]] <- c(paste("x",1:7,sep=""), 
##                                             paste("f",1:3,sep="")) 
## S1 
S1 <- as.mxMatrix(S1) 

## Factor loadings 
Lambda <- create.mxMatrix( c(".3*f1x1",".3*f1x2",".3*f1x3",rep(0,7), 
                             ".3*f2x4",".3*f2x5",rep(0,7),".3*f3x6",".3*f3x7"), 
                           type="Full", ncol=3, nrow=7, as.mxMatrix=FALSE ) 
Zero1 <- matrix(0, nrow=7, ncol=7) 
Zero2 <- matrix(0, nrow=3, ncol=10) 

## Create Amatrix 
A1 <- rbind( cbind(Zero1, Lambda), 
             Zero2 ) 
## dimnames(A1)[[1]] <- dimnames(A1)[[2]] <- c(paste("x",1:7,sep=""), 
##                                             paste("f",1:3,sep="")) 
## A1 
A1 <- as.mxMatrix(A1) 

## Create Fmatrix 
F1 <- create.Fmatrix(c(rep(1,7), rep(0,3))) 

#### Fixed-effects model: Stage 2 analysis 
fixed2 <- tssem2(fixed1, Amatrix=A1, Smatrix=S1, Fmatrix=F1, 
                 intervals.type="LB") 
summary(fixed2) 

## 
## Call:
## wls(Cov = coef.tssem1FEM(tssem1.obj), asyCov = vcov.tssem1FEM(tssem1.obj), 
##     n = sum(tssem1.obj$n), Amatrix = Amatrix, Smatrix = Smatrix, 
##     Fmatrix = Fmatrix, diag.constraints = diag.constraints, cor.analysis = tssem1.obj$cor.analysis, 
##     intervals.type = intervals.type, mx.algebras = mx.algebras, 
##     model.name = model.name, suppressWarnings = suppressWarnings, 
##     silent = silent, run = run)
## 
## 95% confidence intervals: Likelihood-based statistic
## Coefficients:
##         Estimate Std.Error  lbound  ubound z value Pr(>|z|)
## f1x1     0.51977        NA 0.50958 0.52997      NA       NA
## f1x2     0.64077        NA 0.63084 0.65073      NA       NA
## f1x3     0.66364        NA 0.65391 0.67339      NA       NA
## f2x4     0.69218        NA 0.68122 0.70326      NA       NA
## f2x5     0.53312        NA 0.52284 0.54341      NA       NA
## f3x6     0.84930        NA 0.84150 0.85714      NA       NA
## f3x7     0.77874        NA 0.77097 0.78652      NA       NA
## corf2f1  0.52611        NA 0.51176 0.54053      NA       NA
## corf3f1  0.42025        NA 0.40781 0.43266      NA       NA
## corf3f2  0.63540        NA 0.62323 0.64762      NA       NA
## 
## Goodness-of-fit indices:
##                                                 Value
## Sample size                                40596.0000
## Chi-square of target model                  1709.3864
## DF of target model                            11.0000
## p value of target model                        0.0000
## Number of constraints imposed on "Smatrix"     0.0000
## DF manually adjusted                           0.0000
## Chi-square of independence model           75296.1517
## DF of independence model                      21.0000
## RMSEA                                          0.0617
## RMSEA lower 95% CI                             0.0592
## RMSEA upper 95% CI                             0.0642
## SRMR                                           0.0375
## TLI                                            0.9569
## CFI                                            0.9774
## AIC                                         1687.3864
## BIC                                         1592.6607
## OpenMx status1: 0 ("0" or "1": The optimization is considered fine.
## Other values indicate problems.)

## } 

[Package metaSEM version 0.9.8 Index]