issp05 {metaSEM}	R Documentation

Data Set from ISSP (2005)

Description

Thirty-two covariance matrices on work-related attitudes were extracted from the International Social Survey Programme 2005: Work Orientation III (ISSP, 2005). Seven variables were selected for demonstration purposes. They were grouped into three constructs: Importance of Job Prospects measured by job security (JP1), high income (JP2), and opportunity for advancement (JP3); Importance of Job Autonomy measured by work independently (JA1) and decide time of work (JA2); and Importance of Contributions to Society measured by help other people (CS1) and a job useful to society (CS2).

Usage

data(issp05)

Details

A list of data with the following structure:

data

A list of 32 covariance matrices

n

A vector of sample sizes

Author(s)

Mike W.-L. Cheung <mikewlcheung@nus.edu.sg>

Source

ISSP Research Group (2007): International Social Survey Programme 2005: Work Orientation III (ISSP 2005). GESIS Data Archive, Cologne. ZA4350 Data file Version 1.0.0, doi:10.4232/1.4350

See Also

issp89

Examples

data(issp05)

#### Fixed-effects TSSEM
fixed1 <- tssem1(issp05$data, issp05$n, method="FEM")
summary(fixed1)

## 
## Call:
## tssem1FEM(my.df = my.df, n = n, cor.analysis = cor.analysis, 
##     model.name = model.name, cluster = cluster, suppressWarnings = suppressWarnings, 
##     silent = silent, run = run)
## 
## Coefficients:
##         Estimate Std.Error z value  Pr(>|z|)    
## S[1,2] 0.3458534 0.0043972  78.653 < 2.2e-16 ***
## S[1,3] 0.2957095 0.0045561  64.904 < 2.2e-16 ***
## S[1,4] 0.1451601 0.0048892  29.690 < 2.2e-16 ***
## S[1,5] 0.0902532 0.0049425  18.261 < 2.2e-16 ***
## S[1,6] 0.2058959 0.0047692  43.172 < 2.2e-16 ***
## S[1,7] 0.2178547 0.0047436  45.926 < 2.2e-16 ***
## S[2,3] 0.4039976 0.0041675  96.939 < 2.2e-16 ***
## S[2,4] 0.1941407 0.0047956  40.483 < 2.2e-16 ***
## S[2,5] 0.1905024 0.0047941  39.737 < 2.2e-16 ***
## S[2,6] 0.1348607 0.0048961  27.544 < 2.2e-16 ***
## S[2,7] 0.1432349 0.0048851  29.321 < 2.2e-16 ***
## S[3,4] 0.2780695 0.0046175  60.221 < 2.2e-16 ***
## S[3,5] 0.1882147 0.0048153  39.087 < 2.2e-16 ***
## S[3,6] 0.2530853 0.0046819  54.056 < 2.2e-16 ***
## S[3,7] 0.2500478 0.0046872  53.347 < 2.2e-16 ***
## S[4,5] 0.3628185 0.0043406  83.586 < 2.2e-16 ***
## S[4,6] 0.3779793 0.0042869  88.171 < 2.2e-16 ***
## S[4,7] 0.3178023 0.0044903  70.775 < 2.2e-16 ***
## S[5,6] 0.2647940 0.0046379  57.094 < 2.2e-16 ***
## S[5,7] 0.2541328 0.0046685  54.436 < 2.2e-16 ***
## S[6,7] 0.6623115 0.0028146 235.312 < 2.2e-16 ***
## ---
## Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
## 
## Goodness-of-fit indices:
##                                       Value
## Sample size                      40596.0000
## Chi-square of target model        5305.5769
## DF of target model                 651.0000
## p value of target model              0.0000
## Chi-square of independence model 61872.2338
## DF of independence model           672.0000
## RMSEA                                0.0751
## SRMR                                 0.0902
## TLI                                  0.9215
## CFI                                  0.9239
## AIC                               4003.5769
## BIC                              -1602.4607
## OpenMx status1: 6 ("0" or "1": The optimization is considered fine.
## Other values may indicate problems.)

## Prepare a model implied matrix
## Factor correlation matrix
Phi <- create.mxMatrix( c("0.3*corf2f1","0.3*corf3f1","0.3*corf3f2"),
                        type="Stand", as.mxMatrix=FALSE )

## Error variances
Psi <- create.mxMatrix( paste("0.2*e", 1:7, sep=""), type="Diag",
                        as.mxMatrix=FALSE )

## Create Smatrix
S1 <- bdiagMat(list(Psi, Phi))
## dimnames(S1)[[1]] <- dimnames(S1)[[2]] <- c(paste("x",1:7,sep=""),
##                                             paste("f",1:3,sep=""))
## S1
S1 <- as.mxMatrix(S1)

## Factor loadings
Lambda <- create.mxMatrix( c(".3*f1x1",".3*f1x2",".3*f1x3",rep(0,7),
                             ".3*f2x4",".3*f2x5",rep(0,7),".3*f3x6",".3*f3x7"),
                           type="Full", ncol=3, nrow=7, as.mxMatrix=FALSE )
Zero1 <- matrix(0, nrow=7, ncol=7)
Zero2 <- matrix(0, nrow=3, ncol=10)

## Create Amatrix
A1 <- rbind( cbind(Zero1, Lambda),
             Zero2 )
## dimnames(A1)[[1]] <- dimnames(A1)[[2]] <- c(paste("x",1:7,sep=""),
##                                             paste("f",1:3,sep=""))
## A1
A1 <- as.mxMatrix(A1)

## Create Fmatrix
F1 <- create.Fmatrix(c(rep(1,7), rep(0,3)))

#### Fixed-effects model: Stage 2 analysis
fixed2 <- tssem2(fixed1, Amatrix=A1, Smatrix=S1, Fmatrix=F1,
                 intervals.type="LB")
summary(fixed2)

## 
## Call:
## wls(Cov = pooledS, asyCov = tssem1.obj$acovS, n = sum(tssem1.obj$n), 
##     Amatrix = Amatrix, Smatrix = Smatrix, Fmatrix = Fmatrix, 
##     diag.constraints = diag.constraints, cor.analysis = cor.analysis, 
##     intervals.type = intervals.type, mx.algebras = mx.algebras, 
##     model.name = model.name, suppressWarnings = suppressWarnings, 
##     silent = silent, run = run)
## 
## 95% confidence intervals: Likelihood-based statistic
## Coefficients:
##         Estimate Std.Error  lbound  ubound z value Pr(>|z|)
## f1x1     0.51978        NA 0.50958 0.52998      NA       NA
## f1x2     0.64076        NA 0.63083 0.65073      NA       NA
## f1x3     0.66364        NA 0.65391 0.67340      NA       NA
## f2x4     0.69218        NA 0.68120 0.70326      NA       NA
## f2x5     0.53312        NA 0.52284 0.54342      NA       NA
## f3x6     0.84929        NA 0.84149 0.85713      NA       NA
## f3x7     0.77874        NA 0.77097 0.78652      NA       NA
## corf2f1  0.52611        NA 0.51175 0.54054      NA       NA
## corf3f1  0.42026        NA 0.40782 0.43268      NA       NA
## corf3f2  0.63541        NA 0.62324 0.64764      NA       NA
## 
## Goodness-of-fit indices:
##                                                 Value
## Sample size                                40596.0000
## Chi-square of target model                  1707.9305
## DF of target model                            11.0000
## p value of target model                        0.0000
## Number of constraints imposed on "Smatrix"     0.0000
## DF manually adjusted                           0.0000
## Chi-square of independence model           75251.3615
## DF of independence model                      21.0000
## RMSEA                                          0.0616
## SRMR                                           0.0375
## TLI                                            0.9569
## CFI                                            0.9774
## AIC                                         1685.9305
## BIC                                         1591.2048
## OpenMx status1: 0 ("0" or "1": The optimization is considered fine.
## Other values indicate problems.)

[Package metaSEM version 0.9.5-2 Index]