issp89 {metaSEM}	R Documentation

Data Set from Cheung and Chan (2005; 2009)

Description

Eleven covariance matrices on work-related attitudes were extracted from the Inter-University Consortium for Political and Social Research (1989). Nine variables were selected by Cheung and Chan (2005; 2009) for demonstration purposes. They were grouped into three constructs: Job Prospects measured by job security (JP1), income (JP2), and advancement opportunity (JP3); Job Nature measured by interesting job (JN1), independent work (JN2), help other people (JN3), and useful to society (JN4); and Time Demand measured by flexible working hours (TD1) and lots of leisure time (TD2).

Usage

data(issp89)

Details

A list of data with the following structure:

data

A list of 11 studies of covariance matrices

n

A vector of sample sizes

Author(s)

Mike W.-L. Cheung <mikewlcheung@nus.edu.sg>

Source

Inter-University Consortium for Political and Social Research. (1989). International Social Survey Program: Work orientation. Ann Arbor, MI: Author.

References

Cheung, M. W.-L., & Chan, W. (2005). Meta-analytic structural equation modeling: A two-stage approach. Psychological Methods, 10, 40-64.

Cheung, M. W.-L., & Chan, W. (2009). A two-stage approach to synthesizing covariance matrices in meta-analytic structural equation modeling. Structural Equation Modeling, 16, 28-53.

See Also

issp05

Examples

data(issp89)

#### Analysis of correlation structure in Cheung and Chan (2005)
#### Fixed-effects model: Stage 1 analysis
cor1 <- tssem1(issp89$data, issp89$n, method="FEM", cor.analysis=TRUE)
summary(cor1)

## 
## Call:
## tssem1FEM(my.df = my.df, n = n, cor.analysis = cor.analysis, 
##     model.name = model.name, cluster = cluster, suppressWarnings = suppressWarnings, 
##     silent = silent, run = run)
## 
## Coefficients:
##         Estimate Std.Error z value  Pr(>|z|)    
## S[1,2] 0.2849397 0.0108864 26.1740 < 2.2e-16 ***
## S[1,3] 0.2452271 0.0111613 21.9712 < 2.2e-16 ***
## S[1,4] 0.1972493 0.0113946 17.3108 < 2.2e-16 ***
## S[1,5] 0.1637938 0.0115457 14.1865 < 2.2e-16 ***
## S[1,6] 0.1654163 0.0115125 14.3684 < 2.2e-16 ***
## S[1,7] 0.1679298 0.0115042 14.5973 < 2.2e-16 ***
## S[1,8] 0.0631791 0.0118042  5.3523 8.685e-08 ***
## S[1,9] 0.0967567 0.0117318  8.2474 2.220e-16 ***
## S[2,3] 0.4254660 0.0097171 43.7854 < 2.2e-16 ***
## S[2,4] 0.2267410 0.0112400 20.1726 < 2.2e-16 ***
## S[2,5] 0.2033542 0.0113607 17.8998 < 2.2e-16 ***
## S[2,6] 0.0671466 0.0118030  5.6890 1.278e-08 ***
## S[2,7] 0.0337135 0.0118339  2.8489  0.004387 ** 
## S[2,8] 0.1401964 0.0116357 12.0488 < 2.2e-16 ***
## S[2,9] 0.1170028 0.0116911 10.0079 < 2.2e-16 ***
## S[3,4] 0.2855816 0.0108872 26.2311 < 2.2e-16 ***
## S[3,5] 0.1881168 0.0114313 16.4563 < 2.2e-16 ***
## S[3,6] 0.1132342 0.0117030  9.6757 < 2.2e-16 ***
## S[3,7] 0.0844455 0.0117757  7.1712 7.436e-13 ***
## S[3,8] 0.1543047 0.0115813 13.3236 < 2.2e-16 ***
## S[3,9] 0.1154627 0.0116934  9.8742 < 2.2e-16 ***
## S[4,5] 0.4226017 0.0097391 43.3924 < 2.2e-16 ***
## S[4,6] 0.3750854 0.0101844 36.8293 < 2.2e-16 ***
## S[4,7] 0.3594932 0.0103347 34.7849 < 2.2e-16 ***
## S[4,8] 0.1546478 0.0115585 13.3796 < 2.2e-16 ***
## S[4,9] 0.0267351 0.0118564  2.2549  0.024139 *  
## S[5,6] 0.2771064 0.0109499 25.3066 < 2.2e-16 ***
## S[5,7] 0.2162919 0.0112925 19.1535 < 2.2e-16 ***
## S[5,8] 0.2500515 0.0110990 22.5291 < 2.2e-16 ***
## S[5,9] 0.0257200 0.0118508  2.1703  0.029982 *  
## S[6,7] 0.5460047 0.0083277 65.5646 < 2.2e-16 ***
## S[6,8] 0.1092173 0.0117113  9.3258 < 2.2e-16 ***
## S[6,9] 0.0360578 0.0118346  3.0468  0.002313 ** 
## S[7,8] 0.1108463 0.0117080  9.4676 < 2.2e-16 ***
## S[7,9] 0.0373362 0.0118415  3.1530  0.001616 ** 
## S[8,9] 0.1697427 0.0116102 14.6201 < 2.2e-16 ***
## ---
## Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
## 
## Goodness-of-fit indices:
##                                       Value
## Sample size                       7155.0000
## Chi-square of target model         941.4240
## DF of target model                 360.0000
## p value of target model              0.0000
## Chi-square of independence model 10559.1411
## DF of independence model           396.0000
## RMSEA                                0.0498
## SRMR                                 0.0641
## TLI                                  0.9371
## CFI                                  0.9428
## AIC                                221.4240
## BIC                              -2253.7800
## OpenMx status1: 0 ("0" or "1": The optimization is considered fine.
## Other values indicate problems.)

## Prepare a model implied matrix
## Factor correlation matrix
Phi <- create.mxMatrix( c("0.3*corf2f1","0.3*corf3f1","0.3*corf3f2"),
                        type="Stand", as.mxMatrix=FALSE )
## Error variances
Psi <- create.mxMatrix( paste("0.2*e", 1:9, sep=""), type="Diag",
                        as.mxMatrix=FALSE )

## Create Smatrix
S1 <- bdiagMat(list(Psi, Phi))
## dimnames(S1)[[1]] <- dimnames(S1)[[2]] <- c(paste("x",1:9,sep=""),
##                                             paste("f",1:3,sep=""))
## S1
S1 <- as.mxMatrix(S1)

## Factor loadings
Lambda <- create.mxMatrix( c(".3*f1x1",".3*f1x2",".3*f1x3",rep(0,9),
                             ".3*f2x4",".3*f2x5",".3*f2x6",".3*f2x7",
                             rep(0,9),".3*f3x8",".3*f3x9"), type="Full",
                             ncol=3, nrow=9, as.mxMatrix=FALSE )
Zero1 <- matrix(0, nrow=9, ncol=9)
Zero2 <- matrix(0, nrow=3, ncol=12)

## Create Amatrix
A1 <- rbind( cbind(Zero1, Lambda),
             Zero2 )
## dimnames(A1)[[1]] <- dimnames(A1)[[2]] <- c(paste("x",1:9,sep=""),
##                                             paste("f",1:3,sep=""))
## A1
A1 <- as.mxMatrix(A1)

## Create Fmatrix
F1 <- create.Fmatrix(c(rep(1,9), rep(0,3)))

#### Fixed-effects model: Stage 2 analysis
cor2 <- tssem2(cor1, Amatrix=A1, Smatrix=S1, Fmatrix=F1, intervals.type="LB")
summary(cor2)

## 
## Call:
## wls(Cov = pooledS, asyCov = tssem1.obj$acovS, n = sum(tssem1.obj$n), 
##     Amatrix = Amatrix, Smatrix = Smatrix, Fmatrix = Fmatrix, 
##     diag.constraints = diag.constraints, cor.analysis = cor.analysis, 
##     intervals.type = intervals.type, mx.algebras = mx.algebras, 
##     model.name = model.name, suppressWarnings = suppressWarnings, 
##     silent = silent, run = run)
## 
## 95% confidence intervals: Likelihood-based statistic
## Coefficients:
##         Estimate Std.Error  lbound  ubound z value Pr(>|z|)
## f1x1     0.47467        NA 0.44925 0.50003      NA       NA
## f1x2     0.61402        NA 0.59027 0.63803      NA       NA
## f1x3     0.63450        NA 0.61046 0.65881      NA       NA
## f2x4     0.69578        NA 0.67825 0.71320      NA       NA
## f2x5     0.56077        NA 0.53993 0.58143      NA       NA
## f2x6     0.71612        NA 0.69938 0.73290      NA       NA
## f2x7     0.68410        NA 0.66653 0.70171      NA       NA
## f3x8     0.63941        NA 0.57515 0.71819      NA       NA
## f3x9     0.29451        NA 0.25462 0.33422      NA       NA
## corf2f1  0.49112        NA 0.46158 0.52050      NA       NA
## corf3f1  0.37978        NA 0.32093 0.43858      NA       NA
## corf3f2  0.39782        NA 0.35321 0.44418      NA       NA
## 
## Goodness-of-fit indices:
##                                                Value
## Sample size                                7155.0000
## Chi-square of target model                 1272.4563
## DF of target model                           24.0000
## p value of target model                       0.0000
## Number of constraints imposed on "Smatrix"    0.0000
## DF manually adjusted                          0.0000
## Chi-square of independence model           9413.4047
## DF of independence model                     36.0000
## RMSEA                                         0.0853
## SRMR                                          0.0763
## TLI                                           0.8003
## CFI                                           0.8669
## AIC                                        1224.4563
## BIC                                        1059.4427
## OpenMx status1: 0 ("0" or "1": The optimization is considered fine.
## Other values indicate problems.)

#### Analysis of covariance structure in Cheung and Chan (2009)
#### Fixed-effects model: Stage 1 analysis
cov1 <- tssem1(issp89$data, issp89$n, method="FEM", cor.analysis=FALSE)
summary(cov1)

## 
## Call:
## tssem1FEM(my.df = my.df, n = n, cor.analysis = cor.analysis, 
##     model.name = model.name, cluster = cluster, suppressWarnings = suppressWarnings, 
##     silent = silent, run = run)
## 
## Coefficients:
##        Estimate Std.Error z value  Pr(>|z|)    
## S[1,1] 1.104919  0.018488 59.7640 < 2.2e-16 ***
## S[1,2] 0.295168  0.012863 22.9474 < 2.2e-16 ***
## S[1,3] 0.269277  0.013666 19.7045 < 2.2e-16 ***
## S[1,4] 0.187299  0.011757 15.9308 < 2.2e-16 ***
## S[1,5] 0.159311  0.012345 12.9047 < 2.2e-16 ***
## S[1,6] 0.179443  0.013150 13.6461 < 2.2e-16 ***
## S[1,7] 0.172348  0.012348 13.9572 < 2.2e-16 ***
## S[1,8] 0.081752  0.016151  5.0617 4.155e-07 ***
## S[1,9] 0.109184  0.013160  8.2965 < 2.2e-16 ***
## S[2,2] 0.990709  0.016577 59.7635 < 2.2e-16 ***
## S[2,3] 0.449570  0.013662 32.9073 < 2.2e-16 ***
## S[2,4] 0.207704  0.011206 18.5350 < 2.2e-16 ***
## S[2,5] 0.195933  0.011783 16.6283 < 2.2e-16 ***
## S[2,6] 0.074682  0.012320  6.0617 1.347e-09 ***
## S[2,7] 0.034254  0.011540  2.9684 0.0029939 ** 
## S[2,8] 0.182014  0.015417 11.8057 < 2.2e-16 ***
## S[2,9] 0.122277  0.012485  9.7936 < 2.2e-16 ***
## S[3,3] 1.141567  0.019102 59.7608 < 2.2e-16 ***
## S[3,4] 0.284279  0.012209 23.2839 < 2.2e-16 ***
## S[3,5] 0.194638  0.012614 15.4307 < 2.2e-16 ***
## S[3,6] 0.131606  0.013283  9.9081 < 2.2e-16 ***
## S[3,7] 0.095593  0.012431  7.6898 1.465e-14 ***
## S[3,8] 0.210984  0.016577 12.7274 < 2.2e-16 ***
## S[3,9] 0.127952  0.013398  9.5500 < 2.2e-16 ***
## S[4,4] 0.861721  0.014420 59.7592 < 2.2e-16 ***
## S[4,5] 0.378570  0.011669 32.4416 < 2.2e-16 ***
## S[4,6] 0.357258  0.012216 29.2459 < 2.2e-16 ***
## S[4,7] 0.325259  0.011424 28.4725 < 2.2e-16 ***
## S[4,8] 0.185780  0.014407 12.8948 < 2.2e-16 ***
## S[4,9] 0.025314  0.011570  2.1879 0.0286774 *  
## S[5,5] 0.962181  0.016101 59.7596 < 2.2e-16 ***
## S[5,6] 0.278357  0.012550 22.1790 < 2.2e-16 ***
## S[5,7] 0.202481  0.011615 17.4323 < 2.2e-16 ***
## S[5,8] 0.317355  0.015507 20.4652 < 2.2e-16 ***
## S[5,9] 0.023931  0.012225  1.9575 0.0502874 .  
## S[6,6] 1.088649  0.018216 59.7632 < 2.2e-16 ***
## S[6,7] 0.558221  0.013775 40.5235 < 2.2e-16 ***
## S[6,8] 0.141423  0.016090  8.7895 < 2.2e-16 ***
## S[6,9] 0.042626  0.013010  3.2765 0.0010512 ** 
## S[7,7] 0.958835  0.016044 59.7642 < 2.2e-16 ***
## S[7,8] 0.140198  0.015110  9.2788 < 2.2e-16 ***
## S[7,9] 0.046393  0.012213  3.7988 0.0001454 ***
## S[8,8] 1.680194  0.028114 59.7627 < 2.2e-16 ***
## S[8,9] 0.222825  0.016364 13.6168 < 2.2e-16 ***
## S[9,9] 1.108923  0.018554 59.7660 < 2.2e-16 ***
## ---
## Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
## 
## Goodness-of-fit indices:
##                                       Value
## Sample size                       7155.0000
## Chi-square of target model        2509.9297
## DF of target model                 450.0000
## p value of target model              0.0000
## Chi-square of independence model 10559.1411
## DF of independence model           396.0000
## RMSEA                                0.0839
## SRMR                                 0.1220
## TLI                                  0.8216
## CFI                                  0.7973
## AIC                               1609.9297
## BIC                              -1484.0754
## OpenMx status1: 0 ("0" or "1": The optimization is considered fine.
## Other values indicate problems.)

#### Fixed-effects model: Stage 2 analysis
cov2 <- tssem2(cov1, Amatrix=A1, Smatrix=S1, Fmatrix=F1)              
summary(cov2)

## 
## Call:
## wls(Cov = pooledS, asyCov = tssem1.obj$acovS, n = sum(tssem1.obj$n), 
##     Amatrix = Amatrix, Smatrix = Smatrix, Fmatrix = Fmatrix, 
##     diag.constraints = diag.constraints, cor.analysis = cor.analysis, 
##     intervals.type = intervals.type, mx.algebras = mx.algebras, 
##     model.name = model.name, suppressWarnings = suppressWarnings, 
##     silent = silent, run = run)
## 
## 95% confidence intervals: z statistic approximation
## Coefficients:
##         Estimate Std.Error   lbound   ubound z value  Pr(>|z|)    
## f1x1    0.467878  0.016282 0.435965 0.499790  28.736 < 2.2e-16 ***
## f1x2    0.582153  0.015585 0.551607 0.612698  37.354 < 2.2e-16 ***
## f1x3    0.640230  0.016959 0.606992 0.673468  37.753 < 2.2e-16 ***
## f2x4    0.597636  0.013438 0.571298 0.623974  44.473 < 2.2e-16 ***
## f2x5    0.488876  0.015799 0.457910 0.519841  30.943 < 2.2e-16 ***
## f2x6    0.698877  0.014864 0.669744 0.728010  47.018 < 2.2e-16 ***
## f2x7    0.624076  0.014524 0.595610 0.652541  42.970 < 2.2e-16 ***
## f3x8    0.733322  0.048424 0.638413 0.828231  15.144 < 2.2e-16 ***
## f3x9    0.319455  0.025096 0.270268 0.368642  12.729 < 2.2e-16 ***
## e1      0.856325  0.017403 0.822216 0.890433  49.207 < 2.2e-16 ***
## e2      0.577585  0.015998 0.546230 0.608941  36.104 < 2.2e-16 ***
## e3      0.656541  0.018862 0.619571 0.693510  34.807 < 2.2e-16 ***
## e4      0.453736  0.011097 0.431987 0.475485  40.890 < 2.2e-16 ***
## e5      0.623193  0.012993 0.597727 0.648659  47.964 < 2.2e-16 ***
## e6      0.524811  0.014175 0.497028 0.552593  37.023 < 2.2e-16 ***
## e7      0.472836  0.012470 0.448395 0.497276  37.918 < 2.2e-16 ***
## e8      1.048437  0.068106 0.914952 1.181923  15.394 < 2.2e-16 ***
## e9      0.984197  0.020844 0.943344 1.025050  47.218 < 2.2e-16 ***
## corf2f1 0.474884  0.018681 0.438269 0.511499  25.420 < 2.2e-16 ***
## corf3f1 0.396286  0.033675 0.330285 0.462287  11.768 < 2.2e-16 ***
## corf3f2 0.385322  0.025528 0.335288 0.435357  15.094 < 2.2e-16 ***
## ---
## Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
## 
## Goodness-of-fit indices:
##                                                Value
## Sample size                                7155.0000
## Chi-square of target model                 1164.3846
## DF of target model                           24.0000
## p value of target model                       0.0000
## Number of constraints imposed on "Smatrix"    0.0000
## DF manually adjusted                          0.0000
## Chi-square of independence model           4080.4128
## DF of independence model                     36.0000
## RMSEA                                         0.0815
## SRMR                                          0.0673
## TLI                                           0.5771
## CFI                                           0.7180
## AIC                                        1116.3846
## BIC                                         951.3710
## OpenMx status1: 0 ("0" or "1": The optimization is considered fine.
## Other values indicate problems.)

[Package metaSEM version 0.9.3-1 Index]