issp89 {metaSEM}	R Documentation

Data Set from Cheung and Chan (2005; 2009)

Description

Eleven covariance matrices on work-related attitudes were extracted from the Inter-University Consortium for Political and Social Research (1989). Nine variables were selected by Cheung and Chan (2005; 2009) for demonstration purposes. They were grouped into three constructs: Job Prospects measured by job security (JP1), income (JP2), and advancement opportunity (JP3); Job Nature measured by interesting job (JN1), independent work (JN2), help other people (JN3), and useful to society (JN4); and Time Demand measured by flexible working hours (TD1) and lots of leisure time (TD2).

Usage

data(issp89)

Details

A list of data with the following structure:

data

A list of 11 studies of covariance matrices

n

A vector of sample sizes

Author(s)

Mike W.-L. Cheung <mikewlcheung@nus.edu.sg>

Source

Inter-University Consortium for Political and Social Research. (1989). International Social Survey Program: Work orientation. Ann Arbor, MI: Author.

References

Cheung, M. W.-L., & Chan, W. (2005). Meta-analytic structural equation modeling: A two-stage approach. Psychological Methods, 10, 40-64.

Cheung, M. W.-L., & Chan, W. (2009). A two-stage approach to synthesizing covariance matrices in meta-analytic structural equation modeling. Structural Equation Modeling, 16, 28-53.

See Also

issp05

Examples

## \dontrun{ 
data(issp89) 

#### Analysis of correlation structure in Cheung and Chan (2005) 
#### Fixed-effects model: Stage 1 analysis 
cor1 <- tssem1(issp89$data, issp89$n, method="FEM", cor.analysis=TRUE) 
summary(cor1) 

## 
## Call:
## tssem1FEM(my.df = my.df, n = n, cor.analysis = cor.analysis, 
##     model.name = model.name, cluster = cluster, suppressWarnings = suppressWarnings, 
##     silent = silent, run = run)
## 
## Coefficients:
##         Estimate Std.Error z value  Pr(>|z|)    
## S[1,2] 0.2849491 0.0108772 26.1969 < 2.2e-16 ***
## S[1,3] 0.2452415 0.0111519 21.9911 < 2.2e-16 ***
## S[1,4] 0.1972593 0.0113850 17.3263 < 2.2e-16 ***
## S[1,5] 0.1637982 0.0115358 14.1991 < 2.2e-16 ***
## S[1,6] 0.1654192 0.0115029 14.3807 < 2.2e-16 ***
## S[1,7] 0.1679291 0.0114948 14.6092 < 2.2e-16 ***
## S[1,8] 0.0631833 0.0117944  5.3570 8.460e-08 ***
## S[1,9] 0.0967622 0.0117225  8.2544 2.220e-16 ***
## S[2,3] 0.4254753 0.0097088 43.8237 < 2.2e-16 ***
## S[2,4] 0.2267388 0.0112305 20.1896 < 2.2e-16 ***
## S[2,5] 0.2033488 0.0113509 17.9148 < 2.2e-16 ***
## S[2,6] 0.0671540 0.0117928  5.6945 1.238e-08 ***
## S[2,7] 0.0337245 0.0118238  2.8522  0.004341 ** 
## S[2,8] 0.1401967 0.0116260 12.0589 < 2.2e-16 ***
## S[2,9] 0.1170073 0.0116817 10.0163 < 2.2e-16 ***
## S[3,4] 0.2855866 0.0108779 26.2538 < 2.2e-16 ***
## S[3,5] 0.1881115 0.0114213 16.4702 < 2.2e-16 ***
## S[3,6] 0.1132368 0.0116930  9.6842 < 2.2e-16 ***
## S[3,7] 0.0844460 0.0117658  7.1773 7.112e-13 ***
## S[3,8] 0.1543058 0.0115717 13.3348 < 2.2e-16 ***
## S[3,9] 0.1154633 0.0116840  9.8821 < 2.2e-16 ***
## S[4,5] 0.4225988 0.0097310 43.4281 < 2.2e-16 ***
## S[4,6] 0.3750932 0.0101759 36.8609 < 2.2e-16 ***
## S[4,7] 0.3595015 0.0103263 34.8141 < 2.2e-16 ***
## S[4,8] 0.1546517 0.0115492 13.3907 < 2.2e-16 ***
## S[4,9] 0.0267537 0.0118468  2.2583  0.023926 *  
## S[5,6] 0.2771136 0.0109409 25.3283 < 2.2e-16 ***
## S[5,7] 0.2162954 0.0112835 19.1692 < 2.2e-16 ***
## S[5,8] 0.2500501 0.0110903 22.5467 < 2.2e-16 ***
## S[5,9] 0.0257266 0.0118413  2.1726  0.029810 *  
## S[6,7] 0.5460141 0.0083209 65.6195 < 2.2e-16 ***
## S[6,8] 0.1092225 0.0117020  9.3337 < 2.2e-16 ***
## S[6,9] 0.0360659 0.0118252  3.0499  0.002289 ** 
## S[7,8] 0.1108534 0.0116987  9.4757 < 2.2e-16 ***
## S[7,9] 0.0373419 0.0118321  3.1560  0.001600 ** 
## S[8,9] 0.1697426 0.0116010 14.6317 < 2.2e-16 ***
## ---
## Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
## 
## Goodness-of-fit indices:
##                                       Value
## Sample size                       7155.0000
## Chi-square of target model         942.8555
## DF of target model                 360.0000
## p value of target model              0.0000
## Chi-square of independence model 10575.7274
## DF of independence model           396.0000
## RMSEA                                0.0499
## RMSEA lower 95% CI                   0.0461
## RMSEA upper 95% CI                   0.0538
## SRMR                                 0.0641
## TLI                                  0.9370
## CFI                                  0.9427
## AIC                                222.8555
## BIC                              -2252.3485
## OpenMx status1: 6 ("0" or "1": The optimization is considered fine.
## Other values may indicate problems.)

## Prepare a model implied matrix 
## Factor correlation matrix 
Phi <- create.mxMatrix( c("0.3*corf2f1","0.3*corf3f1","0.3*corf3f2"), 
                        type="Stand", as.mxMatrix=FALSE ) 
## Error variances 
Psi <- create.mxMatrix( paste("0.2*e", 1:9, sep=""), type="Diag", 
                        as.mxMatrix=FALSE ) 

## Create Smatrix 
S1 <- bdiagMat(list(Psi, Phi)) 
## dimnames(S1)[[1]] <- dimnames(S1)[[2]] <- c(paste("x",1:9,sep=""), 
##                                             paste("f",1:3,sep="")) 
## S1 
S1 <- as.mxMatrix(S1) 

## Factor loadings 
Lambda <- create.mxMatrix( c(".3*f1x1",".3*f1x2",".3*f1x3",rep(0,9), 
                             ".3*f2x4",".3*f2x5",".3*f2x6",".3*f2x7", 
                             rep(0,9),".3*f3x8",".3*f3x9"), type="Full", 
                             ncol=3, nrow=9, as.mxMatrix=FALSE ) 
Zero1 <- matrix(0, nrow=9, ncol=9) 
Zero2 <- matrix(0, nrow=3, ncol=12) 

## Create Amatrix 
A1 <- rbind( cbind(Zero1, Lambda), 
             Zero2 ) 
## dimnames(A1)[[1]] <- dimnames(A1)[[2]] <- c(paste("x",1:9,sep=""), 
##                                             paste("f",1:3,sep="")) 
## A1 
A1 <- as.mxMatrix(A1) 

## Create Fmatrix 
F1 <- create.Fmatrix(c(rep(1,9), rep(0,3))) 

#### Fixed-effects model: Stage 2 analysis 
cor2 <- tssem2(cor1, Amatrix=A1, Smatrix=S1, Fmatrix=F1, intervals.type="LB") 
summary(cor2) 

## 
## Call:
## wls(Cov = coef.tssem1FEM(tssem1.obj), asyCov = vcov.tssem1FEM(tssem1.obj), 
##     n = sum(tssem1.obj$n), Amatrix = Amatrix, Smatrix = Smatrix, 
##     Fmatrix = Fmatrix, diag.constraints = diag.constraints, cor.analysis = tssem1.obj$cor.analysis, 
##     intervals.type = intervals.type, mx.algebras = mx.algebras, 
##     model.name = model.name, suppressWarnings = suppressWarnings, 
##     silent = silent, run = run)
## 
## 95% confidence intervals: Likelihood-based statistic
## Coefficients:
##         Estimate Std.Error  lbound  ubound z value Pr(>|z|)
## f1x1     0.47466        NA 0.44926 0.50005      NA       NA
## f1x2     0.61402        NA 0.59029 0.63805      NA       NA
## f1x3     0.63449        NA 0.61044 0.65887      NA       NA
## f2x4     0.69577        NA 0.67825 0.71318      NA       NA
## f2x5     0.56075        NA 0.53990 0.58139      NA       NA
## f2x6     0.71613        NA 0.69940 0.73292      NA       NA
## f2x7     0.68411        NA 0.66655 0.70181      NA       NA
## f3x8     0.63944        NA 0.57520 0.67972      NA       NA
## f3x9     0.29448        NA 0.25463 0.33425      NA       NA
## corf2f1  0.49107        NA 0.46153 0.52050      NA       NA
## corf3f1  0.37971        NA 0.32093 0.43844      NA       NA
## corf3f2  0.39779        NA 0.35309 0.44416      NA       NA
## 
## Goodness-of-fit indices:
##                                                Value
## Sample size                                7155.0000
## Chi-square of target model                 1274.3957
## DF of target model                           24.0000
## p value of target model                       0.0000
## Number of constraints imposed on "Smatrix"    0.0000
## DF manually adjusted                          0.0000
## Chi-square of independence model           9431.8521
## DF of independence model                     36.0000
## RMSEA                                         0.0853
## RMSEA lower 95% CI                            0.0814
## RMSEA upper 95% CI                            0.0894
## SRMR                                          0.0763
## TLI                                           0.8004
## CFI                                           0.8669
## AIC                                        1226.3957
## BIC                                        1061.3821
## OpenMx status1: 0 ("0" or "1": The optimization is considered fine.
## Other values indicate problems.)

#### Analysis of covariance structure in Cheung and Chan (2009) 
#### Fixed-effects model: Stage 1 analysis 
cov1 <- tssem1(issp89$data, issp89$n, method="FEM", cor.analysis=FALSE) 
summary(cov1) 

## 
## Call:
## tssem1FEM(my.df = my.df, n = n, cor.analysis = cor.analysis, 
##     model.name = model.name, cluster = cluster, suppressWarnings = suppressWarnings, 
##     silent = silent, run = run)
## 
## Coefficients:
##        Estimate Std.Error z value  Pr(>|z|)    
## S[1,1] 1.103222  0.018446 59.8093 < 2.2e-16 ***
## S[1,2] 0.294716  0.012833 22.9649 < 2.2e-16 ***
## S[1,3] 0.268865  0.013634 19.7201 < 2.2e-16 ***
## S[1,4] 0.187013  0.011730 15.9431 < 2.2e-16 ***
## S[1,5] 0.159068  0.012317 12.9145 < 2.2e-16 ***
## S[1,6] 0.179169  0.013120 13.6563 < 2.2e-16 ***
## S[1,7] 0.172085  0.012320 13.9679 < 2.2e-16 ***
## S[1,8] 0.081628  0.016114  5.0657 4.069e-07 ***
## S[1,9] 0.109017  0.013130  8.3030 < 2.2e-16 ***
## S[2,2] 0.989187  0.016539 59.8096 < 2.2e-16 ***
## S[2,3] 0.448881  0.013630 32.9336 < 2.2e-16 ***
## S[2,4] 0.207387  0.011180 18.5496 < 2.2e-16 ***
## S[2,5] 0.195633  0.011756 16.6411 < 2.2e-16 ***
## S[2,6] 0.074569  0.012292  6.0664 1.308e-09 ***
## S[2,7] 0.034203  0.011513  2.9708 0.0029699 ** 
## S[2,8] 0.181736  0.015382 11.8149 < 2.2e-16 ***
## S[2,9] 0.122091  0.012457  9.8011 < 2.2e-16 ***
## S[3,3] 1.139815  0.019058 59.8086 < 2.2e-16 ***
## S[3,4] 0.283844  0.012181 23.3030 < 2.2e-16 ***
## S[3,5] 0.194341  0.012584 15.4431 < 2.2e-16 ***
## S[3,6] 0.131406  0.013252  9.9156 < 2.2e-16 ***
## S[3,7] 0.095448  0.012402  7.6960 1.399e-14 ***
## S[3,8] 0.210661  0.016538 12.7376 < 2.2e-16 ***
## S[3,9] 0.127757  0.013367  9.5574 < 2.2e-16 ***
## S[4,4] 0.860398  0.014387 59.8056 < 2.2e-16 ***
## S[4,5] 0.377990  0.011642 32.4671 < 2.2e-16 ***
## S[4,6] 0.356711  0.012188 29.2674 < 2.2e-16 ***
## S[4,7] 0.324761  0.011397 28.4945 < 2.2e-16 ***
## S[4,8] 0.185497  0.014374 12.9052 < 2.2e-16 ***
## S[4,9] 0.025276  0.011543  2.1897 0.0285475 *  
## S[5,5] 0.960703  0.016064 59.8062 < 2.2e-16 ***
## S[5,6] 0.277931  0.012522 22.1956 < 2.2e-16 ***
## S[5,7] 0.202172  0.011588 17.4459 < 2.2e-16 ***
## S[5,8] 0.316869  0.015471 20.4815 < 2.2e-16 ***
## S[5,9] 0.023895  0.012197  1.9591 0.0500989 .  
## S[6,6] 1.086978  0.018174 59.8082 < 2.2e-16 ***
## S[6,7] 0.557365  0.013744 40.5539 < 2.2e-16 ***
## S[6,8] 0.141208  0.016053  8.7962 < 2.2e-16 ***
## S[6,9] 0.042562  0.012980  3.2791 0.0010415 ** 
## S[7,7] 0.957363  0.016007 59.8095 < 2.2e-16 ***
## S[7,8] 0.139985  0.015075  9.2860 < 2.2e-16 ***
## S[7,9] 0.046323  0.012184  3.8018 0.0001436 ***
## S[8,8] 1.677613  0.028049 59.8092 < 2.2e-16 ***
## S[8,9] 0.222483  0.016326 13.6273 < 2.2e-16 ***
## S[9,9] 1.107219  0.018512 59.8121 < 2.2e-16 ***
## ---
## Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
## 
## Goodness-of-fit indices:
##                                       Value
## Sample size                       7155.0000
## Chi-square of target model        2514.0348
## DF of target model                 450.0000
## p value of target model              0.0000
## Chi-square of independence model 10575.7274
## DF of independence model           396.0000
## RMSEA                                0.0840
## RMSEA lower 95% CI                   0.0809
## RMSEA upper 95% CI                   0.0873
## SRMR                                 0.1217
## TLI                                  0.8216
## CFI                                  0.7972
## AIC                               1614.0348
## BIC                              -1479.9702
## OpenMx status1: 0 ("0" or "1": The optimization is considered fine.
## Other values may indicate problems.)

#### Fixed-effects model: Stage 2 analysis 
cov2 <- tssem2(cov1, Amatrix=A1, Smatrix=S1, Fmatrix=F1)               
summary(cov2) 

## 
## Call:
## wls(Cov = coef.tssem1FEM(tssem1.obj), asyCov = vcov.tssem1FEM(tssem1.obj), 
##     n = sum(tssem1.obj$n), Amatrix = Amatrix, Smatrix = Smatrix, 
##     Fmatrix = Fmatrix, diag.constraints = diag.constraints, cor.analysis = tssem1.obj$cor.analysis, 
##     intervals.type = intervals.type, mx.algebras = mx.algebras, 
##     model.name = model.name, suppressWarnings = suppressWarnings, 
##     silent = silent, run = run)
## 
## 95% confidence intervals: z statistic approximation
## Coefficients:
##         Estimate Std.Error   lbound   ubound z value  Pr(>|z|)    
## f1x1    0.467533  0.016257 0.435670 0.499397  28.759 < 2.2e-16 ***
## f1x2    0.581714  0.015560 0.551217 0.612212  37.385 < 2.2e-16 ***
## f1x3    0.639752  0.016931 0.606567 0.672937  37.785 < 2.2e-16 ***
## f2x4    0.597189  0.013417 0.570893 0.623485  44.511 < 2.2e-16 ***
## f2x5    0.488520  0.015774 0.457605 0.519436  30.971 < 2.2e-16 ***
## f2x6    0.698333  0.014841 0.669244 0.727421  47.053 < 2.2e-16 ***
## f2x7    0.623595  0.014501 0.595173 0.652017  43.003 < 2.2e-16 ***
## f3x8    0.732766  0.048345 0.638013 0.827520  15.157 < 2.2e-16 ***
## f3x9    0.319213  0.025055 0.270106 0.368321  12.740 < 2.2e-16 ***
## e1      0.855005  0.017362 0.820975 0.889034  49.245 < 2.2e-16 ***
## e2      0.576701  0.015961 0.545419 0.607984  36.133 < 2.2e-16 ***
## e3      0.655532  0.018819 0.618648 0.692416  34.834 < 2.2e-16 ***
## e4      0.453037  0.011071 0.431338 0.474735  40.921 < 2.2e-16 ***
## e5      0.622231  0.012963 0.596824 0.647639  48.000 < 2.2e-16 ***
## e6      0.524014  0.014142 0.496296 0.551732  37.054 < 2.2e-16 ***
## e7      0.472106  0.012441 0.447722 0.496490  37.947 < 2.2e-16 ***
## e8      1.046833  0.067944 0.913665 1.180001  15.407 < 2.2e-16 ***
## e9      0.982685  0.020795 0.941927 1.023442  47.255 < 2.2e-16 ***
## corf2f1 0.474907  0.018666 0.438323 0.511490  25.443 < 2.2e-16 ***
## corf3f1 0.396303  0.033647 0.330357 0.462250  11.778 < 2.2e-16 ***
## corf3f2 0.385354  0.025508 0.335360 0.435349  15.107 < 2.2e-16 ***
## ---
## Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
## 
## Goodness-of-fit indices:
##                                                Value
## Sample size                                7155.0000
## Chi-square of target model                 1166.1825
## DF of target model                           24.0000
## p value of target model                       0.0000
## Number of constraints imposed on "Smatrix"    0.0000
## DF manually adjusted                          0.0000
## Chi-square of independence model           4086.6879
## DF of independence model                     36.0000
## RMSEA                                         0.0816
## RMSEA lower 95% CI                            0.0776
## RMSEA upper 95% CI                            0.0856
## SRMR                                          0.0673
## TLI                                           0.5770
## CFI                                           0.7180
## AIC                                        1118.1825
## BIC                                         953.1689
## OpenMx status1: 0 ("0" or "1": The optimization is considered fine.
## Other values indicate problems.)

## } 

[Package metaSEM version 0.9.8 Index]