wvs94b {metaSEM}	R Documentation

Forty-four Covariance Matrices on Life Satisfaction, Job Satisfaction, and Job Autonomy

Description

Between 1990 and 1993, 57,561 adults aged 18 and above from 42 nations were interviewed by local academic institutes in Eastern European nations and by professional survey organizations in other nations. The covariance matrices among Life Satisfaction, Job Satisfaction, and Job Autonomy were calculated.

Usage

data(wvs94b)

Details

The variables are:

data

Covariance matrix among Life Satisfaction (LS), Job Satisfaction (JS), and Job Autonomy (JA)

n

Sample size in the country

Source

World Values Study Group. (1994). World Values Survey, 1981-1984 and 1990-1993 [Computer file]. Ann Arbor, MI: Inter-university Consortium for Political and Social Research.

References

Au, K., & Cheung, M. W.-L. (2004). Intra-cultural variation and job autonomy in 42 countries. Organization Studies, 25, 1339-1362.

Examples

## Not run: 
##D data(wvs94b)
##D 
##D ## Get the indirect and the direct effects and
##D ## their sampling covariance matrices for each study
##D indirect1 <- indirectEffect(wvs94b$data, wvs94b$n)
##D indirect1
##D 
##D ## Multivariate meta-analysis on the indirect and direct effects
##D indirect2 <- meta(indirect1[, c("ind_eff", "dir_eff")],
##D                   indirect1[, c("ind_var", "ind_dir_cov", "dir_var")])
##D 
##D summary(indirect2)  
## End(Not run)

[Package metaSEM version 0.9.4 Index]